Class to functional changes, remove unwanted files, remove unwanted code, implemented reset functionality
