+++ src/components/box-shadow/index.js
@@ -4,7 +4,7 @@
  */
  import { __ } from '@wordpress/i18n';
  import Range from '../../components/range/Range.js';
- import AdvancedPopColorControl from '../../advanced-pop-color-control'
+ import AdvancedPopColorControl from '../../components/color-control/advanced-pop-color-control'
 
 const {
 	ColorPalette
