--- src/icons.js
@@ -1,1454 +1,0 @@
-const icons = {};
-icons.blockRow = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <path d='M46.506,7.877l-45.012,0c-0.484,0 -0.882,0.398 -0.882,0.882l0,23.645c0,0.484 0.398,0.882 0.882,0.882l45.012,0c0.484,0 0.882,-0.398 0.882,-0.882l0,-23.65c-0.003,-0.482 -0.4,-0.877 -0.882,-0.877Zm-0.882,23.65l-43.248,0l0,-21.886l43.248,0l0,21.886Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <rect x='4.587' y='11.533' width='38.825' height='18.096' fill='#0058b0'
-    />
-    <path d='M43.413,29.63l-38.826,0l0,-18.097l38.826,0l0,18.097Zm-37.061,-16.332l0,14.567l35.296,0l0,-14.567l-35.296,0Z'
-    fill='#0058b0' />
-    <path d='M24,40.123c4.365,0 7.957,-3.592 7.957,-7.957c0,-4.365 -3.592,-7.957 -7.957,-7.957c-4.365,0 -7.957,3.592 -7.957,7.957c0,0.001 0,0.002 0,0.003c0,4.363 3.591,7.954 7.954,7.954c0.001,0 0.002,0 0.003,0Z'
-    fill='#0058b0' fillRule='nonzero' stroke='#fff' strokeWidth='1.76' />
-    <path d='M20.811,31.539l2.307,0l0,-2.309c0,-0.484 0.398,-0.882 0.882,-0.882c0.484,0 0.882,0.398 0.882,0.882l0,2.309l2.307,0c0.483,0 0.882,0.398 0.882,0.882c0,0.484 -0.399,0.882 -0.882,0.882l-2.307,0l0,2.307c0,0.484 -0.398,0.882 -0.882,0.882c-0.484,0 -0.882,-0.398 -0.882,-0.882l0,-2.307l-2.307,0c-0.483,0 -0.882,-0.398 -0.882,-0.882c0,-0.484 0.399,-0.882 0.882,-0.882Z'
-    fill='#fff' fillRule='nonzero' />
-</svg>;
-icons.row = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.collapserow = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="16.000" width="60.000" height="14.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.collapserowthree = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="9.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="10.500" width="60.000" height="9.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="21.000" width="60.000" height="9.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.collapserowfour = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="6.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="8.000" width="60.000" height="6.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="16.000" width="60.000" height="6.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="24.000" width="60.000" height="6.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.collapserowfive = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="5.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="6.000" width="60.000" height="5.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="12.000" width="60.000" height="5.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="18.000" width="60.000" height="5.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="24.000" width="60.000" height="5.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.collapserowsix = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="5.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="10.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="15.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="20.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="25.000" width="60.000" height="4.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.twocol = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="31.000" y="0.000" width="29.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="0.000" width="29.000" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.grid = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="31.000" y="0.000" width="29.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="0.000" width="29.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="31.000" y="16.000" width="29.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="16.000" width="29.000" height="14.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.threecol = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="18.500" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="20.500" y="0.000" width="19.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="41.500" y="0.000" width="18.500" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.threegrid = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="18.500" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="20.500" y="0.000" width="19.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="41.500" y="0.000" width="18.500" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="16.000" width="18.500" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="20.500" y="16.000" width="19.000" height="14.000" fill="#d5dadf"
-	/>
-	<rect x="41.500" y="16.000" width="18.500" height="14.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.lastrow = <svg viewBox='0 0 60 30' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='31' width='29' height='14' fill='#d5dadf' />
-    <rect x='-0.024' width='29' height='14' fill='#d5dadf' />
-    <rect x='-0.024' y='16' width='60' height='14' fill='#d5dadf' />
-</svg>;
-icons.firstrow = <svg viewBox='0 0 60 30' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='31' y='16' width='29' height='14' fill='#d5dadf' />
-    <rect x='-0.024' y='16' width='29' height='14' fill='#d5dadf' />
-    <rect x='-0.024' y='-0.003' width='60' height='14' fill='#d5dadf' />
-</svg>;
-icons.twoleftgolden = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="41.000" y="0.000" width="19.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="0.000" width="39.000" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.tworightgolden = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="21.000" y="0.000" width="39.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="0.000" y="0.000" width="19.000" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.lefthalf = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="29.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="31" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="46.500" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.righthalf = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="15.500" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="31.000" y="0.000" width="29.000" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.centerhalf = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-	clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="15.500" y="0.000" width="29.000" height="30.000" fill="#d5dadf"
-	/>
-	<rect x="46.500" y="0.000" width="13.500" height="30.000" fill="#d5dadf"
-	/>
-</svg>;
-icons.widecenter = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="11.000" height="30.000" fill="#d5dadf" />
-	<rect x="13.000" y="0.000" width="34.000" height="30.000" fill="#d5dadf" />
-	<rect x="49.000" y="0.000" width="11.000" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.exwidecenter = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="7.200" height="30.000" fill="#d5dadf" />
-	<rect x="9.200" y="0.000" width="41.600" height="30.000" fill="#d5dadf" />
-	<rect x="52.800" y="0.000" width="7.200" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.fourcol = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="13.500" height="30.000" fill="#d5dadf" />
-	<rect x="15.500" y="0.000" width="13.500" height="30.000" fill="#d5dadf" />
-	<rect x="31.000" y="0.000" width="13.500" height="30.000" fill="#d5dadf" />
-	<rect x="46.500" y="0.000" width="13.500" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.lfourforty = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="21.600" height="30.000" fill="#d5dadf" />
-	<rect x="23.600" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-	<rect x="36.400" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-	<rect x="49.200" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.rfourforty = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-	<rect x="12.800" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-	<rect x="25.600" y="0.000" width="10.800" height="30.000" fill="#d5dadf" />
-	<rect x="38.400" y="0.000" width="21.600" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.fivecol = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="10.400" height="30.000" fill="#d5dadf" />
-	<rect x="12.400" y="0.000" width="10.400" height="30.000" fill="#d5dadf" />
-	<rect x="24.800" y="0.000" width="10.400" height="30.000" fill="#d5dadf" />
-	<rect x="37.200" y="0.000" width="10.400" height="30.000" fill="#d5dadf" />
-	<rect x="49.600" y="0.000" width="10.400" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.sixcol = <svg viewBox="0 0 60 30" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="0.000" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-	<rect x="10.330" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-	<rect x="20.660" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-	<rect x="30.990" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-	<rect x="41.320" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-	<rect x="51.650" y="0.000" width="8.350" height="30.000" fill="#d5dadf" />
-</svg>;
-icons.aligntop = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="4.545" width="11.022" height="2.512" />
-</svg>;
-icons.alignmiddle = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="8.744" width="11.022" height="2.512" />
-</svg>;
-icons.alignbottom = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fill="currentColor" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M17.294,17.287l-14.588,0l0,-14.574l14.588,0c0,4.858 0,9.716 0,14.574Zm-13.738,-0.85l12.888,0l0,-12.874l-12.888,0c0,4.291 0,8.583 0,12.874Z" fillRule="nonzero" />
-	<rect x="4.489" y="12.802" width="11.022" height="2.512" />
-</svg>;
-icons.blockColumn = <svg width="20px" height="20px" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M47.612,40.008l-19.999,0l0,-32.016l13.566,0l6.433,0.023l0,31.993Z" fill="#177ee5" fillRule="nonzero" />
-	<path d="M20.383,40l-20.036,0l0,-32l20.036,0.019l0,31.981Z" fill="#0058b0" fillRule="nonzero" />
-</svg>;
-icons.outlinetop = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1.048" height="9.017" fill="#555d66" />
-	<rect x="16.265" y="5.498" width="1.023" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.186" width="8.964" height="2.482" fill="#272b2f" />
-	<rect x="5.487" y="16.261" width="9.026" height="1.037" fill="#555d66" />
-</svg>;
-icons.outlineright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1.046" height="9.017" fill="#555d66" />
-	<rect x="15.244" y="5.498" width="2.518" height="9.003" fill="#272b2f" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.954" fill="#555d66" />
-	<rect x="5.487" y="16.308" width="9.026" height="0.99" fill="#555d66" />
-</svg>;
-icons.outlinebottom = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.714" y="5.492" width="1" height="9.017" fill="#555d66" />
-	<rect x="16.261" y="5.498" width="1.027" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.968" fill="#555d66" />
-	<rect x="5.487" y="15.28" width="9.026" height="2.499" fill="#272b2f" />
-</svg>;
-icons.outlineleft = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<rect x="2.202" y="5.492" width="2.503" height="9.017" fill="#272b2f" />
-	<rect x="16.276" y="5.498" width="1.012" height="9.003" fill="#555d66" />
-	<rect x="5.518" y="2.719" width="8.964" height="0.966" fill="#555d66" />
-	<rect x="5.487" y="16.303" width="9.026" height="0.995" fill="#555d66" />
-</svg>;
-icons.blocktabs = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <path d='M13.864,6.719l0.02,0l0.019,0.001l0.019,0.001l0.02,0.002l0.019,0.002l0.018,0.003l0.019,0.003l0.019,0.003l0.018,0.004l0.018,0.005l0.019,0.005l0.017,0.005l0.018,0.006l0.018,0.006l0.017,0.007l0.017,0.007l0.017,0.007l0.017,0.008l0.016,0.008l0.016,0.009l0.016,0.008l0.016,0.01l0.015,0.009l0.016,0.011l0.015,0.01l0.014,0.011l0.015,0.011l0.014,0.011l0.014,0.012l0.014,0.012l0.013,0.013l0.013,0.012l0.013,0.013l0.012,0.013l0.012,0.014l0.012,0.014l0.011,0.014l0.011,0.015l0.011,0.014l0.011,0.015l0.01,0.016l0.009,0.015l0.01,0.016l0.008,0.016l0.009,0.016l0.008,0.017l0.008,0.016l0.007,0.017l0.007,0.017l0.007,0.017l0.006,0.018l0.006,0.018l0.005,0.017l0.005,0.019l0.005,0.018l0.004,0.018l0.003,0.019l0.003,0.019l0.003,0.019l0.002,0.018l0.002,0.02l0.001,0.019l0.001,0.019l0,0.02l0,5.225l32.775,0.007l0,27.817l0,0.02l-0.001,0.019l-0.001,0.019l-0.002,0.02l-0.002,0.019l-0.003,0.019l-0.003,0.018l-0.004,0.019l-0.004,0.018l-0.004,0.018l-0.005,0.018l-0.005,0.018l-0.006,0.018l-0.006,0.018l-0.007,0.017l-0.007,0.017l-0.007,0.017l-0.008,0.017l-0.008,0.016l-0.009,0.016l-0.009,0.016l-0.009,0.016l-0.01,0.016l-0.01,0.015l-0.01,0.015l-0.011,0.014l-0.011,0.015l-0.011,0.014l-0.012,0.014l-0.012,0.013l-0.012,0.014l-0.013,0.013l-0.013,0.013l-0.014,0.012l-0.013,0.012l-0.014,0.012l-0.014,0.011l-0.015,0.011l-0.014,0.011l-0.015,0.01l-0.016,0.01l-0.015,0.01l-0.016,0.009l-0.016,0.009l-0.016,0.009l-0.016,0.008l-0.017,0.008l-0.017,0.007l-0.017,0.007l-0.018,0.007l-0.017,0.006l-0.018,0.006l-0.018,0.005l-0.018,0.005l-0.018,0.004l-0.018,0.004l-0.019,0.004l-0.019,0.003l-0.018,0.003l-0.02,0.002l-0.019,0.002l-0.019,0.001l-0.019,0.001l-0.02,0l-45.28,0l-0.02,0l-0.019,-0.001l-0.019,-0.001l-0.019,-0.002l-0.02,-0.002l-0.019,-0.003l-0.018,-0.003l-0.019,-0.004l-0.018,-0.004l-0.019,-0.004l-0.017,-0.005l-0.018,-0.005l-0.018,-0.006l-0.018,-0.006l-0.017,-0.007l-0.017,-0.007l-0.017,-0.007l-0.017,-0.008l-0.016,-0.008l-0.016,-0.009l-0.016,-0.009l-0.016,-0.009l-0.016,-0.01l-0.015,-0.01l-0.015,-0.01l-0.014,-0.011l-0.015,-0.011l-0.014,-0.011l-0.014,-0.012l-0.014,-0.012l-0.013,-0.012l-0.013,-0.013l-0.013,-0.013l-0.012,-0.014l-0.012,-0.013l-0.012,-0.014l-0.011,-0.014l-0.011,-0.015l-0.011,-0.014l-0.01,-0.015l-0.01,-0.015l-0.01,-0.016l-0.009,-0.016l-0.009,-0.016l-0.009,-0.016l-0.008,-0.016l-0.008,-0.017l-0.007,-0.017l-0.007,-0.017l-0.007,-0.017l-0.006,-0.018l-0.006,-0.018l-0.005,-0.018l-0.005,-0.018l-0.004,-0.018l-0.004,-0.018l-0.004,-0.019l-0.003,-0.018l-0.003,-0.019l-0.002,-0.019l-0.002,-0.02l-0.001,-0.019l-0.001,-0.019l0,-0.02l0,-33.052l0,-0.019l0.001,-0.02l0.001,-0.019l0.002,-0.019l0.002,-0.019l0.003,-0.019l0.003,-0.019l0.004,-0.019l0.004,-0.018l0.004,-0.018l0.005,-0.018l0.005,-0.018l0.006,-0.018l0.006,-0.017l0.007,-0.018l0.007,-0.017l0.007,-0.017l0.008,-0.016l0.008,-0.017l0.009,-0.016l0.009,-0.016l0.009,-0.016l0.01,-0.015l0.01,-0.015l0.01,-0.016l0.011,-0.014l0.011,-0.015l0.011,-0.014l0.012,-0.014l0.012,-0.013l0.012,-0.014l0.013,-0.013l0.013,-0.012l0.013,-0.013l0.014,-0.012l0.014,-0.011l0.014,-0.012l0.015,-0.011l0.014,-0.011l0.015,-0.01l0.015,-0.01l0.016,-0.01l0.016,-0.009l0.016,-0.009l0.016,-0.009l0.016,-0.008l0.017,-0.008l0.017,-0.007l0.017,-0.007l0.017,-0.007l0.018,-0.006l0.018,-0.005l0.018,-0.006l0.017,-0.005l0.019,-0.004l0.018,-0.004l0.019,-0.004l0.018,-0.003l0.019,-0.002l0.02,-0.003l0.019,-0.002l0.019,-0.001l0.019,-0.001l0.02,0l12.505,0.003Zm-11.496,32.798l43.262,0l0,-25.042l-32.775,-0.006l0,-5.984l-10.487,-0.002l0,31.034Z'
-    fill='#0058b0' />
-    <path d='M30.246,6.72l0.021,0.001l0.02,0.001l0.02,0.002l0.02,0.002l0.019,0.003l0.02,0.003l0.019,0.004l0.019,0.004l0.019,0.005l0.019,0.005l0.019,0.006l0.018,0.006l0.019,0.006l0.018,0.007l0.018,0.007l0.017,0.008l0.018,0.008l0.017,0.009l0.017,0.009l0.016,0.009l0.017,0.01l0.016,0.01l0.016,0.01l0.016,0.011l0.015,0.011l0.015,0.012l0.015,0.012l0.029,0.024l0.013,0.014l0.014,0.013l0.013,0.013l0.013,0.014l0.013,0.014l0.012,0.015l0.012,0.015l0.012,0.015l0.011,0.015l0.01,0.016l0.011,0.016l0.01,0.016l0.01,0.016l0.009,0.017l0.009,0.017l0.009,0.017l0.008,0.017l0.007,0.018l0.008,0.018l0.007,0.018l0.006,0.018l0.006,0.019l0.005,0.018l0.006,0.019l0.004,0.019l0.005,0.019l0.003,0.02l0.003,0.019l0.003,0.02l0.002,0.02l0.002,0.02l0.001,0.02l0.001,0.02l0.001,0.021l0,6.165l-14.034,0l0,-6.165l0.001,-0.021l0.001,-0.02l0.001,-0.02l0.002,-0.02l0.002,-0.02l0.003,-0.02l0.003,-0.019l0.003,-0.02l0.005,-0.019l0.004,-0.019l0.006,-0.019l0.005,-0.018l0.006,-0.019l0.006,-0.018l0.007,-0.018l0.008,-0.018l0.007,-0.018l0.008,-0.017l0.009,-0.017l0.009,-0.017l0.009,-0.017l0.01,-0.016l0.01,-0.016l0.011,-0.016l0.01,-0.016l0.011,-0.015l0.012,-0.015l0.012,-0.015l0.012,-0.015l0.013,-0.014l0.013,-0.014l0.013,-0.013l0.014,-0.013l0.013,-0.014l0.015,-0.012l0.014,-0.012l0.015,-0.012l0.015,-0.012l0.015,-0.011l0.016,-0.011l0.016,-0.01l0.016,-0.01l0.017,-0.01l0.016,-0.009l0.017,-0.009l0.017,-0.009l0.018,-0.008l0.017,-0.008l0.018,-0.007l0.018,-0.007l0.019,-0.006l0.018,-0.006l0.019,-0.006l0.019,-0.005l0.019,-0.005l0.019,-0.004l0.019,-0.004l0.02,-0.003l0.019,-0.003l0.02,-0.002l0.02,-0.002l0.02,-0.001l0.021,-0.001l0.02,0l12.452,0l0.02,0Z'
-    fill='#0058b0' />
-    <path d='M46.626,6.72l0.02,0.001l0.02,0.001l0.02,0.002l0.02,0.002l0.02,0.003l0.019,0.003l0.02,0.004l0.019,0.004l0.019,0.005l0.019,0.005l0.019,0.006l0.018,0.006l0.018,0.006l0.018,0.007l0.018,0.007l0.018,0.008l0.017,0.008l0.017,0.009l0.017,0.009l0.017,0.009l0.017,0.01l0.016,0.01l0.016,0.01l0.015,0.011l0.015,0.011l0.016,0.012l0.014,0.012l0.029,0.024l0.014,0.014l0.013,0.013l0.014,0.013l0.013,0.014l0.012,0.014l0.013,0.015l0.012,0.015l0.011,0.015l0.011,0.015l0.011,0.016l0.011,0.016l0.01,0.016l0.009,0.016l0.01,0.017l0.009,0.017l0.008,0.017l0.008,0.017l0.008,0.018l0.007,0.018l0.007,0.018l0.007,0.018l0.006,0.019l0.005,0.018l0.005,0.019l0.005,0.019l0.004,0.019l0.004,0.02l0.003,0.019l0.003,0.02l0.002,0.02l0.002,0.02l0.001,0.02l0.001,0.02l0,0.021l0,6.165l-14.033,0l0,-6.165l0,-0.021l0.001,-0.02l0.001,-0.02l0.002,-0.02l0.002,-0.02l0.003,-0.02l0.003,-0.019l0.004,-0.02l0.004,-0.019l0.005,-0.019l0.005,-0.019l0.006,-0.018l0.005,-0.019l0.007,-0.018l0.007,-0.018l0.007,-0.018l0.008,-0.018l0.008,-0.017l0.009,-0.017l0.008,-0.017l0.01,-0.017l0.009,-0.016l0.011,-0.016l0.01,-0.016l0.011,-0.016l0.011,-0.015l0.012,-0.015l0.012,-0.015l0.012,-0.015l0.013,-0.014l0.012,-0.014l0.014,-0.013l0.013,-0.013l0.014,-0.014l0.014,-0.012l0.015,-0.012l0.015,-0.012l0.015,-0.012l0.015,-0.011l0.016,-0.011l0.016,-0.01l0.016,-0.01l0.016,-0.01l0.017,-0.009l0.017,-0.009l0.017,-0.009l0.017,-0.008l0.018,-0.008l0.018,-0.007l0.018,-0.007l0.018,-0.006l0.019,-0.006l0.018,-0.006l0.019,-0.005l0.019,-0.005l0.019,-0.004l0.02,-0.004l0.019,-0.003l0.02,-0.003l0.02,-0.002l0.02,-0.002l0.02,-0.001l0.02,-0.001l0.021,0l12.452,0l0.02,0Z'
-    fill='#0058b0' />
-    <path d='M19.753,18.081l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M19.753,23.146l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.884c0,0.484 0.399,0.883 0.884,0.883l11.238,0c0.484,0 0.883,-0.399 0.883,-0.883c0,-0.485 -0.399,-0.884 -0.883,-0.884Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M19.753,28.681l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M19.753,33.746l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M39.461,18.081l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M39.461,23.146l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.884c0,0.484 0.399,0.883 0.884,0.883l11.238,0c0.484,0 0.883,-0.399 0.883,-0.883c0,-0.485 -0.399,-0.884 -0.883,-0.884Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M39.461,28.681l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M39.461,33.746l-11.238,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.884 0.884,0.884l11.238,0c0.484,0 0.883,-0.399 0.883,-0.884c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-    fill='#0058b0' fillRule='nonzero' />
-</svg>;
-icons.tabs = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<path d='M17.294,15.966l-14.588,0l0,-11.932l4.122,0.001l-0.002,3.473l10.468,0.007c0,4.339 0,4.112 0,8.451Zm-13.738,-0.76l12.888,0l0,-6.932l-10.448,0l-0.008,-3.541l-2.432,0.003c0,3.832 0,6.638 0,10.47Z' fill='#555d66' fillRule='nonzero' />
-	<rect x='7.919' y='4.042' width='4.152' height='2.669' fill='#555d66' />
-	<rect x='13.135' y='4.045' width='4.152' height='2.657' fill='#555d66' />
-</svg>;
-icons.vtabs = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<path d='M17.227,17.294l0,-14.588l-14.454,0l0.002,3.319l3.473,-0.003l0.006,11.272c4.339,0 6.634,0 10.973,0Zm-0.759,-13.738l0,12.888l-9.455,0l0.001,-11.251l-3.542,-0.008l0.002,-1.629c3.833,0 9.161,0 12.994,0Z' fill='#555d66' fillRule='nonzero' />
-	<path d='M2.82,9.846l0,-3.033l2.669,0l0,3.033l-2.669,0Z' fill='#555d66' />
-	<path d='M2.823,13.676l0,-3.011l2.656,0l0,3.011l-2.656,0Z' fill='#555d66' />
-</svg>;
-icons.accordion = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<path d='M17.275,9.664l-14.607,0l0,-3.036l14.607,0l0,3.036Zm-13.807,-2.236l0,1.436l13.007,0l0,-1.436l-13.007,0Z' fill='#555d66' />
-	<path d='M17.305,6.28l-14.607,0l0,-3.036l14.607,0l0,3.036Zm-13.807,-2.236l0,1.436l13.007,0l0,-1.436l-13.007,0Z' fill='#555d66' />
-	<path d='M17.316,16.771l-14.608,0l0,-3.036l14.608,0l0,3.036Zm-13.808,-2.236l0,1.436l13.008,0l0,-1.436l-13.008,0Z' fill='#555d66' />
-	<path d='M16.392,13.423l-12.815,0l0,-4.553l12.815,0l0,4.553Zm-12.015,-3.753l0,2.953l11.215,0l0,-2.953l-11.215,0Z' fill='#555d66' />
-</svg>;
-icons.linked = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<path d='M17.888,17.895l-15.776,0l0,-15.79l15.776,0l0,15.79Zm-13.276,-13.29l0,10.79l10.776,0l0,-10.79l-10.776,0Z' />
-</svg>;
-icons.individual = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd' clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-	<rect x='2.024' y='5.486' width='2.61' height='9.027' />
-	<rect x='15.379' y='5.497' width='2.56' height='9.007' />
-	<rect x='5.509' y='1.999' width='8.983' height='2.582' />
-	<rect x='5.51' y='15.36' width='8.979' height='2.563' />
-</svg>;
-icons.radiuslinked = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,7.881l-0.001,0.102l-0.004,0.101l-0.006,0.1l-0.009,0.1l-0.011,0.098l-0.014,0.099l-0.016,0.097l-0.019,0.097l-0.021,0.095l-0.023,0.095l-0.025,0.094l-0.028,0.093l-0.03,0.092l-0.032,0.091l-0.034,0.09l-0.037,0.089l-0.038,0.088l-0.041,0.087l-0.042,0.085l-0.045,0.085l-0.046,0.083l-0.049,0.082l-0.05,0.08l-0.052,0.079l-0.054,0.078l-0.056,0.077l-0.058,0.075l-0.059,0.074l-0.061,0.072l-0.063,0.071l-0.064,0.069l-0.066,0.068l-0.068,0.066l-0.069,0.064l-0.071,0.063l-0.072,0.061l-0.074,0.06l-0.075,0.057l-0.077,0.056l-0.078,0.054l-0.079,0.052l-0.081,0.051l-0.082,0.048l-0.083,0.047l-0.084,0.044l-0.086,0.043l-0.086,0.04l-0.088,0.039l-0.089,0.036l-0.09,0.034l-0.091,0.032l-0.092,0.03l-0.093,0.028l-0.094,0.025l-0.095,0.023l-0.096,0.021l-0.096,0.019l-0.098,0.016l-0.098,0.014l-0.099,0.011l-0.099,0.009l-0.101,0.006l-0.101,0.004l-0.101,0.001l-7.905,0l-0.101,-0.001l-0.101,-0.004l-0.101,-0.006l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.019l-0.096,-0.021l-0.095,-0.023l-0.094,-0.025l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.067,-0.066l-0.067,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.077l-0.054,-0.078l-0.052,-0.079l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.095l-0.021,-0.095l-0.019,-0.097l-0.016,-0.097l-0.014,-0.099l-0.011,-0.098l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-7.881l0.001,-0.102l0.004,-0.101l0.006,-0.1l0.009,-0.1l0.011,-0.099l0.014,-0.098l0.016,-0.097l0.019,-0.097l0.021,-0.096l0.023,-0.094l0.025,-0.094l0.028,-0.093l0.03,-0.092l0.032,-0.091l0.034,-0.09l0.037,-0.089l0.038,-0.088l0.041,-0.087l0.042,-0.085l0.045,-0.085l0.046,-0.083l0.049,-0.082l0.05,-0.08l0.052,-0.08l0.054,-0.078l0.056,-0.076l0.058,-0.075l0.059,-0.074l0.061,-0.072l0.063,-0.071l0.064,-0.069l0.067,-0.068l0.067,-0.066l0.069,-0.064l0.071,-0.063l0.072,-0.061l0.074,-0.06l0.075,-0.057l0.077,-0.056l0.078,-0.054l0.079,-0.052l0.081,-0.051l0.082,-0.048l0.083,-0.047l0.084,-0.044l0.086,-0.043l0.086,-0.04l0.088,-0.039l0.089,-0.036l0.09,-0.034l0.091,-0.032l0.092,-0.03l0.093,-0.028l0.094,-0.026l0.095,-0.023l0.096,-0.021l0.096,-0.018l0.098,-0.016l0.098,-0.014l0.099,-0.011l0.099,-0.009l0.101,-0.007l0.101,-0.003l0.101,-0.002l7.905,0l0.101,0.002Zm-7.99,2.498l-0.054,0.001l-0.038,0.001l-0.037,0.003l-0.036,0.003l-0.037,0.004l-0.035,0.005l-0.036,0.006l-0.035,0.007l-0.034,0.007l-0.034,0.009l-0.034,0.009l-0.034,0.01l-0.033,0.011l-0.033,0.011l-0.032,0.013l-0.032,0.013l-0.032,0.014l-0.032,0.014l-0.031,0.016l-0.03,0.016l-0.031,0.017l-0.03,0.018l-0.029,0.018l-0.029,0.019l-0.029,0.02l-0.028,0.021l-0.028,0.021l-0.027,0.022l-0.027,0.022l-0.026,0.024l-0.026,0.023l-0.025,0.025l-0.024,0.025l-0.024,0.026l-0.023,0.026l-0.023,0.026l-0.022,0.028l-0.021,0.027l-0.02,0.029l-0.02,0.028l-0.019,0.029l-0.019,0.03l-0.018,0.03l-0.017,0.03l-0.016,0.031l-0.015,0.031l-0.015,0.031l-0.014,0.032l-0.013,0.032l-0.012,0.033l-0.012,0.033l-0.011,0.033l-0.01,0.033l-0.009,0.034l-0.008,0.034l-0.008,0.035l-0.006,0.035l-0.006,0.035l-0.005,0.036l-0.004,0.036l-0.004,0.037l-0.002,0.037l-0.002,0.038l0,0.054l0,7.849l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.034l0.008,0.035l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.032l0.013,0.033l0.014,0.031l0.015,0.032l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.019,0.029l0.019,0.03l0.02,0.028l0.02,0.028l0.021,0.028l0.022,0.027l0.023,0.027l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.024l0.026,0.024l0.026,0.023l0.027,0.023l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.034,0.009l0.034,0.008l0.034,0.008l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.002l0.038,0.002l0.054,0l7.873,0l0.054,0l0.038,-0.002l0.037,-0.002l0.036,-0.003l0.037,-0.004l0.035,-0.005l0.036,-0.006l0.035,-0.007l0.034,-0.008l0.034,-0.008l0.034,-0.009l0.034,-0.01l0.033,-0.011l0.033,-0.011l0.032,-0.013l0.032,-0.013l0.032,-0.014l0.032,-0.014l0.031,-0.016l0.03,-0.016l0.031,-0.017l0.03,-0.018l0.029,-0.018l0.029,-0.019l0.029,-0.02l0.028,-0.021l0.028,-0.021l0.027,-0.022l0.027,-0.023l0.026,-0.023l0.026,-0.024l0.025,-0.024l0.024,-0.025l0.024,-0.026l0.023,-0.026l0.023,-0.027l0.022,-0.027l0.021,-0.028l0.021,-0.028l0.019,-0.028l0.02,-0.03l0.018,-0.029l0.018,-0.03l0.017,-0.03l0.016,-0.031l0.015,-0.031l0.015,-0.032l0.014,-0.031l0.013,-0.033l0.012,-0.032l0.012,-0.033l0.011,-0.033l0.01,-0.033l0.009,-0.034l0.008,-0.035l0.008,-0.034l0.006,-0.035l0.006,-0.035l0.005,-0.036l0.004,-0.036l0.004,-0.037l0.002,-0.037l0.002,-0.038l0,-0.054l0,-7.849l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-7.873,0Z" fill="#555d66" />
-</svg>;
-icons.radiusindividual = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,1.273l-5.531,-5.214l1.591,0l0.101,0.002Zm-1.692,-0.002l5.531,5.214l-2.5,-0.012l0,-1.245l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-1.57,0l-0.005,-2.5Z" fill="#555d66" />
-	<path d="M14.044,17.873l0.101,-0.003l0.101,-0.007l0.099,-0.009l0.099,-0.011l0.098,-0.014l0.098,-0.016l0.096,-0.018l0.096,-0.021l0.095,-0.023l0.094,-0.026l0.093,-0.028l0.092,-0.03l0.091,-0.032l0.09,-0.034l0.089,-0.036l0.088,-0.039l0.086,-0.04l0.086,-0.043l0.084,-0.044l0.083,-0.047l0.082,-0.048l0.081,-0.051l0.079,-0.052l0.078,-0.054l0.077,-0.056l0.075,-0.057l0.074,-0.06l0.072,-0.061l0.071,-0.063l0.069,-0.064l0.068,-0.066l0.066,-0.068l0.064,-0.069l0.063,-0.071l0.061,-0.072l0.059,-0.074l0.058,-0.075l0.056,-0.076l0.054,-0.078l0.052,-0.08l0.05,-0.08l0.049,-0.082l0.046,-0.083l0.045,-0.085l0.042,-0.085l0.041,-0.087l0.038,-0.088l0.037,-0.089l0.034,-0.09l0.032,-0.091l0.03,-0.092l0.028,-0.093l0.025,-0.094l0.023,-0.094l0.021,-0.096l0.019,-0.097l0.016,-0.097l0.014,-0.098l0.011,-0.099l0.009,-0.1l0.006,-0.1l0.004,-0.101l0.001,-0.102l0,-1.272l-5.531,5.213l1.591,0l0.101,-0.002Zm-1.692,0.002l5.531,-5.213l-2.5,0.012l0,1.244l0,0.054l-0.002,0.038l-0.002,0.037l-0.004,0.037l-0.004,0.036l-0.005,0.036l-0.006,0.035l-0.006,0.035l-0.008,0.035l-0.008,0.034l-0.009,0.034l-0.01,0.033l-0.011,0.033l-0.012,0.033l-0.012,0.033l-0.013,0.032l-0.014,0.032l-0.015,0.031l-0.015,0.031l-0.016,0.031l-0.017,0.03l-0.018,0.03l-0.018,0.03l-0.02,0.029l-0.019,0.028l-0.021,0.029l-0.021,0.027l-0.022,0.028l-0.023,0.026l-0.023,0.026l-0.024,0.026l-0.024,0.025l-0.025,0.025l-0.026,0.023l-0.026,0.024l-0.027,0.022l-0.027,0.022l-0.028,0.021l-0.028,0.021l-0.029,0.02l-0.029,0.019l-0.029,0.018l-0.03,0.018l-0.031,0.017l-0.03,0.016l-0.031,0.016l-0.032,0.014l-0.032,0.014l-0.032,0.013l-0.032,0.013l-0.033,0.011l-0.033,0.011l-0.034,0.01l-0.033,0.009l-0.035,0.009l-0.034,0.007l-0.035,0.007l-0.036,0.006l-0.035,0.005l-0.037,0.004l-0.036,0.003l-0.037,0.003l-0.038,0.001l-0.054,0.001l-1.57,0l-0.005,2.5Z" fill="#555d66" />
-	<path d="M6.077,17.873l-0.101,-0.003l-0.101,-0.007l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.018l-0.096,-0.021l-0.095,-0.023l-0.094,-0.026l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.068,-0.066l-0.066,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.076l-0.054,-0.078l-0.052,-0.08l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.094l-0.021,-0.096l-0.019,-0.097l-0.016,-0.097l-0.014,-0.098l-0.011,-0.099l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-1.272l5.531,5.213l-1.591,0l-0.101,-0.002Zm1.692,0.002l-5.531,-5.213l2.5,0.012l0,1.244l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.035l0.008,0.034l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.033l0.013,0.032l0.014,0.032l0.015,0.031l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.018,0.03l0.02,0.029l0.019,0.028l0.021,0.029l0.021,0.027l0.022,0.028l0.023,0.026l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.025l0.026,0.023l0.026,0.024l0.027,0.022l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.033,0.009l0.035,0.009l0.034,0.007l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.003l0.038,0.001l0.054,0.001l1.569,0l0.006,2.5Z" fill="#555d66" />
-	<path d="M6.077,2.124l-0.101,0.003l-0.101,0.007l-0.099,0.009l-0.099,0.011l-0.098,0.014l-0.098,0.016l-0.096,0.018l-0.096,0.021l-0.095,0.023l-0.094,0.026l-0.093,0.028l-0.092,0.03l-0.091,0.032l-0.09,0.034l-0.089,0.036l-0.088,0.039l-0.086,0.04l-0.086,0.043l-0.084,0.044l-0.083,0.047l-0.082,0.048l-0.081,0.051l-0.079,0.052l-0.078,0.054l-0.077,0.056l-0.075,0.057l-0.074,0.06l-0.072,0.061l-0.071,0.063l-0.069,0.064l-0.068,0.066l-0.066,0.068l-0.064,0.069l-0.063,0.071l-0.061,0.072l-0.059,0.074l-0.058,0.075l-0.056,0.076l-0.054,0.078l-0.052,0.08l-0.05,0.08l-0.049,0.082l-0.046,0.083l-0.045,0.085l-0.042,0.085l-0.041,0.087l-0.038,0.088l-0.037,0.089l-0.034,0.09l-0.032,0.091l-0.03,0.092l-0.028,0.093l-0.025,0.094l-0.023,0.094l-0.021,0.096l-0.019,0.097l-0.016,0.097l-0.014,0.098l-0.011,0.099l-0.009,0.1l-0.006,0.1l-0.004,0.101l-0.001,0.102l0,1.272l5.531,-5.213l-1.591,0l-0.101,0.002Zm1.692,-0.002l-5.531,5.213l2.5,-0.011l0,-1.245l0,-0.054l0.002,-0.038l0.002,-0.037l0.004,-0.037l0.004,-0.036l0.005,-0.036l0.006,-0.035l0.006,-0.035l0.008,-0.035l0.008,-0.034l0.009,-0.034l0.01,-0.033l0.011,-0.033l0.012,-0.033l0.012,-0.033l0.013,-0.032l0.014,-0.032l0.015,-0.031l0.015,-0.031l0.016,-0.031l0.017,-0.03l0.018,-0.03l0.018,-0.03l0.02,-0.029l0.019,-0.028l0.021,-0.029l0.021,-0.027l0.022,-0.028l0.023,-0.026l0.023,-0.026l0.024,-0.026l0.024,-0.025l0.025,-0.025l0.026,-0.023l0.026,-0.024l0.027,-0.022l0.027,-0.022l0.028,-0.021l0.028,-0.021l0.029,-0.02l0.029,-0.019l0.029,-0.018l0.03,-0.018l0.031,-0.017l0.03,-0.016l0.031,-0.016l0.032,-0.014l0.032,-0.014l0.032,-0.013l0.032,-0.013l0.033,-0.011l0.033,-0.011l0.034,-0.01l0.033,-0.009l0.035,-0.009l0.034,-0.007l0.035,-0.007l0.036,-0.006l0.035,-0.005l0.037,-0.004l0.036,-0.003l0.037,-0.003l0.038,-0.001l0.054,-0.001l1.569,0l0.006,-2.5Z" fill="#555d66" />
-</svg>;
-icons.topright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd" clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M14.044,2.12l0.101,0.003l0.101,0.007l0.099,0.009l0.099,0.011l0.098,0.014l0.098,0.016l0.096,0.018l0.096,0.021l0.095,0.023l0.094,0.026l0.093,0.028l0.092,0.03l0.091,0.032l0.09,0.034l0.089,0.036l0.088,0.039l0.086,0.04l0.086,0.043l0.084,0.044l0.083,0.047l0.082,0.048l0.081,0.051l0.079,0.052l0.078,0.054l0.077,0.056l0.075,0.057l0.074,0.06l0.072,0.061l0.071,0.063l0.069,0.064l0.068,0.066l0.066,0.068l0.064,0.069l0.063,0.071l0.061,0.072l0.059,0.074l0.058,0.075l0.056,0.076l0.054,0.078l0.052,0.08l0.05,0.08l0.049,0.082l0.046,0.083l0.045,0.085l0.042,0.085l0.041,0.087l0.038,0.088l0.037,0.089l0.034,0.09l0.032,0.091l0.03,0.092l0.028,0.093l0.025,0.094l0.023,0.094l0.021,0.096l0.019,0.097l0.016,0.097l0.014,0.098l0.011,0.099l0.009,0.1l0.006,0.1l0.004,0.101l0.001,0.102l0,1.273l-5.531,-5.214l1.591,0l0.101,0.002Zm-1.692,-0.002l5.531,5.214l-2.5,-0.012l0,-1.245l0,-0.054l-0.002,-0.038l-0.002,-0.037l-0.004,-0.037l-0.004,-0.036l-0.005,-0.036l-0.006,-0.035l-0.006,-0.035l-0.008,-0.035l-0.008,-0.034l-0.009,-0.034l-0.01,-0.033l-0.011,-0.033l-0.012,-0.033l-0.012,-0.033l-0.013,-0.032l-0.014,-0.032l-0.015,-0.031l-0.015,-0.031l-0.016,-0.031l-0.017,-0.03l-0.018,-0.03l-0.018,-0.03l-0.02,-0.029l-0.019,-0.028l-0.021,-0.029l-0.021,-0.027l-0.022,-0.028l-0.023,-0.026l-0.023,-0.026l-0.024,-0.026l-0.024,-0.025l-0.025,-0.025l-0.026,-0.023l-0.026,-0.024l-0.027,-0.022l-0.027,-0.022l-0.028,-0.021l-0.028,-0.021l-0.029,-0.02l-0.029,-0.019l-0.029,-0.018l-0.03,-0.018l-0.031,-0.017l-0.03,-0.016l-0.031,-0.016l-0.032,-0.014l-0.032,-0.014l-0.032,-0.013l-0.032,-0.013l-0.033,-0.011l-0.033,-0.011l-0.034,-0.01l-0.033,-0.009l-0.035,-0.009l-0.034,-0.007l-0.035,-0.007l-0.036,-0.006l-0.035,-0.005l-0.037,-0.004l-0.036,-0.003l-0.037,-0.003l-0.038,-0.001l-0.054,-0.001l-1.57,0l-0.005,-2.5Z" fill="#555d66" />
-	<path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z" fill="#555d66" />
-	<path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z" fill="#555d66" />
-	<path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z" fill="#555d66" />
-</svg>;
-icons.topleft = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M6.077,2.124l-0.101,0.003l-0.101,0.007l-0.099,0.009l-0.099,0.011l-0.098,0.014l-0.098,0.016l-0.096,0.018l-0.096,0.021l-0.095,0.023l-0.094,0.026l-0.093,0.028l-0.092,0.03l-0.091,0.032l-0.09,0.034l-0.089,0.036l-0.088,0.039l-0.086,0.04l-0.086,0.043l-0.084,0.044l-0.083,0.047l-0.082,0.048l-0.081,0.051l-0.079,0.052l-0.078,0.054l-0.077,0.056l-0.075,0.057l-0.074,0.06l-0.072,0.061l-0.071,0.063l-0.069,0.064l-0.068,0.066l-0.066,0.068l-0.064,0.069l-0.063,0.071l-0.061,0.072l-0.059,0.074l-0.058,0.075l-0.056,0.076l-0.054,0.078l-0.052,0.08l-0.05,0.08l-0.049,0.082l-0.046,0.083l-0.045,0.085l-0.042,0.085l-0.041,0.087l-0.038,0.088l-0.037,0.089l-0.034,0.09l-0.032,0.091l-0.03,0.092l-0.028,0.093l-0.025,0.094l-0.023,0.094l-0.021,0.096l-0.019,0.097l-0.016,0.097l-0.014,0.098l-0.011,0.099l-0.009,0.1l-0.006,0.1l-0.004,0.101l-0.001,0.102l0,1.272l5.531,-5.213l-1.591,0l-0.101,0.002Zm1.692,-0.002l-5.531,5.213l2.5,-0.011l0,-1.245l0,-0.054l0.002,-0.038l0.002,-0.037l0.004,-0.037l0.004,-0.036l0.005,-0.036l0.006,-0.035l0.006,-0.035l0.008,-0.035l0.008,-0.034l0.009,-0.034l0.01,-0.033l0.011,-0.033l0.012,-0.033l0.012,-0.033l0.013,-0.032l0.014,-0.032l0.015,-0.031l0.015,-0.031l0.016,-0.031l0.017,-0.03l0.018,-0.03l0.018,-0.03l0.02,-0.029l0.019,-0.028l0.021,-0.029l0.021,-0.027l0.022,-0.028l0.023,-0.026l0.023,-0.026l0.024,-0.026l0.024,-0.025l0.025,-0.025l0.026,-0.023l0.026,-0.024l0.027,-0.022l0.027,-0.022l0.028,-0.021l0.028,-0.021l0.029,-0.02l0.029,-0.019l0.029,-0.018l0.03,-0.018l0.031,-0.017l0.03,-0.016l0.031,-0.016l0.032,-0.014l0.032,-0.014l0.032,-0.013l0.032,-0.013l0.033,-0.011l0.033,-0.011l0.034,-0.01l0.033,-0.009l0.035,-0.009l0.034,-0.007l0.035,-0.007l0.036,-0.006l0.035,-0.005l0.037,-0.004l0.036,-0.003l0.037,-0.003l0.038,-0.001l0.054,-0.001l1.569,0l0.006,-2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z"
-    fill="#555d66" />
-    <path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.bottomleft= <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M6.077,17.873l-0.101,-0.003l-0.101,-0.007l-0.099,-0.009l-0.099,-0.011l-0.098,-0.014l-0.098,-0.016l-0.096,-0.018l-0.096,-0.021l-0.095,-0.023l-0.094,-0.026l-0.093,-0.028l-0.092,-0.03l-0.091,-0.032l-0.09,-0.034l-0.089,-0.036l-0.088,-0.039l-0.086,-0.04l-0.086,-0.043l-0.084,-0.044l-0.083,-0.047l-0.082,-0.048l-0.081,-0.051l-0.079,-0.052l-0.078,-0.054l-0.077,-0.056l-0.075,-0.057l-0.074,-0.06l-0.072,-0.061l-0.071,-0.063l-0.069,-0.064l-0.068,-0.066l-0.066,-0.068l-0.064,-0.069l-0.063,-0.071l-0.061,-0.072l-0.059,-0.074l-0.058,-0.075l-0.056,-0.076l-0.054,-0.078l-0.052,-0.08l-0.05,-0.08l-0.049,-0.082l-0.046,-0.083l-0.045,-0.085l-0.042,-0.085l-0.041,-0.087l-0.038,-0.088l-0.037,-0.089l-0.034,-0.09l-0.032,-0.091l-0.03,-0.092l-0.028,-0.093l-0.025,-0.094l-0.023,-0.094l-0.021,-0.096l-0.019,-0.097l-0.016,-0.097l-0.014,-0.098l-0.011,-0.099l-0.009,-0.1l-0.006,-0.1l-0.004,-0.101l-0.001,-0.102l0,-1.272l5.531,5.213l-1.591,0l-0.101,-0.002Zm1.692,0.002l-5.531,-5.213l2.5,0.012l0,1.244l0,0.054l0.002,0.038l0.002,0.037l0.004,0.037l0.004,0.036l0.005,0.036l0.006,0.035l0.006,0.035l0.008,0.035l0.008,0.034l0.009,0.034l0.01,0.033l0.011,0.033l0.012,0.033l0.012,0.033l0.013,0.032l0.014,0.032l0.015,0.031l0.015,0.031l0.016,0.031l0.017,0.03l0.018,0.03l0.018,0.03l0.02,0.029l0.019,0.028l0.021,0.029l0.021,0.027l0.022,0.028l0.023,0.026l0.023,0.026l0.024,0.026l0.024,0.025l0.025,0.025l0.026,0.023l0.026,0.024l0.027,0.022l0.027,0.022l0.028,0.021l0.028,0.021l0.029,0.02l0.029,0.019l0.029,0.018l0.03,0.018l0.031,0.017l0.03,0.016l0.031,0.016l0.032,0.014l0.032,0.014l0.032,0.013l0.032,0.013l0.033,0.011l0.033,0.011l0.034,0.01l0.033,0.009l0.035,0.009l0.034,0.007l0.035,0.007l0.036,0.006l0.035,0.005l0.037,0.004l0.036,0.003l0.037,0.003l0.038,0.001l0.054,0.001l1.569,0l0.006,2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M13.856,17.518l0.096,-0.003l0.096,-0.006l0.094,-0.009l0.094,-0.01l0.093,-0.014l0.094,-0.015l0.091,-0.017l0.091,-0.02l0.09,-0.022l0.089,-0.024l0.089,-0.027l0.087,-0.029l0.087,-0.03l0.085,-0.032l0.085,-0.034l0.083,-0.038l0.082,-0.038l0.082,-0.04l0.079,-0.042l0.079,-0.045l0.078,-0.045l0.077,-0.049l0.075,-0.049l0.074,-0.052l0.073,-0.053l0.072,-0.054l0.07,-0.057l0.068,-0.058l0.068,-0.06l0.065,-0.06l0.065,-0.063l0.063,-0.065l0.06,-0.065l0.06,-0.068l0.058,-0.068l0.056,-0.07l0.055,-0.072l0.053,-0.072l0.052,-0.074l0.049,-0.076l0.048,-0.076l0.046,-0.078l0.044,-0.079l0.043,-0.08l0.039,-0.081l0.039,-0.083l0.037,-0.083l0.035,-0.085l0.032,-0.085l0.03,-0.087l0.029,-0.087l0.026,-0.089l0.024,-0.089l0.022,-0.089l0.02,-0.091l0.018,-0.092l0.015,-0.093l0.014,-0.093l0.01,-0.094l0.009,-0.095l0.005,-0.095l0.004,-0.096l0.001,-0.096l0,-1.043l-5.125,4.786l1.382,0l0.096,-0.002Zm-1.478,0.002l5.125,-4.786l-0.95,0.003l0,1.034l-0.001,0.078l-0.003,0.072l-0.004,0.071l-0.006,0.071l-0.009,0.07l-0.009,0.07l-0.012,0.068l-0.013,0.069l-0.014,0.067l-0.016,0.068l-0.018,0.066l-0.02,0.066l-0.021,0.065l-0.023,0.064l-0.025,0.064l-0.025,0.063l-0.027,0.062l-0.028,0.062l-0.031,0.06l-0.031,0.059l-0.033,0.059l-0.035,0.058l-0.036,0.057l-0.037,0.057l-0.038,0.055l-0.04,0.054l-0.04,0.054l-0.042,0.052l-0.044,0.051l-0.045,0.05l-0.045,0.05l-0.048,0.048l-0.047,0.047l-0.05,0.045l-0.05,0.045l-0.051,0.044l-0.052,0.042l-0.054,0.041l-0.054,0.039l-0.056,0.039l-0.056,0.037l-0.057,0.035l-0.058,0.035l-0.059,0.033l-0.06,0.031l-0.06,0.031l-0.061,0.028l-0.063,0.028l-0.063,0.025l-0.063,0.024l-0.065,0.023l-0.064,0.021l-0.066,0.02l-0.066,0.018l-0.068,0.016l-0.067,0.015l-0.069,0.013l-0.069,0.012l-0.069,0.009l-0.071,0.008l-0.07,0.007l-0.071,0.004l-0.072,0.003l-0.079,0.001l-1.393,0l0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.bottomright = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M14.044,17.873l0.101,-0.003l0.101,-0.007l0.099,-0.009l0.099,-0.011l0.098,-0.014l0.098,-0.016l0.096,-0.018l0.096,-0.021l0.095,-0.023l0.094,-0.026l0.093,-0.028l0.092,-0.03l0.091,-0.032l0.09,-0.034l0.089,-0.036l0.088,-0.039l0.086,-0.04l0.086,-0.043l0.084,-0.044l0.083,-0.047l0.082,-0.048l0.081,-0.051l0.079,-0.052l0.078,-0.054l0.077,-0.056l0.075,-0.057l0.074,-0.06l0.072,-0.061l0.071,-0.063l0.069,-0.064l0.068,-0.066l0.066,-0.068l0.064,-0.069l0.063,-0.071l0.061,-0.072l0.059,-0.074l0.058,-0.075l0.056,-0.076l0.054,-0.078l0.052,-0.08l0.05,-0.08l0.049,-0.082l0.046,-0.083l0.045,-0.085l0.042,-0.085l0.041,-0.087l0.038,-0.088l0.037,-0.089l0.034,-0.09l0.032,-0.091l0.03,-0.092l0.028,-0.093l0.025,-0.094l0.023,-0.094l0.021,-0.096l0.019,-0.097l0.016,-0.097l0.014,-0.098l0.011,-0.099l0.009,-0.1l0.006,-0.1l0.004,-0.101l0.001,-0.102l0,-1.272l-5.531,5.213l1.591,0l0.101,-0.002Zm-1.692,0.002l5.531,-5.213l-2.5,0.012l0,1.244l0,0.054l-0.002,0.038l-0.002,0.037l-0.004,0.037l-0.004,0.036l-0.005,0.036l-0.006,0.035l-0.006,0.035l-0.008,0.035l-0.008,0.034l-0.009,0.034l-0.01,0.033l-0.011,0.033l-0.012,0.033l-0.012,0.033l-0.013,0.032l-0.014,0.032l-0.015,0.031l-0.015,0.031l-0.016,0.031l-0.017,0.03l-0.018,0.03l-0.018,0.03l-0.02,0.029l-0.019,0.028l-0.021,0.029l-0.021,0.027l-0.022,0.028l-0.023,0.026l-0.023,0.026l-0.024,0.026l-0.024,0.025l-0.025,0.025l-0.026,0.023l-0.026,0.024l-0.027,0.022l-0.027,0.022l-0.028,0.021l-0.028,0.021l-0.029,0.02l-0.029,0.019l-0.029,0.018l-0.03,0.018l-0.031,0.017l-0.03,0.016l-0.031,0.016l-0.032,0.014l-0.032,0.014l-0.032,0.013l-0.032,0.013l-0.033,0.011l-0.033,0.011l-0.034,0.01l-0.033,0.009l-0.035,0.009l-0.034,0.007l-0.035,0.007l-0.036,0.006l-0.035,0.005l-0.037,0.004l-0.036,0.003l-0.037,0.003l-0.038,0.001l-0.054,0.001l-1.57,0l-0.005,2.5Z"
-    fill="#555d66" />
-    <path d="M13.85,2.516l0.096,0.003l0.096,0.006l0.094,0.009l0.094,0.01l0.093,0.014l0.093,0.015l0.092,0.017l0.091,0.02l0.09,0.022l0.089,0.024l0.089,0.027l0.087,0.029l0.087,0.03l0.085,0.032l0.085,0.034l0.083,0.037l0.082,0.038l0.082,0.041l0.079,0.042l0.079,0.045l0.078,0.045l0.077,0.049l0.075,0.049l0.074,0.051l0.073,0.054l0.072,0.054l0.07,0.057l0.068,0.058l0.068,0.06l0.065,0.06l0.065,0.063l0.062,0.065l0.061,0.065l0.06,0.068l0.058,0.068l0.056,0.07l0.055,0.072l0.053,0.072l0.052,0.074l0.049,0.076l0.048,0.076l0.046,0.078l0.044,0.079l0.043,0.08l0.039,0.081l0.039,0.083l0.036,0.083l0.036,0.085l0.032,0.085l0.03,0.087l0.029,0.087l0.026,0.088l0.024,0.09l0.022,0.089l0.02,0.091l0.018,0.092l0.015,0.093l0.013,0.093l0.011,0.094l0.008,0.095l0.006,0.095l0.004,0.096l0.001,0.096l0,1.043l-5.125,-4.786l1.382,0l0.096,0.002Zm-1.478,-0.002l5.125,4.786l-0.95,-0.004l0,-1.033l-0.001,-0.078l-0.003,-0.072l-0.004,-0.071l-0.006,-0.071l-0.009,-0.07l-0.009,-0.07l-0.012,-0.068l-0.013,-0.069l-0.014,-0.067l-0.016,-0.068l-0.018,-0.066l-0.02,-0.066l-0.021,-0.065l-0.023,-0.064l-0.025,-0.064l-0.025,-0.063l-0.027,-0.062l-0.028,-0.062l-0.031,-0.06l-0.031,-0.059l-0.033,-0.059l-0.035,-0.058l-0.036,-0.057l-0.037,-0.057l-0.038,-0.055l-0.04,-0.054l-0.041,-0.054l-0.041,-0.052l-0.044,-0.051l-0.045,-0.051l-0.045,-0.049l-0.048,-0.048l-0.047,-0.047l-0.05,-0.045l-0.05,-0.045l-0.051,-0.044l-0.053,-0.042l-0.053,-0.041l-0.054,-0.039l-0.056,-0.039l-0.056,-0.037l-0.057,-0.035l-0.058,-0.035l-0.059,-0.033l-0.06,-0.031l-0.06,-0.031l-0.061,-0.028l-0.063,-0.028l-0.063,-0.025l-0.063,-0.024l-0.065,-0.023l-0.064,-0.021l-0.066,-0.02l-0.066,-0.018l-0.068,-0.016l-0.067,-0.015l-0.069,-0.013l-0.069,-0.012l-0.069,-0.009l-0.071,-0.008l-0.07,-0.007l-0.071,-0.004l-0.072,-0.003l-0.079,-0.001l-1.393,0l0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.136,2.516l-0.096,0.003l-0.096,0.006l-0.094,0.009l-0.094,0.01l-0.093,0.014l-0.093,0.015l-0.091,0.017l-0.091,0.02l-0.09,0.022l-0.09,0.024l-0.088,0.027l-0.087,0.029l-0.087,0.03l-0.085,0.032l-0.085,0.034l-0.084,0.037l-0.081,0.038l-0.082,0.041l-0.08,0.042l-0.079,0.045l-0.077,0.045l-0.077,0.049l-0.075,0.049l-0.075,0.051l-0.073,0.054l-0.071,0.054l-0.07,0.057l-0.069,0.058l-0.067,0.06l-0.066,0.06l-0.064,0.063l-0.063,0.065l-0.061,0.065l-0.059,0.068l-0.058,0.068l-0.056,0.07l-0.056,0.072l-0.053,0.072l-0.051,0.074l-0.049,0.076l-0.048,0.076l-0.046,0.078l-0.044,0.079l-0.043,0.08l-0.04,0.081l-0.039,0.083l-0.036,0.083l-0.035,0.085l-0.032,0.085l-0.031,0.087l-0.028,0.087l-0.027,0.088l-0.024,0.09l-0.021,0.089l-0.02,0.091l-0.018,0.092l-0.016,0.093l-0.013,0.093l-0.01,0.094l-0.009,0.095l-0.006,0.095l-0.003,0.096l-0.001,0.096l0,1.043l5.124,-4.786l-1.382,0l-0.096,0.002Zm1.478,-0.002l-5.124,4.786l0.95,-0.004l0,-1.033l0.001,-0.078l0.002,-0.072l0.004,-0.071l0.007,-0.071l0.008,-0.07l0.01,-0.07l0.011,-0.068l0.014,-0.069l0.014,-0.067l0.016,-0.068l0.018,-0.066l0.02,-0.066l0.021,-0.065l0.023,-0.064l0.024,-0.064l0.026,-0.063l0.027,-0.062l0.028,-0.062l0.03,-0.06l0.032,-0.059l0.033,-0.059l0.034,-0.058l0.036,-0.057l0.037,-0.057l0.038,-0.055l0.04,-0.054l0.041,-0.054l0.042,-0.052l0.044,-0.051l0.044,-0.051l0.046,-0.049l0.047,-0.048l0.048,-0.047l0.049,-0.045l0.051,-0.045l0.051,-0.044l0.052,-0.042l0.053,-0.041l0.054,-0.039l0.057,-0.039l0.056,-0.037l0.057,-0.035l0.057,-0.035l0.059,-0.033l0.06,-0.031l0.061,-0.031l0.061,-0.028l0.063,-0.028l0.062,-0.025l0.064,-0.024l0.064,-0.023l0.065,-0.021l0.066,-0.02l0.066,-0.018l0.068,-0.016l0.067,-0.015l0.068,-0.013l0.07,-0.012l0.069,-0.009l0.07,-0.008l0.071,-0.007l0.071,-0.004l0.072,-0.003l0.079,-0.001l1.392,0l-0.017,-0.95Z"
-    fill="#555d66" />
-    <path d="M6.191,17.518l-0.096,-0.003l-0.096,-0.006l-0.094,-0.009l-0.094,-0.01l-0.093,-0.014l-0.093,-0.015l-0.091,-0.017l-0.092,-0.02l-0.09,-0.022l-0.089,-0.024l-0.088,-0.027l-0.088,-0.029l-0.086,-0.03l-0.086,-0.032l-0.084,-0.034l-0.084,-0.038l-0.082,-0.038l-0.081,-0.04l-0.08,-0.042l-0.079,-0.045l-0.078,-0.045l-0.077,-0.049l-0.075,-0.049l-0.074,-0.052l-0.073,-0.053l-0.071,-0.054l-0.071,-0.057l-0.068,-0.058l-0.067,-0.06l-0.066,-0.06l-0.064,-0.063l-0.063,-0.065l-0.061,-0.065l-0.06,-0.068l-0.058,-0.068l-0.056,-0.07l-0.055,-0.072l-0.053,-0.072l-0.051,-0.074l-0.05,-0.076l-0.047,-0.076l-0.047,-0.078l-0.044,-0.079l-0.042,-0.08l-0.04,-0.081l-0.039,-0.083l-0.036,-0.083l-0.035,-0.085l-0.033,-0.085l-0.03,-0.087l-0.029,-0.087l-0.026,-0.089l-0.024,-0.089l-0.022,-0.089l-0.02,-0.091l-0.018,-0.092l-0.015,-0.093l-0.013,-0.093l-0.011,-0.094l-0.008,-0.095l-0.006,-0.095l-0.004,-0.096l-0.001,-0.096l0,-1.043l5.125,4.786l-1.382,0l-0.096,-0.002Zm1.478,0.002l-5.125,-4.786l0.95,0.003l0,1.034l0.001,0.078l0.003,0.072l0.004,0.071l0.006,0.071l0.009,0.07l0.01,0.07l0.011,0.068l0.013,0.069l0.014,0.067l0.017,0.068l0.018,0.066l0.02,0.066l0.021,0.065l0.022,0.064l0.025,0.064l0.026,0.063l0.026,0.062l0.029,0.062l0.03,0.06l0.031,0.059l0.034,0.059l0.034,0.058l0.036,0.057l0.037,0.057l0.038,0.055l0.04,0.054l0.041,0.054l0.042,0.052l0.043,0.051l0.045,0.05l0.046,0.05l0.047,0.048l0.048,0.047l0.049,0.045l0.05,0.045l0.052,0.044l0.052,0.042l0.053,0.041l0.054,0.039l0.056,0.039l0.056,0.037l0.057,0.035l0.058,0.035l0.059,0.033l0.06,0.031l0.061,0.031l0.06,0.028l0.063,0.028l0.063,0.025l0.063,0.024l0.065,0.023l0.065,0.021l0.065,0.02l0.067,0.018l0.067,0.016l0.068,0.015l0.068,0.013l0.069,0.012l0.07,0.009l0.07,0.008l0.07,0.007l0.071,0.004l0.073,0.003l0.079,0.001l1.392,0l-0.017,0.95Z"
-    fill="#555d66" />
-</svg>;
-icons.h1 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M12.993,9.914c0.232,-0.096 0.474,-0.206 0.726,-0.33c0.252,-0.124 0.498,-0.26 0.738,-0.408c0.24,-0.148 0.47,-0.304 0.69,-0.468c0.22,-0.164 0.418,-0.338 0.594,-0.522l1.248,0l0,8.316l-1.788,0l0,-5.988c-0.24,0.16 -0.508,0.306 -0.804,0.438c-0.296,0.132 -0.584,0.246 -0.864,0.342l-0.54,-1.38Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h2 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-	<path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-	/>
-	<path d="M18.019,10.342c0,0.304 -0.06,0.596 -0.18,0.876c-0.12,0.28 -0.276,0.55 -0.468,0.81c-0.192,0.26 -0.408,0.51 -0.648,0.75c-0.24,0.24 -0.476,0.468 -0.708,0.684c-0.12,0.112 -0.25,0.238 -0.39,0.378c-0.14,0.14 -0.274,0.282 -0.402,0.426c-0.128,0.144 -0.242,0.278 -0.342,0.402c-0.1,0.124 -0.162,0.226 -0.186,0.306l3.576,0l0,1.5l-5.544,0c-0.016,-0.088 -0.024,-0.2 -0.024,-0.336c0,-0.136 0,-0.232 0,-0.288c0,-0.384 0.062,-0.736 0.186,-1.056c0.124,-0.32 0.286,-0.618 0.486,-0.894c0.2,-0.276 0.426,-0.534 0.678,-0.774c0.252,-0.24 0.502,-0.48 0.75,-0.72c0.192,-0.184 0.372,-0.358 0.54,-0.522c0.168,-0.164 0.316,-0.324 0.444,-0.48c0.128,-0.156 0.228,-0.312 0.3,-0.468c0.072,-0.156 0.108,-0.314 0.108,-0.474c0,-0.352 -0.1,-0.6 -0.3,-0.744c-0.2,-0.144 -0.448,-0.216 -0.744,-0.216c-0.216,0 -0.418,0.034 -0.606,0.102c-0.188,0.068 -0.36,0.148 -0.516,0.24c-0.156,0.092 -0.29,0.184 -0.402,0.276c-0.112,0.092 -0.196,0.166 -0.252,0.222l-0.888,-1.248c0.352,-0.328 0.762,-0.598 1.23,-0.81c0.468,-0.212 0.97,-0.318 1.506,-0.318c0.488,0 0.908,0.056 1.26,0.168c0.352,0.112 0.642,0.27 0.87,0.474c0.228,0.204 0.396,0.452 0.504,0.744c0.108,0.292 0.162,0.622 0.162,0.99Z"
-	/>
-</svg>;
-icons.h3 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M14.884,16.485c-0.216,0 -0.444,-0.014 -0.684,-0.042c-0.24,-0.028 -0.472,-0.066 -0.696,-0.114c-0.224,-0.048 -0.428,-0.1 -0.612,-0.156c-0.184,-0.056 -0.328,-0.108 -0.432,-0.156l0.348,-1.488c0.208,0.088 0.474,0.182 0.798,0.282c0.324,0.1 0.726,0.15 1.206,0.15c0.552,0 0.956,-0.104 1.212,-0.312c0.256,-0.208 0.384,-0.488 0.384,-0.84c0,-0.216 -0.046,-0.398 -0.138,-0.546c-0.092,-0.148 -0.218,-0.268 -0.378,-0.36c-0.16,-0.092 -0.35,-0.156 -0.57,-0.192c-0.22,-0.036 -0.454,-0.054 -0.702,-0.054l-0.696,0l0,-1.44l0.792,0c0.176,0 0.346,-0.016 0.51,-0.048c0.164,-0.032 0.31,-0.086 0.438,-0.162c0.128,-0.076 0.23,-0.18 0.306,-0.312c0.076,-0.132 0.114,-0.298 0.114,-0.498c0,-0.152 -0.032,-0.284 -0.096,-0.396c-0.064,-0.112 -0.146,-0.204 -0.246,-0.276c-0.1,-0.072 -0.216,-0.126 -0.348,-0.162c-0.132,-0.036 -0.266,-0.054 -0.402,-0.054c-0.344,0 -0.662,0.052 -0.954,0.156c-0.292,0.104 -0.558,0.232 -0.798,0.384l-0.636,-1.308c0.128,-0.08 0.278,-0.164 0.45,-0.252c0.172,-0.088 0.362,-0.168 0.57,-0.24c0.208,-0.072 0.43,-0.132 0.666,-0.18c0.236,-0.048 0.486,-0.072 0.75,-0.072c0.488,0 0.91,0.058 1.266,0.174c0.356,0.116 0.65,0.28 0.882,0.492c0.232,0.212 0.404,0.46 0.516,0.744c0.112,0.284 0.168,0.594 0.168,0.93c0,0.328 -0.092,0.646 -0.276,0.954c-0.184,0.308 -0.432,0.542 -0.744,0.702c0.432,0.176 0.766,0.438 1.002,0.786c0.236,0.348 0.354,0.766 0.354,1.254c0,0.384 -0.064,0.738 -0.192,1.062c-0.128,0.324 -0.328,0.604 -0.6,0.84c-0.272,0.236 -0.618,0.42 -1.038,0.552c-0.42,0.132 -0.918,0.198 -1.494,0.198Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h4 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M15.966,10.329c-0.312,0.408 -0.644,0.852 -0.996,1.332c-0.352,0.48 -0.648,0.972 -0.888,1.476l1.884,0l0,-2.808Zm1.752,-2.184l0,4.992l0.888,0l0,1.464l-0.888,0l0,1.86l-1.752,0l0,-1.86l-3.588,0l0,-1.308c0.176,-0.352 0.396,-0.75 0.66,-1.194c0.264,-0.444 0.558,-0.9 0.882,-1.368c0.324,-0.468 0.666,-0.926 1.026,-1.374c0.36,-0.448 0.72,-0.852 1.08,-1.212l1.692,0Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h5 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M16.468,13.857c0,-0.232 -0.052,-0.438 -0.156,-0.618c-0.104,-0.18 -0.284,-0.33 -0.54,-0.45c-0.256,-0.12 -0.6,-0.212 -1.032,-0.276c-0.432,-0.064 -0.976,-0.096 -1.632,-0.096c0.088,-0.744 0.16,-1.5 0.216,-2.268c0.056,-0.768 0.1,-1.488 0.132,-2.16l4.5,0l0,1.5l-3,0c-0.016,0.288 -0.034,0.566 -0.054,0.834c-0.02,0.268 -0.042,0.502 -0.066,0.702c1.168,0.08 2.03,0.354 2.586,0.822c0.556,0.468 0.834,1.118 0.834,1.95c0,0.384 -0.068,0.74 -0.204,1.068c-0.136,0.328 -0.34,0.612 -0.612,0.852c-0.272,0.24 -0.614,0.428 -1.026,0.564c-0.412,0.136 -0.894,0.204 -1.446,0.204c-0.216,0 -0.442,-0.014 -0.678,-0.042c-0.236,-0.028 -0.464,-0.064 -0.684,-0.108c-0.22,-0.044 -0.418,-0.09 -0.594,-0.138c-0.176,-0.048 -0.312,-0.096 -0.408,-0.144l0.324,-1.476c0.2,0.088 0.464,0.174 0.792,0.258c0.328,0.084 0.72,0.126 1.176,0.126c0.552,0 0.952,-0.11 1.2,-0.33c0.248,-0.22 0.372,-0.478 0.372,-0.774Z"
-    fillRule="nonzero" />
-</svg>;
-icons.h6 = <svg width="20px" height="20px" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M9.229,3.5l0,5.2l-5,0l0,-5.2l-2.5,0l0,13l2.5,0l0,-5.2l5,0l0,5.2l2.5,0l0,-13l-2.5,0Z"
-    fillRule="nonzero" />
-    <path d="M15.436,12.393c-0.176,0 -0.364,0.016 -0.564,0.048c-0.2,0.032 -0.356,0.076 -0.468,0.132c0,0.032 -0.002,0.08 -0.006,0.144c-0.004,0.064 -0.006,0.12 -0.006,0.168c0,0.288 0.02,0.56 0.06,0.816c0.04,0.256 0.108,0.48 0.204,0.672c0.096,0.192 0.224,0.342 0.384,0.45c0.16,0.108 0.36,0.162 0.6,0.162c0.2,0 0.37,-0.042 0.51,-0.126c0.14,-0.084 0.256,-0.19 0.348,-0.318c0.092,-0.128 0.16,-0.27 0.204,-0.426c0.044,-0.156 0.066,-0.306 0.066,-0.45c0,-0.408 -0.102,-0.722 -0.306,-0.942c-0.204,-0.22 -0.546,-0.33 -1.026,-0.33Zm0.228,-1.392c0.52,0 0.962,0.07 1.326,0.21c0.364,0.14 0.662,0.33 0.894,0.57c0.232,0.24 0.4,0.52 0.504,0.84c0.104,0.32 0.156,0.66 0.156,1.02c0,0.32 -0.058,0.648 -0.174,0.984c-0.116,0.336 -0.292,0.64 -0.528,0.912c-0.236,0.272 -0.536,0.496 -0.9,0.672c-0.364,0.176 -0.794,0.264 -1.29,0.264c-1,0 -1.76,-0.316 -2.28,-0.948c-0.52,-0.632 -0.78,-1.504 -0.78,-2.616c0,-0.8 0.124,-1.508 0.372,-2.124c0.248,-0.616 0.602,-1.134 1.062,-1.554c0.46,-0.42 1.016,-0.74 1.668,-0.96c0.652,-0.22 1.386,-0.334 2.202,-0.342c0.016,0.248 0.032,0.49 0.048,0.726c0.016,0.236 0.032,0.482 0.048,0.738c-0.408,0.008 -0.79,0.046 -1.146,0.114c-0.356,0.068 -0.68,0.174 -0.972,0.318c-0.292,0.144 -0.546,0.33 -0.762,0.558c-0.216,0.228 -0.384,0.506 -0.504,0.834c0.176,-0.08 0.356,-0.136 0.54,-0.168c0.184,-0.032 0.356,-0.048 0.516,-0.048Z"
-    fillRule="nonzero" />
-</svg>;
-icons.infobox = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fill='#0058b0' fillRule='nonzero'>
-        <path d='M24,26.945c0.001,0 0.002,0 0.003,0c4.371,0 7.968,-3.597 7.968,-7.969c0,-4.371 -3.597,-7.968 -7.968,-7.968c-4.372,0 -7.969,3.597 -7.969,7.968c0,0.001 0,0.002 0,0.003c0,4.37 3.596,7.966 7.966,7.966Zm-4.098,-9.541l2.057,-0.297c0.218,-0.036 0.406,-0.175 0.504,-0.374l0.919,-1.862c0.115,-0.232 0.352,-0.379 0.61,-0.379c0.259,0 0.496,0.147 0.611,0.379l0.918,1.862c0.098,0.199 0.287,0.338 0.505,0.374l2.057,0.297c0.333,0.049 0.582,0.338 0.582,0.675c0,0.183 -0.073,0.358 -0.204,0.486l-1.489,1.451c-0.159,0.157 -0.231,0.382 -0.194,0.603l0.351,2.049c0.006,0.037 0.009,0.075 0.009,0.114c0,0.373 -0.307,0.681 -0.681,0.681c-0.111,0 -0.219,-0.027 -0.317,-0.079l-1.837,-0.966c-0.199,-0.103 -0.437,-0.103 -0.636,0l-1.837,0.966c-0.098,0.052 -0.207,0.079 -0.317,0.079c-0.374,0 -0.682,-0.308 -0.682,-0.681c0,-0.039 0.004,-0.077 0.01,-0.114l0.351,-2.049c0.037,-0.221 -0.036,-0.446 -0.195,-0.603l-1.488,-1.451c-0.13,-0.128 -0.203,-0.303 -0.203,-0.485c0,-0.341 0.257,-0.633 0.596,-0.676Z'
-        />
-        <path d='M46.529,6.719l-45.058,0c-0.485,0 -0.884,0.398 -0.884,0.883l0,32.796c0,0.485 0.399,0.883 0.884,0.883l45.058,0c0.485,0 0.884,-0.398 0.884,-0.883l0,-32.801c-0.003,-0.483 -0.401,-0.878 -0.884,-0.878Zm-0.883,32.801l-43.292,0l0,-31.035l43.292,0l0,31.035Z'
-        />
-        <path d='M29.659,30.091l-11.318,0c-0.485,0 -0.884,0.399 -0.884,0.883c0,0.485 0.399,0.883 0.884,0.883l11.318,0c0.485,0 0.884,-0.398 0.884,-0.883c0,-0.484 -0.399,-0.883 -0.884,-0.883Z'
-        />
-        <path d='M29.425,34.769l-10.85,0c-0.484,0 -0.883,0.399 -0.883,0.883c0,0.485 0.399,0.883 0.883,0.883l10.85,0c0.484,0 0.883,-0.398 0.883,-0.883c0,-0.484 -0.399,-0.883 -0.883,-0.883Z'
-        />
-    </g>
-</svg>;
-icons.accordionBlock = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fill='#0058b0'>
-        <path d='M46.524,6.723l-45.048,0c-0.485,0 -0.883,0.398 -0.883,0.883l0,32.788c0,0.485 0.398,0.883 0.883,0.883l45.048,0c0.485,0 0.883,-0.398 0.883,-0.883l0,-32.793c-0.002,-0.483 -0.4,-0.878 -0.883,-0.878Zm-17.451,3.985c0,0.485 -0.399,0.883 -0.883,0.883l-18.249,0c-0.484,0 -0.882,-0.398 -0.882,-0.883c0,-0.484 0.398,-0.883 0.882,-0.883l18.249,0c0.484,0 0.883,0.399 0.883,0.883Zm-21.846,0c0,0.678 -0.558,1.236 -1.236,1.236c-0.678,0 -1.236,-0.558 -1.236,-1.236c0,-0.678 0.558,-1.236 1.236,-1.236c0.678,0 1.236,0.558 1.236,1.236Zm38.414,3.784l0,15.893l-43.282,0l0,-15.893l43.282,0Zm0,25.009l-43.282,0l0,-7.35l43.282,0l0,7.35Z'
-        fillRule='nonzero' />
-        <path d='M7.578,20.774l32.476,0.04c0.484,0 0.883,-0.399 0.883,-0.883c0,-0.484 -0.399,-0.883 -0.883,-0.883l-32.474,-0.04c-0.484,0 -0.883,0.398 -0.883,0.883c0,0.484 0.399,0.883 0.883,0.883l-0.002,0Z'
-        fillRule='nonzero' />
-        <path d='M40.046,24.194l-32.466,-0.04c-0.484,0 -0.883,0.398 -0.883,0.883c0,0.484 0.399,0.883 0.883,0.883l32.466,0.04c0.485,0 0.883,-0.398 0.883,-0.883c0,-0.484 -0.398,-0.883 -0.883,-0.883Z'
-        fillRule='nonzero' />
-        <path d='M10.925,36.734l14.942,0c0.484,0 0.883,-0.398 0.883,-0.883c0,-0.484 -0.399,-0.883 -0.883,-0.883l-14.942,0c-0.484,0 -0.883,0.399 -0.883,0.883c0,0.485 0.399,0.883 0.883,0.883Z'
-        fillRule='nonzero' />
-        <circle cx='6.072' cy='35.851' r='1.236' />
-    </g>
-</svg>;
-icons.tabsSimple = <svg viewBox="0 0 400 300" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M170.032,39.692l0.085,0.001l0.083,0.001l0.084,0.003l0.083,0.004l0.084,0.005l0.082,0.005l0.083,0.007l0.082,0.008l0.082,0.009l0.082,0.01l0.081,0.011l0.081,0.011l0.081,0.013l0.08,0.014l0.08,0.015l0.08,0.016l0.079,0.016l0.079,0.018l0.079,0.019l0.078,0.02l0.078,0.02l0.077,0.022l0.077,0.022l0.077,0.023l0.076,0.025l0.076,0.025l0.075,0.026l0.075,0.027l0.075,0.028l0.074,0.028l0.074,0.03l0.073,0.031l0.073,0.031l0.072,0.032l0.072,0.033l0.072,0.034l0.071,0.035l0.07,0.035l0.07,0.037l0.07,0.037l0.069,0.038l0.068,0.039l0.068,0.04l0.067,0.04l0.067,0.041l0.067,0.042l0.065,0.043l0.066,0.043l0.064,0.045l0.065,0.045l0.063,0.046l0.063,0.046l0.063,0.047l0.061,0.048l0.062,0.049l0.06,0.05l0.06,0.05l0.06,0.051l0.059,0.051l0.058,0.052l0.057,0.053l0.057,0.054l0.056,0.054l0.056,0.055l0.055,0.056l0.054,0.056l0.054,0.057l0.053,0.058l0.052,0.058l0.052,0.059l0.05,0.059l0.051,0.06l0.049,0.061l0.049,0.061l0.048,0.062l0.047,0.062l0.047,0.063l0.045,0.064l0.045,0.064l0.045,0.065l0.043,0.065l0.043,0.066l0.042,0.066l0.041,0.067l0.041,0.067l0.039,0.068l0.039,0.069l0.038,0.069l0.037,0.069l0.037,0.07l0.035,0.071l0.035,0.071l0.034,0.071l0.033,0.072l0.032,0.072l0.032,0.073l0.03,0.073l0.03,0.074l0.029,0.074l0.027,0.075l0.027,0.075l0.026,0.075l0.026,0.076l0.024,0.076l0.023,0.077l0.023,0.077l0.021,0.078l0.021,0.078l0.019,0.078l0.019,0.078l0.018,0.079l0.017,0.08l0.015,0.079l0.015,0.08l0.014,0.081l0.013,0.08l0.012,0.081l0.011,0.082l0.009,0.081l0.009,0.082l0.008,0.082l0.007,0.083l0.006,0.083l0.004,0.083l0.004,0.083l0.003,0.084l0.001,0.084l0.001,0.084l0,14.143l-23.704,0.019l-50.727,-0.028l0,-14.229l0.001,-0.085l0.001,-0.083l0.003,-0.084l0.004,-0.083l0.005,-0.084l0.005,-0.082l0.007,-0.083l0.008,-0.082l0.009,-0.082l0.01,-0.082l0.011,-0.081l0.011,-0.081l0.013,-0.081l0.014,-0.08l0.015,-0.08l0.016,-0.08l0.017,-0.079l0.017,-0.079l0.019,-0.078l0.02,-0.079l0.02,-0.078l0.022,-0.077l0.022,-0.077l0.023,-0.077l0.025,-0.076l0.025,-0.076l0.026,-0.075l0.027,-0.075l0.028,-0.075l0.029,-0.074l0.029,-0.074l0.031,-0.073l0.031,-0.073l0.032,-0.072l0.033,-0.072l0.034,-0.072l0.035,-0.071l0.035,-0.07l0.037,-0.07l0.037,-0.07l0.038,-0.069l0.039,-0.068l0.04,-0.068l0.04,-0.067l0.041,-0.067l0.042,-0.067l0.043,-0.065l0.044,-0.066l0.044,-0.064l0.045,-0.065l0.046,-0.063l0.046,-0.063l0.048,-0.063l0.048,-0.061l0.048,-0.062l0.05,-0.06l0.05,-0.06l0.051,-0.06l0.051,-0.059l0.053,-0.058l0.052,-0.057l0.054,-0.057l0.054,-0.056l0.055,-0.056l0.056,-0.055l0.056,-0.054l0.057,-0.054l0.058,-0.053l0.058,-0.052l0.059,-0.052l0.059,-0.05l0.06,-0.051l0.061,-0.049l0.061,-0.049l0.062,-0.048l0.062,-0.047l0.063,-0.047l0.064,-0.045l0.064,-0.045l0.065,-0.045l0.065,-0.043l0.066,-0.043l0.066,-0.042l0.067,-0.041l0.067,-0.041l0.068,-0.039l0.069,-0.039l0.069,-0.038l0.069,-0.037l0.07,-0.037l0.071,-0.035l0.071,-0.035l0.071,-0.034l0.072,-0.033l0.072,-0.032l0.073,-0.032l0.073,-0.03l0.074,-0.03l0.074,-0.029l0.075,-0.027l0.075,-0.027l0.075,-0.026l0.076,-0.026l0.077,-0.024l0.076,-0.023l0.077,-0.023l0.078,-0.021l0.078,-0.021l0.078,-0.019l0.078,-0.019l0.079,-0.018l0.08,-0.016l0.079,-0.016l0.08,-0.015l0.081,-0.014l0.08,-0.013l0.081,-0.012l0.082,-0.011l0.081,-0.009l0.082,-0.009l0.082,-0.008l0.083,-0.007l0.083,-0.006l0.083,-0.004l0.083,-0.004l0.084,-0.003l0.084,-0.001l0.084,-0.001l61.401,0.095Zm-61.473,0.906l-0.071,0.001l-0.071,0.002l-0.07,0.003l-0.071,0.004l-0.07,0.005l-0.07,0.006l-0.07,0.007l-0.069,0.007l-0.069,0.008l-0.069,0.01l-0.068,0.01l-0.069,0.011l-0.068,0.011l-0.067,0.013l-0.068,0.013l-0.067,0.014l-0.066,0.015l-0.067,0.016l-0.066,0.017l-0.066,0.017l-0.065,0.018l-0.065,0.019l-0.065,0.02l-0.065,0.02l-0.064,0.022l-0.064,0.022l-0.063,0.023l-0.063,0.023l-0.063,0.024l-0.062,0.025l-0.062,0.026l-0.062,0.027l-0.061,0.027l-0.061,0.028l-0.06,0.029l-0.06,0.029l-0.06,0.03l-0.059,0.031l-0.059,0.031l-0.058,0.033l-0.058,0.033l-0.058,0.033l-0.057,0.034l-0.056,0.035l-0.056,0.036l-0.056,0.036l-0.055,0.037l-0.055,0.037l-0.055,0.039l-0.053,0.038l-0.054,0.04l-0.053,0.04l-0.052,0.04l-0.052,0.042l-0.051,0.042l-0.051,0.042l-0.051,0.043l-0.049,0.044l-0.05,0.044l-0.049,0.045l-0.048,0.046l-0.047,0.046l-0.048,0.046l-0.046,0.047l-0.046,0.048l-0.046,0.048l-0.044,0.049l-0.045,0.049l-0.043,0.05l-0.044,0.05l-0.042,0.051l-0.042,0.052l-0.041,0.052l-0.041,0.052l-0.04,0.053l-0.039,0.053l-0.039,0.054l-0.038,0.054l-0.038,0.055l-0.037,0.056l-0.036,0.055l-0.036,0.056l-0.034,0.057l-0.035,0.057l-0.033,0.058l-0.033,0.058l-0.032,0.058l-0.032,0.059l-0.031,0.059l-0.03,0.06l-0.029,0.06l-0.029,0.06l-0.028,0.061l-0.027,0.061l-0.026,0.062l-0.026,0.062l-0.025,0.062l-0.025,0.063l-0.023,0.063l-0.023,0.063l-0.022,0.064l-0.021,0.064l-0.021,0.065l-0.019,0.064l-0.019,0.065l-0.019,0.066l-0.017,0.066l-0.017,0.066l-0.015,0.066l-0.015,0.067l-0.015,0.067l-0.013,0.067l-0.012,0.068l-0.012,0.068l-0.011,0.068l-0.01,0.069l-0.009,0.069l-0.009,0.069l-0.007,0.069l-0.007,0.07l-0.006,0.07l-0.004,0.07l-0.004,0.07l-0.004,0.071l-0.002,0.071l-0.001,0.071l-0.001,0.075l0,13.227l49.727,0.027l22.704,-0.018l0,-13.141l-0.001,-0.075l-0.001,-0.071l-0.002,-0.071l-0.003,-0.071l-0.004,-0.07l-0.005,-0.07l-0.006,-0.07l-0.007,-0.07l-0.007,-0.069l-0.008,-0.069l-0.01,-0.069l-0.01,-0.069l-0.011,-0.068l-0.011,-0.068l-0.013,-0.067l-0.013,-0.068l-0.014,-0.067l-0.015,-0.067l-0.016,-0.066l-0.017,-0.066l-0.017,-0.066l-0.018,-0.065l-0.019,-0.065l-0.02,-0.065l-0.02,-0.065l-0.022,-0.064l-0.022,-0.064l-0.023,-0.063l-0.023,-0.063l-0.024,-0.063l-0.025,-0.062l-0.026,-0.062l-0.027,-0.062l-0.027,-0.061l-0.028,-0.061l-0.029,-0.06l-0.029,-0.06l-0.03,-0.06l-0.031,-0.059l-0.031,-0.059l-0.033,-0.058l-0.032,-0.058l-0.034,-0.058l-0.034,-0.057l-0.035,-0.056l-0.036,-0.056l-0.036,-0.056l-0.037,-0.056l-0.037,-0.054l-0.038,-0.055l-0.039,-0.054l-0.04,-0.053l-0.04,-0.053l-0.04,-0.052l-0.042,-0.052l-0.042,-0.051l-0.042,-0.051l-0.043,-0.051l-0.044,-0.049l-0.044,-0.05l-0.045,-0.049l-0.046,-0.048l-0.046,-0.048l-0.046,-0.047l-0.047,-0.046l-0.048,-0.046l-0.048,-0.046l-0.049,-0.045l-0.049,-0.044l-0.05,-0.044l-0.05,-0.043l-0.051,-0.042l-0.052,-0.042l-0.052,-0.041l-0.052,-0.041l-0.053,-0.04l-0.053,-0.039l-0.054,-0.039l-0.054,-0.038l-0.055,-0.038l-0.055,-0.037l-0.056,-0.036l-0.056,-0.036l-0.057,-0.034l-0.057,-0.035l-0.058,-0.033l-0.058,-0.033l-0.058,-0.032l-0.059,-0.032l-0.059,-0.031l-0.06,-0.03l-0.06,-0.029l-0.06,-0.029l-0.061,-0.028l-0.061,-0.027l-0.062,-0.027l-0.062,-0.025l-0.062,-0.025l-0.063,-0.025l-0.063,-0.023l-0.063,-0.023l-0.064,-0.022l-0.064,-0.021l-0.064,-0.021l-0.065,-0.02l-0.065,-0.019l-0.066,-0.018l-0.066,-0.017l-0.066,-0.017l-0.066,-0.015l-0.067,-0.015l-0.067,-0.015l-0.067,-0.013l-0.068,-0.012l-0.068,-0.012l-0.068,-0.011l-0.069,-0.01l-0.069,-0.009l-0.069,-0.009l-0.069,-0.007l-0.07,-0.007l-0.07,-0.006l-0.07,-0.004l-0.07,-0.005l-0.071,-0.003l-0.071,-0.002l-0.071,-0.001l-0.075,-0.001l-61.396,-0.095l-0.074,0.001Z"
-    fill="#b1b1b1" />
-    <path d="M252.783,39.729l0.084,0l0.084,0.002l0.084,0.002l0.083,0.004l0.083,0.005l0.083,0.006l0.083,0.007l0.082,0.007l0.082,0.009l0.082,0.01l0.081,0.011l0.081,0.012l0.081,0.013l0.08,0.014l0.08,0.014l0.08,0.016l0.079,0.017l0.079,0.018l0.078,0.018l0.079,0.02l0.077,0.021l0.078,0.021l0.077,0.022l0.077,0.024l0.076,0.024l0.076,0.025l0.075,0.026l0.075,0.027l0.075,0.028l0.074,0.029l0.074,0.029l0.073,0.031l0.073,0.031l0.072,0.033l0.072,0.033l0.072,0.034l0.071,0.034l0.07,0.036l0.07,0.036l0.069,0.037l0.069,0.038l0.069,0.039l0.068,0.04l0.067,0.04l0.067,0.042l0.067,0.042l0.065,0.042l0.066,0.044l0.064,0.044l0.064,0.045l0.064,0.046l0.063,0.047l0.062,0.047l0.062,0.048l0.062,0.049l0.06,0.049l0.06,0.05l0.06,0.051l0.058,0.052l0.059,0.052l0.057,0.053l0.057,0.053l0.056,0.055l0.056,0.055l0.055,0.055l0.054,0.057l0.054,0.056l0.053,0.058l0.052,0.058l0.051,0.059l0.051,0.059l0.05,0.06l0.05,0.061l0.049,0.061l0.048,0.062l0.047,0.063l0.046,0.063l0.046,0.063l0.045,0.064l0.045,0.065l0.043,0.065l0.043,0.066l0.042,0.066l0.041,0.067l0.041,0.068l0.039,0.068l0.039,0.068l0.038,0.069l0.037,0.07l0.037,0.07l0.035,0.07l0.035,0.071l0.034,0.071l0.033,0.072l0.032,0.073l0.032,0.072l0.03,0.074l0.03,0.073l0.028,0.075l0.028,0.074l0.027,0.075l0.026,0.076l0.025,0.076l0.025,0.076l0.023,0.077l0.023,0.077l0.021,0.077l0.021,0.078l0.019,0.078l0.019,0.079l0.018,0.079l0.016,0.079l0.016,0.08l0.015,0.08l0.014,0.08l0.013,0.081l0.012,0.081l0.01,0.081l0.01,0.082l0.009,0.082l0.008,0.082l0.007,0.082l0.006,0.083l0.004,0.083l0.004,0.084l0.003,0.083l0.001,0.084l0.001,0.084l0,14.143l-23.704,0.019l-50.727,-0.027l0,-14.23l0.001,-0.084l0.001,-0.084l0.003,-0.084l0.004,-0.083l0.004,-0.083l0.006,-0.083l0.007,-0.082l0.008,-0.083l0.009,-0.082l0.01,-0.081l0.01,-0.082l0.012,-0.081l0.013,-0.08l0.014,-0.081l0.015,-0.08l0.016,-0.079l0.016,-0.08l0.018,-0.078l0.019,-0.079l0.019,-0.078l0.021,-0.078l0.022,-0.078l0.022,-0.077l0.023,-0.076l0.025,-0.077l0.025,-0.075l0.026,-0.076l0.027,-0.075l0.028,-0.075l0.028,-0.074l0.03,-0.073l0.03,-0.074l0.032,-0.073l0.032,-0.072l0.033,-0.072l0.034,-0.071l0.035,-0.071l0.035,-0.071l0.037,-0.07l0.037,-0.069l0.038,-0.069l0.039,-0.069l0.039,-0.068l0.041,-0.067l0.041,-0.067l0.042,-0.066l0.043,-0.066l0.043,-0.065l0.045,-0.065l0.045,-0.064l0.046,-0.064l0.046,-0.063l0.047,-0.062l0.048,-0.062l0.049,-0.061l0.05,-0.061l0.05,-0.06l0.051,-0.059l0.051,-0.059l0.052,-0.058l0.053,-0.058l0.054,-0.057l0.054,-0.056l0.055,-0.056l0.056,-0.054l0.056,-0.055l0.057,-0.053l0.057,-0.053l0.059,-0.053l0.058,-0.051l0.06,-0.051l0.06,-0.05l0.061,-0.05l0.061,-0.048l0.062,-0.048l0.062,-0.048l0.063,-0.046l0.064,-0.046l0.064,-0.045l0.064,-0.044l0.066,-0.044l0.066,-0.043l0.066,-0.042l0.067,-0.041l0.067,-0.04l0.068,-0.04l0.069,-0.039l0.069,-0.038l0.069,-0.037l0.07,-0.036l0.071,-0.036l0.07,-0.035l0.072,-0.034l0.072,-0.033l0.072,-0.032l0.073,-0.031l0.073,-0.031l0.074,-0.029l0.074,-0.029l0.075,-0.028l0.075,-0.027l0.075,-0.026l0.076,-0.025l0.076,-0.024l0.077,-0.024l0.077,-0.022l0.078,-0.022l0.077,-0.02l0.079,-0.02l0.078,-0.019l0.079,-0.017l0.08,-0.017l0.079,-0.016l0.08,-0.015l0.081,-0.014l0.08,-0.012l0.081,-0.012l0.081,-0.011l0.082,-0.01l0.082,-0.009l0.082,-0.008l0.083,-0.007l0.083,-0.005l0.083,-0.005l0.083,-0.004l0.084,-0.002l0.084,-0.002l0.084,-0.001l61.401,0.096Zm-61.473,0.905l-0.071,0.001l-0.071,0.003l-0.071,0.003l-0.07,0.004l-0.07,0.005l-0.07,0.005l-0.07,0.007l-0.069,0.008l-0.069,0.008l-0.069,0.009l-0.069,0.01l-0.068,0.011l-0.068,0.012l-0.067,0.012l-0.068,0.014l-0.067,0.014l-0.067,0.015l-0.066,0.016l-0.066,0.016l-0.066,0.018l-0.065,0.018l-0.066,0.019l-0.064,0.019l-0.065,0.021l-0.064,0.021l-0.064,0.022l-0.063,0.023l-0.063,0.024l-0.063,0.024l-0.062,0.025l-0.062,0.026l-0.062,0.026l-0.061,0.027l-0.061,0.028l-0.06,0.029l-0.06,0.029l-0.06,0.031l-0.059,0.03l-0.059,0.032l-0.058,0.032l-0.058,0.033l-0.058,0.033l-0.057,0.035l-0.056,0.035l-0.057,0.035l-0.055,0.036l-0.056,0.037l-0.054,0.038l-0.055,0.038l-0.054,0.039l-0.053,0.039l-0.053,0.04l-0.052,0.041l-0.052,0.041l-0.052,0.042l-0.05,0.043l-0.051,0.043l-0.05,0.043l-0.049,0.045l-0.049,0.045l-0.048,0.045l-0.048,0.046l-0.047,0.047l-0.046,0.047l-0.046,0.047l-0.046,0.049l-0.045,0.048l-0.044,0.05l-0.044,0.05l-0.043,0.05l-0.042,0.051l-0.042,0.051l-0.041,0.052l-0.041,0.052l-0.04,0.053l-0.04,0.054l-0.038,0.054l-0.038,0.054l-0.038,0.055l-0.037,0.055l-0.036,0.056l-0.036,0.056l-0.035,0.057l-0.034,0.057l-0.033,0.057l-0.033,0.058l-0.032,0.058l-0.032,0.059l-0.031,0.059l-0.03,0.06l-0.029,0.06l-0.029,0.06l-0.028,0.061l-0.027,0.062l-0.027,0.061l-0.025,0.062l-0.025,0.062l-0.025,0.063l-0.023,0.063l-0.023,0.064l-0.022,0.063l-0.021,0.064l-0.021,0.065l-0.02,0.065l-0.019,0.065l-0.018,0.065l-0.017,0.066l-0.017,0.066l-0.016,0.067l-0.014,0.066l-0.015,0.067l-0.013,0.068l-0.013,0.068l-0.011,0.067l-0.011,0.069l-0.01,0.068l-0.009,0.069l-0.009,0.069l-0.007,0.069l-0.007,0.07l-0.006,0.07l-0.005,0.07l-0.004,0.071l-0.003,0.07l-0.002,0.071l-0.001,0.072l-0.001,0.074l0,13.227l49.727,0.027l22.704,-0.018l0,-13.141l-0.001,-0.074l-0.001,-0.071l-0.002,-0.072l-0.003,-0.07l-0.004,-0.071l-0.005,-0.07l-0.006,-0.07l-0.007,-0.069l-0.007,-0.07l-0.009,-0.069l-0.009,-0.069l-0.01,-0.068l-0.011,-0.068l-0.011,-0.068l-0.013,-0.068l-0.013,-0.067l-0.014,-0.067l-0.015,-0.067l-0.016,-0.066l-0.017,-0.067l-0.017,-0.066l-0.018,-0.065l-0.019,-0.065l-0.02,-0.065l-0.021,-0.064l-0.021,-0.064l-0.022,-0.064l-0.023,-0.064l-0.023,-0.063l-0.025,-0.062l-0.025,-0.063l-0.025,-0.062l-0.027,-0.061l-0.027,-0.062l-0.028,-0.06l-0.029,-0.061l-0.029,-0.06l-0.03,-0.059l-0.031,-0.06l-0.032,-0.058l-0.032,-0.059l-0.033,-0.058l-0.033,-0.057l-0.034,-0.057l-0.035,-0.057l-0.036,-0.056l-0.036,-0.056l-0.037,-0.055l-0.037,-0.055l-0.039,-0.054l-0.038,-0.054l-0.04,-0.053l-0.04,-0.053l-0.041,-0.053l-0.041,-0.052l-0.042,-0.051l-0.042,-0.051l-0.043,-0.05l-0.044,-0.05l-0.044,-0.049l-0.045,-0.049l-0.046,-0.048l-0.046,-0.048l-0.046,-0.047l-0.047,-0.047l-0.048,-0.046l-0.048,-0.045l-0.049,-0.045l-0.049,-0.044l-0.05,-0.044l-0.051,-0.043l-0.05,-0.043l-0.052,-0.042l-0.052,-0.041l-0.052,-0.041l-0.053,-0.04l-0.053,-0.039l-0.054,-0.039l-0.055,-0.038l-0.054,-0.038l-0.056,-0.036l-0.055,-0.037l-0.057,-0.035l-0.056,-0.035l-0.057,-0.034l-0.058,-0.034l-0.058,-0.033l-0.058,-0.032l-0.059,-0.031l-0.059,-0.031l-0.06,-0.03l-0.06,-0.03l-0.06,-0.028l-0.061,-0.028l-0.061,-0.028l-0.062,-0.026l-0.062,-0.026l-0.062,-0.025l-0.063,-0.024l-0.063,-0.024l-0.063,-0.023l-0.064,-0.022l-0.064,-0.021l-0.065,-0.02l-0.064,-0.02l-0.066,-0.019l-0.065,-0.018l-0.066,-0.018l-0.066,-0.016l-0.066,-0.016l-0.067,-0.015l-0.067,-0.014l-0.068,-0.014l-0.067,-0.012l-0.068,-0.012l-0.068,-0.011l-0.069,-0.01l-0.069,-0.009l-0.069,-0.008l-0.069,-0.008l-0.07,-0.006l-0.07,-0.006l-0.07,-0.005l-0.07,-0.004l-0.071,-0.003l-0.071,-0.002l-0.071,-0.002l-0.076,0l-61.395,-0.096l-0.074,0.001Z"
-    fill="#b1b1b1" />
-    <path d="M380.009,259.376l-360,0l0,-200l360,0l0,200Zm-359,-199l0,198l358,0l0,-198l-358,0Z"
-    fill="#757575" />
-    <path d="M87.907,39.721l0.085,0l0.084,0.002l0.083,0.002l0.083,0.004l0.084,0.005l0.082,0.006l0.083,0.006l0.082,0.008l0.082,0.009l0.082,0.01l0.081,0.011l0.081,0.012l0.081,0.012l0.08,0.014l0.08,0.015l0.08,0.016l0.079,0.017l0.079,0.017l0.079,0.019l0.078,0.02l0.078,0.02l0.077,0.022l0.077,0.022l0.077,0.024l0.076,0.024l0.076,0.025l0.075,0.026l0.075,0.027l0.075,0.028l0.074,0.029l0.074,0.029l0.073,0.031l0.073,0.031l0.072,0.032l0.072,0.033l0.072,0.034l0.071,0.035l0.07,0.036l0.07,0.036l0.07,0.037l0.069,0.038l0.068,0.039l0.068,0.04l0.067,0.04l0.067,0.041l0.067,0.042l0.065,0.043l0.066,0.044l0.064,0.044l0.065,0.045l0.063,0.046l0.063,0.046l0.063,0.048l0.061,0.048l0.062,0.048l0.06,0.05l0.06,0.05l0.06,0.051l0.059,0.051l0.058,0.053l0.057,0.053l0.057,0.053l0.056,0.055l0.056,0.054l0.055,0.056l0.054,0.056l0.054,0.057l0.053,0.058l0.052,0.058l0.052,0.059l0.05,0.059l0.051,0.06l0.049,0.061l0.049,0.061l0.048,0.062l0.047,0.062l0.047,0.063l0.045,0.064l0.045,0.064l0.045,0.065l0.043,0.065l0.043,0.066l0.042,0.066l0.041,0.067l0.041,0.068l0.039,0.067l0.039,0.069l0.038,0.069l0.037,0.069l0.037,0.07l0.035,0.071l0.035,0.071l0.034,0.071l0.033,0.072l0.032,0.072l0.032,0.073l0.03,0.074l0.03,0.073l0.029,0.074l0.027,0.075l0.027,0.075l0.026,0.076l0.026,0.075l0.024,0.077l0.023,0.076l0.023,0.077l0.021,0.078l0.021,0.078l0.019,0.078l0.019,0.079l0.018,0.079l0.017,0.079l0.015,0.079l0.015,0.08l0.014,0.081l0.013,0.08l0.012,0.081l0.011,0.082l0.009,0.081l0.009,0.082l0.008,0.083l0.007,0.082l0.006,0.083l0.004,0.083l0.004,0.083l0.003,0.084l0.001,0.084l0.001,0.084l0,14.143l-23.704,0.019l-50.727,-0.027l0,-14.23l0.001,-0.084l0.001,-0.084l0.003,-0.084l0.004,-0.083l0.005,-0.083l0.005,-0.083l0.007,-0.083l0.008,-0.082l0.009,-0.082l0.01,-0.081l0.011,-0.082l0.011,-0.081l0.013,-0.08l0.014,-0.081l0.015,-0.08l0.016,-0.079l0.017,-0.08l0.017,-0.079l0.019,-0.078l0.02,-0.078l0.02,-0.078l0.022,-0.078l0.022,-0.077l0.023,-0.077l0.025,-0.076l0.025,-0.076l0.026,-0.075l0.027,-0.075l0.028,-0.075l0.029,-0.074l0.029,-0.074l0.031,-0.073l0.031,-0.073l0.032,-0.072l0.033,-0.072l0.034,-0.071l0.035,-0.071l0.035,-0.071l0.037,-0.07l0.037,-0.069l0.038,-0.069l0.039,-0.069l0.04,-0.068l0.04,-0.067l0.041,-0.067l0.042,-0.066l0.043,-0.066l0.044,-0.065l0.044,-0.065l0.045,-0.064l0.046,-0.064l0.046,-0.063l0.048,-0.062l0.048,-0.062l0.048,-0.061l0.05,-0.061l0.05,-0.06l0.051,-0.059l0.051,-0.059l0.053,-0.058l0.052,-0.058l0.054,-0.057l0.054,-0.056l0.055,-0.056l0.056,-0.055l0.056,-0.054l0.057,-0.054l0.058,-0.053l0.058,-0.052l0.059,-0.051l0.059,-0.051l0.06,-0.05l0.061,-0.05l0.061,-0.048l0.062,-0.049l0.062,-0.047l0.063,-0.046l0.064,-0.046l0.064,-0.045l0.065,-0.044l0.065,-0.044l0.066,-0.043l0.066,-0.042l0.067,-0.041l0.067,-0.04l0.068,-0.04l0.069,-0.039l0.069,-0.038l0.069,-0.037l0.07,-0.037l0.071,-0.035l0.071,-0.035l0.071,-0.034l0.072,-0.033l0.072,-0.032l0.073,-0.031l0.073,-0.031l0.074,-0.03l0.074,-0.028l0.075,-0.028l0.075,-0.027l0.075,-0.026l0.076,-0.025l0.077,-0.025l0.076,-0.023l0.077,-0.022l0.078,-0.022l0.078,-0.02l0.078,-0.02l0.078,-0.019l0.079,-0.017l0.08,-0.017l0.079,-0.016l0.08,-0.015l0.081,-0.014l0.08,-0.013l0.081,-0.011l0.082,-0.011l0.081,-0.01l0.082,-0.009l0.082,-0.008l0.083,-0.007l0.083,-0.005l0.083,-0.005l0.083,-0.004l0.084,-0.003l0.084,-0.001l0.084,-0.001l61.401,0.096Zm-61.473,0.905l-0.071,0.001l-0.071,0.002l-0.07,0.004l-0.071,0.004l-0.07,0.004l-0.07,0.006l-0.07,0.007l-0.069,0.007l-0.069,0.009l-0.069,0.009l-0.068,0.01l-0.069,0.011l-0.068,0.012l-0.067,0.012l-0.068,0.013l-0.067,0.015l-0.066,0.015l-0.067,0.015l-0.066,0.017l-0.066,0.017l-0.065,0.019l-0.065,0.018l-0.065,0.02l-0.065,0.021l-0.064,0.021l-0.064,0.022l-0.063,0.023l-0.063,0.023l-0.063,0.025l-0.062,0.025l-0.062,0.026l-0.062,0.026l-0.061,0.027l-0.061,0.028l-0.06,0.029l-0.06,0.029l-0.06,0.03l-0.059,0.031l-0.059,0.032l-0.058,0.032l-0.058,0.033l-0.058,0.033l-0.057,0.035l-0.056,0.034l-0.056,0.036l-0.056,0.036l-0.055,0.037l-0.055,0.038l-0.055,0.038l-0.053,0.039l-0.054,0.039l-0.053,0.04l-0.052,0.041l-0.052,0.041l-0.051,0.042l-0.051,0.042l-0.051,0.044l-0.049,0.043l-0.05,0.044l-0.049,0.045l-0.048,0.046l-0.047,0.046l-0.048,0.046l-0.046,0.048l-0.046,0.047l-0.046,0.048l-0.044,0.049l-0.045,0.049l-0.043,0.05l-0.044,0.051l-0.042,0.051l-0.042,0.051l-0.041,0.052l-0.041,0.052l-0.04,0.053l-0.039,0.054l-0.039,0.053l-0.038,0.055l-0.038,0.055l-0.037,0.055l-0.036,0.056l-0.036,0.056l-0.034,0.056l-0.035,0.057l-0.033,0.058l-0.033,0.058l-0.032,0.058l-0.032,0.059l-0.031,0.059l-0.03,0.06l-0.029,0.06l-0.029,0.06l-0.028,0.061l-0.027,0.061l-0.026,0.062l-0.026,0.062l-0.025,0.062l-0.025,0.063l-0.023,0.063l-0.023,0.063l-0.022,0.064l-0.021,0.064l-0.021,0.065l-0.019,0.065l-0.019,0.065l-0.019,0.065l-0.017,0.066l-0.017,0.066l-0.015,0.066l-0.015,0.067l-0.015,0.067l-0.013,0.068l-0.012,0.067l-0.012,0.068l-0.011,0.068l-0.01,0.069l-0.009,0.069l-0.009,0.069l-0.007,0.069l-0.007,0.07l-0.006,0.07l-0.004,0.07l-0.004,0.07l-0.004,0.071l-0.002,0.071l-0.001,0.071l-0.001,0.075l0,13.227l49.727,0.027l22.704,-0.018l0,-13.141l0,-0.074l-0.002,-0.072l-0.002,-0.071l-0.003,-0.07l-0.004,-0.071l-0.005,-0.07l-0.006,-0.07l-0.007,-0.069l-0.007,-0.07l-0.008,-0.069l-0.01,-0.069l-0.01,-0.068l-0.011,-0.069l-0.011,-0.067l-0.013,-0.068l-0.013,-0.068l-0.014,-0.067l-0.015,-0.066l-0.016,-0.067l-0.017,-0.066l-0.017,-0.066l-0.018,-0.065l-0.019,-0.065l-0.02,-0.065l-0.02,-0.065l-0.022,-0.064l-0.022,-0.063l-0.023,-0.064l-0.023,-0.063l-0.024,-0.063l-0.025,-0.062l-0.026,-0.062l-0.027,-0.061l-0.027,-0.062l-0.028,-0.06l-0.029,-0.061l-0.029,-0.06l-0.03,-0.06l-0.031,-0.059l-0.031,-0.059l-0.033,-0.058l-0.032,-0.058l-0.034,-0.057l-0.034,-0.057l-0.035,-0.057l-0.036,-0.056l-0.036,-0.056l-0.037,-0.055l-0.037,-0.055l-0.038,-0.054l-0.039,-0.054l-0.04,-0.054l-0.04,-0.053l-0.04,-0.052l-0.042,-0.052l-0.042,-0.051l-0.042,-0.051l-0.043,-0.05l-0.044,-0.05l-0.044,-0.05l-0.045,-0.048l-0.046,-0.049l-0.046,-0.047l-0.046,-0.047l-0.047,-0.047l-0.048,-0.046l-0.048,-0.045l-0.049,-0.045l-0.049,-0.045l-0.05,-0.043l-0.05,-0.043l-0.051,-0.043l-0.052,-0.042l-0.052,-0.041l-0.052,-0.041l-0.053,-0.04l-0.053,-0.039l-0.054,-0.039l-0.054,-0.038l-0.055,-0.038l-0.055,-0.037l-0.056,-0.036l-0.056,-0.035l-0.057,-0.035l-0.057,-0.034l-0.058,-0.034l-0.058,-0.033l-0.058,-0.032l-0.059,-0.032l-0.059,-0.03l-0.06,-0.031l-0.06,-0.029l-0.06,-0.029l-0.061,-0.028l-0.061,-0.027l-0.062,-0.026l-0.062,-0.026l-0.062,-0.025l-0.063,-0.024l-0.063,-0.024l-0.063,-0.023l-0.064,-0.022l-0.064,-0.021l-0.064,-0.021l-0.065,-0.019l-0.065,-0.019l-0.066,-0.018l-0.066,-0.018l-0.066,-0.016l-0.066,-0.016l-0.067,-0.015l-0.067,-0.014l-0.067,-0.014l-0.068,-0.012l-0.068,-0.012l-0.068,-0.011l-0.069,-0.01l-0.069,-0.009l-0.069,-0.008l-0.069,-0.008l-0.07,-0.007l-0.07,-0.005l-0.07,-0.005l-0.07,-0.004l-0.071,-0.003l-0.071,-0.003l-0.071,-0.001l-0.075,0l-61.396,-0.096l-0.074,0.001Z"
-    fill="#757575" />
-    <rect x="21.028" y="58.382" width="72.386" height="11.056" fill="#fff"
-    />
-    <g fill="#bcbcbc">
-        <rect x="55" y="93.191" width="290" height="15.846" />
-        <rect x="55" y="122.266" width="290" height="15.846" />
-        <rect x="55" y="151.341" width="290" height="15.846" />
-        <rect x="55" y="180.416" width="290" height="15.846" />
-        <rect x="55" y="209.491" width="226.139" height="15.846" />
-    </g>
-</svg>;
-icons.tabsBold = <svg viewBox="0 0 400 300" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M176.547,46.207c0,-3.596 -2.919,-6.515 -6.515,-6.515l-61.401,-0.095c-3.596,0 -6.515,2.919 -6.515,6.515l0,14.229l50.727,0.028l23.704,-0.019l0,-14.143Z"
-    fill="#ebebeb" />
-    <path d="M259.298,46.243c0,-3.595 -2.919,-6.514 -6.515,-6.514l-61.401,-0.096c-3.596,0 -6.515,2.92 -6.515,6.515l0,14.23l50.727,0.027l23.704,-0.019l0,-14.143Z"
-    fill="#ebebeb" />
-    <path d="M94.422,46.235c0,-3.595 -2.919,-6.514 -6.515,-6.514l-61.401,-0.096c-3.596,0 -6.515,2.919 -6.515,6.515l0,14.23l50.727,0.027l23.704,-0.019l0,-14.143Z"
-    fill="#0a6689" />
-    <rect x="19.883" y="58.847" width="360.036" height="2" fill="#0a6689"
-    />
-    <g fill="#bcbcbc">
-        <rect x="55" y="93.191" width="290" height="15.846" />
-        <rect x="55" y="122.266" width="290" height="15.846" />
-        <rect x="55" y="151.341" width="290" height="15.846" />
-        <rect x="55" y="180.416" width="290" height="15.846" />
-        <rect x="55" y="209.491" width="226.139" height="15.846" />
-    </g>
-</svg>;
-icons.tabsCenter = <svg viewBox="0 0 400 300" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <rect x="19.883" y="59.342" width="360.036" height="1" fill="#bcbcbc"
-    />
-    <g fill="#bcbcbc">
-        <rect x="55" y="93.191" width="290" height="15.846" />
-        <rect x="55" y="122.266" width="290" height="15.846" />
-        <rect x="55" y="151.341" width="290" height="15.846" />
-        <rect x="55" y="180.416" width="290" height="15.846" />
-        <rect x="55" y="209.491" width="226.139" height="15.846" />
-    </g>
-    <rect x="162.653" y="42.087" width="74.142" height="8.82" fill="#bcbcbc"
-    />
-    <rect x="80.653" y="42.087" width="74.142" height="8.82" fill="#0a6689"
-    />
-    <rect x="80.653" y="57.306" width="74.142" height="3" fill="#0a6689" />
-    <rect x="245.559" y="42.087" width="74.142" height="8.82" fill="#bcbcbc"
-    />
-</svg>;
-icons.tabsVertical = <svg viewBox="0 0 400 300" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M95.571,83.809c0,-2.764 -2.244,-5.007 -5.008,-5.007l-65.67,0c-2.763,0 -5.007,2.243 -5.007,5.007l0,10.014c0,2.763 2.244,5.007 5.007,5.007l65.67,0c2.764,0 5.008,-2.244 5.008,-5.007l0,-10.014Z"
-    fill="#d3d3d3" />
-    <path d="M95.798,54.938c0,-2.762 -2.242,-5.004 -5.004,-5.004l-65.677,0c-2.761,0 -5.003,2.242 -5.003,5.004l0,10.007c0,2.762 2.242,5.004 5.003,5.004l65.677,0c2.762,0 5.004,-2.242 5.004,-5.004l0,-10.007Z"
-    fill="#fff" />
-    <path d="M90.859,49.934l0.064,0.001l0.064,0.002l0.064,0.003l0.064,0.004l0.064,0.004l0.063,0.006l0.063,0.006l0.063,0.006l0.063,0.008l0.062,0.008l0.063,0.009l0.062,0.01l0.061,0.011l0.062,0.011l0.061,0.013l0.061,0.012l0.061,0.014l0.06,0.014l0.06,0.015l0.06,0.016l0.059,0.017l0.059,0.017l0.059,0.018l0.059,0.019l0.058,0.019l0.058,0.02l0.058,0.021l0.057,0.021l0.057,0.022l0.113,0.046l0.112,0.049l0.11,0.052l0.108,0.054l0.107,0.056l0.106,0.059l0.104,0.062l0.102,0.064l0.101,0.066l0.099,0.068l0.097,0.071l0.096,0.074l0.093,0.075l0.092,0.078l0.09,0.079l0.088,0.082l0.086,0.084l0.084,0.086l0.081,0.088l0.08,0.09l0.078,0.091l0.075,0.094l0.073,0.096l0.071,0.097l0.069,0.099l0.066,0.1l0.064,0.103l0.062,0.104l0.059,0.105l0.056,0.107l0.054,0.109l0.052,0.11l0.048,0.112l0.047,0.113l0.022,0.057l0.021,0.057l0.021,0.058l0.02,0.057l0.019,0.059l0.019,0.058l0.018,0.059l0.017,0.059l0.016,0.06l0.016,0.06l0.015,0.06l0.015,0.06l0.013,0.061l0.013,0.061l0.012,0.061l0.012,0.061l0.01,0.062l0.01,0.062l0.009,0.062l0.009,0.063l0.007,0.062l0.007,0.063l0.006,0.064l0.005,0.063l0.005,0.064l0.003,0.063l0.003,0.064l0.002,0.065l0.002,0.064l0,0.065l0,10.007l0,0.065l-0.002,0.065l-0.002,0.064l-0.003,0.064l-0.003,0.064l-0.005,0.063l-0.005,0.064l-0.006,0.063l-0.007,0.063l-0.007,0.062l-0.009,0.063l-0.009,0.062l-0.01,0.062l-0.01,0.062l-0.012,0.061l-0.012,0.061l-0.013,0.061l-0.013,0.061l-0.015,0.06l-0.015,0.06l-0.016,0.06l-0.016,0.06l-0.017,0.059l-0.018,0.059l-0.019,0.058l-0.019,0.059l-0.02,0.058l-0.021,0.057l-0.021,0.058l-0.022,0.057l-0.047,0.112l-0.048,0.112l-0.052,0.11l-0.054,0.109l-0.056,0.107l-0.059,0.105l-0.062,0.104l-0.064,0.103l-0.066,0.1l-0.069,0.099l-0.071,0.098l-0.073,0.095l-0.075,0.094l-0.078,0.091l-0.08,0.09l-0.081,0.088l-0.084,0.086l-0.086,0.084l-0.088,0.082l-0.09,0.08l-0.092,0.077l-0.093,0.076l-0.096,0.073l-0.097,0.071l-0.099,0.068l-0.101,0.067l-0.102,0.064l-0.104,0.061l-0.106,0.059l-0.107,0.057l-0.108,0.054l-0.11,0.051l-0.112,0.049l-0.113,0.046l-0.057,0.022l-0.057,0.022l-0.058,0.02l-0.058,0.02l-0.058,0.02l-0.059,0.018l-0.059,0.018l-0.059,0.017l-0.059,0.017l-0.06,0.016l-0.06,0.015l-0.06,0.014l-0.061,0.014l-0.061,0.013l-0.061,0.012l-0.062,0.011l-0.061,0.011l-0.062,0.01l-0.063,0.009l-0.062,0.008l-0.063,0.008l-0.063,0.007l-0.063,0.006l-0.063,0.005l-0.064,0.004l-0.064,0.004l-0.064,0.003l-0.064,0.002l-0.064,0.001l-0.065,0l-65.677,0l-0.064,0l-0.065,-0.001l-0.064,-0.002l-0.064,-0.003l-0.064,-0.004l-0.063,-0.004l-0.064,-0.005l-0.063,-0.006l-0.063,-0.007l-0.062,-0.008l-0.063,-0.008l-0.062,-0.009l-0.062,-0.01l-0.062,-0.011l-0.061,-0.011l-0.061,-0.012l-0.061,-0.013l-0.061,-0.014l-0.06,-0.014l-0.06,-0.015l-0.06,-0.016l-0.06,-0.017l-0.059,-0.017l-0.059,-0.018l-0.058,-0.018l-0.059,-0.02l-0.058,-0.02l-0.057,-0.02l-0.058,-0.022l-0.057,-0.022l-0.112,-0.046l-0.112,-0.049l-0.11,-0.051l-0.109,-0.054l-0.107,-0.057l-0.105,-0.059l-0.104,-0.061l-0.103,-0.064l-0.1,-0.067l-0.099,-0.068l-0.098,-0.071l-0.095,-0.073l-0.094,-0.076l-0.091,-0.077l-0.09,-0.08l-0.088,-0.082l-0.086,-0.084l-0.084,-0.086l-0.082,-0.088l-0.08,-0.09l-0.077,-0.091l-0.076,-0.094l-0.073,-0.095l-0.071,-0.098l-0.068,-0.099l-0.067,-0.1l-0.064,-0.103l-0.061,-0.104l-0.059,-0.105l-0.057,-0.107l-0.054,-0.109l-0.051,-0.11l-0.049,-0.112l-0.046,-0.112l-0.022,-0.057l-0.022,-0.058l-0.02,-0.057l-0.02,-0.058l-0.02,-0.059l-0.018,-0.058l-0.018,-0.059l-0.017,-0.059l-0.017,-0.06l-0.016,-0.06l-0.015,-0.06l-0.014,-0.06l-0.014,-0.061l-0.013,-0.061l-0.012,-0.061l-0.011,-0.061l-0.011,-0.062l-0.01,-0.062l-0.009,-0.062l-0.008,-0.063l-0.008,-0.062l-0.007,-0.063l-0.006,-0.063l-0.005,-0.064l-0.004,-0.063l-0.004,-0.064l-0.003,-0.064l-0.002,-0.064l-0.001,-0.065l0,-0.065l0,-10.007l0,-0.065l0.001,-0.064l0.002,-0.065l0.003,-0.064l0.004,-0.063l0.004,-0.064l0.005,-0.063l0.006,-0.064l0.007,-0.063l0.008,-0.062l0.008,-0.063l0.009,-0.062l0.01,-0.062l0.011,-0.062l0.011,-0.061l0.012,-0.061l0.013,-0.061l0.014,-0.061l0.014,-0.06l0.015,-0.06l0.016,-0.06l0.017,-0.06l0.017,-0.059l0.018,-0.059l0.018,-0.058l0.02,-0.059l0.02,-0.057l0.02,-0.058l0.022,-0.057l0.022,-0.057l0.046,-0.113l0.049,-0.112l0.051,-0.11l0.054,-0.109l0.057,-0.107l0.059,-0.105l0.061,-0.104l0.064,-0.103l0.067,-0.1l0.068,-0.099l0.071,-0.097l0.073,-0.096l0.076,-0.094l0.077,-0.091l0.08,-0.09l0.082,-0.088l0.084,-0.086l0.086,-0.084l0.088,-0.082l0.09,-0.079l0.091,-0.078l0.094,-0.075l0.095,-0.074l0.098,-0.071l0.099,-0.068l0.1,-0.066l0.103,-0.064l0.104,-0.062l0.105,-0.059l0.107,-0.056l0.109,-0.054l0.11,-0.052l0.112,-0.049l0.112,-0.046l0.057,-0.022l0.058,-0.021l0.057,-0.021l0.058,-0.02l0.059,-0.019l0.058,-0.019l0.059,-0.018l0.059,-0.017l0.06,-0.017l0.06,-0.016l0.06,-0.015l0.06,-0.014l0.061,-0.014l0.061,-0.012l0.061,-0.013l0.061,-0.011l0.062,-0.011l0.062,-0.01l0.062,-0.009l0.063,-0.008l0.062,-0.008l0.063,-0.006l0.063,-0.006l0.064,-0.006l0.063,-0.004l0.064,-0.004l0.064,-0.003l0.064,-0.002l0.065,-0.001l0.064,0l65.677,0l0.065,0Zm-65.735,2l-0.046,0l-0.039,0.001l-0.039,0.001l-0.039,0.002l-0.038,0.002l-0.038,0.003l-0.039,0.003l-0.037,0.003l-0.038,0.004l-0.038,0.005l-0.037,0.005l-0.038,0.006l-0.037,0.005l-0.037,0.007l-0.037,0.007l-0.036,0.007l-0.037,0.008l-0.036,0.008l-0.036,0.008l-0.036,0.009l-0.035,0.01l-0.036,0.01l-0.035,0.01l-0.035,0.011l-0.035,0.011l-0.035,0.011l-0.035,0.012l-0.034,0.013l-0.034,0.012l-0.029,0.011l-0.073,0.03l-0.067,0.029l-0.065,0.031l-0.066,0.033l-0.064,0.033l-0.063,0.036l-0.062,0.037l-0.062,0.038l-0.061,0.04l-0.059,0.041l-0.059,0.043l-0.057,0.044l-0.056,0.046l-0.056,0.046l-0.054,0.048l-0.053,0.05l-0.052,0.05l-0.05,0.052l-0.05,0.053l-0.048,0.054l-0.047,0.056l-0.045,0.056l-0.044,0.058l-0.043,0.058l-0.041,0.06l-0.04,0.06l-0.038,0.062l-0.037,0.062l-0.036,0.063l-0.034,0.065l-0.032,0.065l-0.031,0.066l-0.029,0.066l-0.03,0.073l-0.011,0.029l-0.013,0.034l-0.012,0.034l-0.012,0.035l-0.012,0.035l-0.011,0.035l-0.01,0.035l-0.011,0.035l-0.01,0.036l-0.009,0.036l-0.009,0.035l-0.009,0.036l-0.008,0.037l-0.007,0.036l-0.008,0.037l-0.006,0.036l-0.007,0.037l-0.006,0.037l-0.005,0.038l-0.005,0.037l-0.005,0.038l-0.004,0.038l-0.003,0.038l-0.004,0.038l-0.002,0.038l-0.002,0.039l-0.002,0.038l-0.001,0.039l-0.001,0.039l0,0.046l0,9.995l0,0.046l0.001,0.039l0.001,0.039l0.002,0.038l0.002,0.039l0.002,0.038l0.004,0.039l0.003,0.037l0.004,0.038l0.005,0.038l0.005,0.037l0.005,0.038l0.006,0.037l0.007,0.037l0.006,0.036l0.008,0.037l0.007,0.036l0.008,0.037l0.009,0.036l0.009,0.036l0.009,0.035l0.01,0.036l0.011,0.035l0.01,0.035l0.011,0.035l0.012,0.035l0.012,0.035l0.012,0.034l0.013,0.034l0.011,0.029l0.03,0.073l0.029,0.067l0.031,0.065l0.032,0.065l0.034,0.065l0.036,0.063l0.037,0.062l0.038,0.062l0.04,0.06l0.041,0.06l0.043,0.059l0.044,0.057l0.045,0.056l0.047,0.056l0.048,0.054l0.05,0.053l0.05,0.052l0.052,0.05l0.053,0.05l0.054,0.048l0.056,0.047l0.056,0.045l0.057,0.044l0.059,0.043l0.06,0.041l0.06,0.04l0.062,0.038l0.062,0.037l0.063,0.036l0.064,0.034l0.066,0.032l0.065,0.031l0.067,0.029l0.073,0.03l0.029,0.011l0.034,0.013l0.034,0.012l0.035,0.012l0.035,0.012l0.035,0.011l0.035,0.01l0.035,0.011l0.036,0.009l0.035,0.01l0.036,0.009l0.036,0.008l0.037,0.009l0.036,0.007l0.036,0.008l0.037,0.006l0.037,0.007l0.037,0.006l0.038,0.005l0.037,0.005l0.038,0.005l0.038,0.004l0.038,0.003l0.038,0.004l0.038,0.002l0.039,0.002l0.038,0.002l0.039,0.001l0.039,0.001l0.046,0l65.664,0l0.045,0l0.039,-0.001l0.039,-0.001l0.039,-0.002l0.039,-0.002l0.038,-0.002l0.038,-0.004l0.038,-0.003l0.038,-0.004l0.037,-0.005l0.038,-0.005l0.037,-0.005l0.037,-0.006l0.037,-0.007l0.037,-0.006l0.036,-0.008l0.037,-0.007l0.036,-0.009l0.036,-0.008l0.036,-0.009l0.036,-0.01l0.035,-0.009l0.036,-0.011l0.035,-0.01l0.035,-0.011l0.035,-0.012l0.034,-0.012l0.035,-0.012l0.034,-0.013l0.028,-0.011l0.073,-0.03l0.067,-0.029l0.066,-0.031l0.065,-0.032l0.064,-0.034l0.064,-0.036l0.062,-0.037l0.061,-0.038l0.061,-0.04l0.06,-0.041l0.058,-0.043l0.058,-0.044l0.056,-0.045l0.055,-0.047l0.054,-0.048l0.054,-0.05l0.051,-0.05l0.051,-0.052l0.049,-0.053l0.048,-0.054l0.047,-0.056l0.046,-0.056l0.044,-0.057l0.042,-0.059l0.042,-0.06l0.04,-0.06l0.038,-0.062l0.037,-0.062l0.035,-0.063l0.034,-0.065l0.033,-0.065l0.03,-0.065l0.029,-0.067l0.03,-0.073l0.011,-0.029l0.013,-0.034l0.013,-0.034l0.011,-0.035l0.012,-0.035l0.011,-0.035l0.011,-0.035l0.01,-0.035l0.01,-0.036l0.009,-0.035l0.009,-0.036l0.009,-0.036l0.008,-0.037l0.008,-0.036l0.007,-0.037l0.007,-0.036l0.006,-0.037l0.006,-0.037l0.006,-0.038l0.005,-0.037l0.004,-0.038l0.004,-0.038l0.004,-0.038l0.003,-0.038l0.003,-0.038l0.002,-0.038l0.002,-0.039l0.001,-0.039l0.001,-0.039l0,-0.046l0,-9.995l0,-0.046l-0.001,-0.039l-0.001,-0.039l-0.002,-0.038l-0.002,-0.039l-0.003,-0.038l-0.003,-0.038l-0.004,-0.038l-0.004,-0.038l-0.004,-0.038l-0.005,-0.037l-0.006,-0.038l-0.006,-0.037l-0.006,-0.037l-0.007,-0.036l-0.007,-0.037l-0.008,-0.036l-0.008,-0.037l-0.009,-0.036l-0.009,-0.035l-0.009,-0.036l-0.01,-0.036l-0.01,-0.035l-0.011,-0.035l-0.011,-0.035l-0.011,-0.035l-0.012,-0.034l-0.013,-0.035l-0.013,-0.034l-0.011,-0.029l-0.03,-0.073l-0.029,-0.066l-0.03,-0.066l-0.033,-0.065l-0.034,-0.065l-0.035,-0.063l-0.037,-0.062l-0.038,-0.062l-0.04,-0.06l-0.042,-0.06l-0.042,-0.058l-0.044,-0.058l-0.046,-0.056l-0.047,-0.056l-0.048,-0.054l-0.049,-0.053l-0.051,-0.052l-0.052,-0.05l-0.053,-0.05l-0.054,-0.048l-0.055,-0.046l-0.056,-0.046l-0.058,-0.044l-0.058,-0.043l-0.06,-0.041l-0.061,-0.04l-0.061,-0.038l-0.062,-0.037l-0.064,-0.036l-0.064,-0.033l-0.065,-0.033l-0.066,-0.031l-0.067,-0.029l-0.073,-0.03l-0.028,-0.011l-0.034,-0.012l-0.035,-0.013l-0.034,-0.012l-0.035,-0.011l-0.035,-0.011l-0.035,-0.011l-0.036,-0.01l-0.035,-0.01l-0.036,-0.01l-0.036,-0.009l-0.036,-0.008l-0.036,-0.008l-0.036,-0.008l-0.037,-0.007l-0.037,-0.007l-0.037,-0.007l-0.037,-0.005l-0.037,-0.006l-0.037,-0.005l-0.038,-0.005l-0.038,-0.004l-0.038,-0.003l-0.038,-0.003l-0.038,-0.003l-0.039,-0.002l-0.039,-0.002l-0.039,-0.001l-0.039,-0.001l-0.045,0l-65.664,0Z"
-    fill="#d3d3d3" />
-    <path d="M95.583,112.542c0,-2.762 -2.243,-5.004 -5.005,-5.004l-65.675,0c-2.762,0 -5.004,2.242 -5.004,5.004l0,10.01c0,2.762 2.242,5.005 5.004,5.005l65.675,0c2.762,0 5.005,-2.243 5.005,-5.005l0,-10.01Z"
-    fill="#d3d3d3" />
-    <rect x="90" y="50" width="290" height="200" fill="#fff" />
-    <path d="M380,250l-290,0l0,-200l290,0l0,200Zm-288,-198l0,196l286,0l0,-196l-286,0Z"
-    fill="#d3d3d3" />
-    <rect x="84.867" y="52.004" width="15.81" height="15.928" fill="#fff"
-    />
-    <g fill="#bcbcbc">
-        <rect x="121.802" y="85.088" width="223.198" height="15.846" />
-        <rect x="121.802" y="114.163" width="223.198" height="15.846" />
-        <rect x="121.802" y="143.238" width="223.198" height="15.846" />
-        <rect x="121.802" y="172.313" width="223.198" height="15.846" />
-        <rect x="121.802" y="201.388" width="174.047" height="15.846" />
-    </g>
-</svg>;
-icons.spacerblock = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fill='#0058b0'>
-        <circle cx='14.437' cy='23.593' r='2.77' />
-        <circle cx='24.054' cy='23.593' r='2.77' />
-        <circle cx='33.563' cy='23.593' r='2.77' />
-        <path d='M46.88,14.158l-45.76,0c-0.492,0 -0.896,0.405 -0.896,0.897c0,0.492 0.404,0.897 0.896,0.897l45.76,0c0.492,0 0.896,-0.405 0.896,-0.897c0,-0.492 -0.404,-0.897 -0.896,-0.897Z'
-        fillRule='nonzero' />
-        <path d='M46.88,32.048l-18.939,0c-0.492,0 -0.897,0.405 -0.897,0.897c0,0.492 0.405,0.897 0.897,0.897l18.939,0c0.492,0 0.896,-0.405 0.896,-0.897c0,-0.492 -0.404,-0.897 -0.896,-0.897Z'
-        fillRule='nonzero' />
-        <path d='M0.224,32.945c0,0.492 0.404,0.897 0.896,0.897l18.919,0c0.492,0 0.896,-0.405 0.896,-0.897c0,-0.492 -0.404,-0.897 -0.896,-0.897l-18.919,0c-0.492,0 -0.896,0.405 -0.896,0.897Z'
-        fillRule='nonzero' />
-    </g>
-</svg>;
-icons.fontfamily = <svg width="16px" height="16px" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M39.939,7.124l0,-3.751l-11.251,0l0,3.751l3.75,0l0,33.752l-3.75,0l0,3.751l11.251,0l0,-3.751l-3.751,0l0,-33.752l3.751,0Zm-11.251,4.391c0,-0.195 -0.366,-0.429 -0.6,-0.596c-1.033,-0.771 -2.175,-1.273 -3.443,-1.506c-1.268,-0.234 -2.747,-0.35 -4.45,-0.35c-1.234,0 -2.501,0.195 -3.797,0.592c-1.297,0.396 -2.4,0.876 -3.311,1.446c-1.056,0.667 -1.911,1.429 -2.569,2.278c-0.658,0.852 -0.988,1.744 -0.988,2.676c0,0.893 0.238,1.697 0.714,2.417c0.476,0.72 1.221,1.078 2.233,1.078c1.135,0 2.022,-0.292 2.661,-0.883c0.639,-0.585 0.958,-1.234 0.958,-1.944c0,-0.668 -0.097,-1.397 -0.289,-2.189c-0.195,-0.789 -0.288,-1.387 -0.288,-1.792c0.324,-0.345 0.887,-0.674 1.687,-0.989c0.799,-0.315 1.667,-0.47 2.599,-0.47c1.358,0 2.406,0.277 3.147,0.834c0.74,0.559 1.312,1.212 1.717,1.962c0.364,0.669 0.269,2.186 0.269,2.186l0,4.472c0,0.527 -2.545,1.257 -5.333,2.189c-2.789,0.931 -4.484,1.569 -5.397,1.914c-0.728,0.285 -1.429,0.666 -2.261,1.14c-0.831,0.478 -1.463,1.018 -1.971,1.628c-0.649,0.729 -1.116,1.549 -1.44,2.46c-0.325,0.911 -0.475,1.944 -0.475,3.101c0,2.23 0.73,4.034 2.179,5.412c1.448,1.376 3.3,2.068 5.549,2.068c2.128,0 3.947,-0.834 5.456,-1.603c1.514,-0.774 2.928,-1.92 4.244,-3.795l0.184,0c0.264,1.875 0.851,2.946 1.792,3.686l1.223,0.064l0,-27.486Zm-3.259,22.286c-0.75,0.955 -1.673,1.78 -2.766,2.483c-1.095,0.699 -2.363,1.052 -3.801,1.052c-1.359,0 -2.477,-0.396 -3.358,-1.191c-0.883,-0.791 -1.322,-2.029 -1.322,-3.712c0,-1.3 0.292,-2.425 0.883,-3.368c0.585,-0.947 1.388,-1.763 2.398,-2.453c1.118,-0.733 2.322,-1.341 3.619,-1.83c1.296,-0.488 2.622,-0.934 4.347,-1.438l0,10.457Z"
-    fill="#0e9cd1" fillRule="nonzero" />
-</svg>;
-icons.infoSimple = <svg viewBox="0 0 400 400" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <rect x="49.173" y="26.451" width="301.653" height="347.098" fill="#f2f2f2"
-    />
-    <path d="M329.981,254.814c0,-3.137 -2.548,-5.684 -5.683,-5.684l-248.596,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l248.596,0c3.135,0 5.683,-2.546 5.683,-5.683l0,-2.674Z"
-    fill="#888" />
-    <path d="M329.981,279.093c0,-3.137 -2.548,-5.684 -5.683,-5.684l-248.596,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l248.596,0c3.135,0 5.683,-2.546 5.683,-5.683l0,-2.674Z"
-    fill="#888" />
-    <path d="M329.981,303.372c0,-3.137 -2.548,-5.684 -5.683,-5.684l-248.596,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l248.596,0c3.135,0 5.683,-2.546 5.683,-5.683l0,-2.674Z"
-    fill="#888" />
-    <path d="M306.87,327.889c0,-2.579 -2.095,-4.674 -4.673,-4.674l-204.394,0c-2.579,0 -4.673,2.095 -4.673,4.674l0,2.198c0,2.58 2.094,4.673 4.673,4.673l204.394,0c2.578,0 4.673,-2.093 4.673,-4.673l0,-2.198Z"
-    fill="#888" />
-    <path d="M289.818,221.146c0,-4.194 -3.406,-7.6 -7.6,-7.6l-164.435,0c-4.194,0 -7.6,3.406 -7.6,7.6l0,3.801c0,4.194 3.406,7.6 7.6,7.6l164.435,0c4.194,0 7.6,-3.406 7.6,-7.6l0,-3.801Z"
-    fill="#444" />
-    <path d="M200.87,55.888l0.871,0.014l0.87,0.027l0.864,0.041l0.857,0.048l0.863,0.061l0.85,0.068l0.857,0.081l0.85,0.096l0.843,0.102l0.843,0.108l1.673,0.259l0.836,0.143l0.83,0.156l0.823,0.163l0.823,0.17l0.816,0.184l0.816,0.197l0.809,0.204l0.809,0.211l0.802,0.224l0.803,0.231l0.795,0.245l0.789,0.252l0.782,0.258l0.782,0.272l0.782,0.279l0.775,0.285l0.769,0.3l0.761,0.306l0.762,0.319l0.755,0.327l0.748,0.333l0.748,0.34l0.734,0.353l0.741,0.361l0.728,0.367l0.727,0.381l0.721,0.381l0.714,0.394l0.707,0.401l0.707,0.415l0.694,0.415l0.694,0.428l0.693,0.435l0.68,0.442l1.347,0.912l0.666,0.469l0.66,0.476l0.652,0.483l0.646,0.489l0.64,0.497l0.632,0.503l0.632,0.517l0.619,0.516l0.619,0.531l0.612,0.53l0.598,0.544l0.599,0.551l0.591,0.551l0.578,0.564l0.578,0.571l0.571,0.578l0.565,0.578l0.551,0.592l0.55,0.598l0.544,0.599l0.531,0.612l0.53,0.618l0.517,0.619l0.517,0.633l0.503,0.632l0.496,0.639l0.49,0.646l0.483,0.653l0.476,0.66l0.469,0.666l0.456,0.673l0.448,0.673l0.449,0.68l0.435,0.694l0.429,0.693l0.415,0.694l0.414,0.707l0.402,0.707l0.394,0.714l0.381,0.721l0.381,0.728l0.367,0.727l0.36,0.742l0.354,0.734l0.34,0.748l0.333,0.748l0.326,0.755l0.32,0.761l0.306,0.762l0.299,0.768l0.286,0.775l0.278,0.782l0.272,0.782l0.259,0.782l0.251,0.789l0.245,0.796l0.231,0.802l0.225,0.803l0.211,0.809l0.204,0.809l0.197,0.816l0.183,0.816l0.17,0.823l0.164,0.823l0.156,0.829l0.143,0.837l0.258,1.672l0.109,0.843l0.102,0.844l0.095,0.85l0.082,0.856l0.068,0.85l0.061,0.864l0.048,0.857l0.04,0.863l0.028,0.871l0.013,0.87l0.007,0.871l-0.007,0.87l-0.013,0.87l-0.028,0.871l-0.04,0.863l-0.048,0.857l-0.061,0.864l-0.068,0.85l-0.082,0.856l-0.095,0.85l-0.102,0.844l-0.109,0.843l-0.129,0.836l-0.129,0.837l-0.143,0.836l-0.156,0.83l-0.164,0.822l-0.17,0.823l-0.183,0.816l-0.197,0.816l-0.204,0.809l-0.211,0.809l-0.225,0.803l-0.231,0.802l-0.245,0.789l-0.251,0.796l-0.259,0.782l-0.272,0.782l-0.278,0.782l-0.286,0.768l-0.299,0.775l-0.306,0.762l-0.32,0.761l-0.326,0.755l-0.333,0.748l-0.34,0.748l-0.354,0.735l-0.36,0.741l-0.367,0.727l-0.381,0.728l-0.381,0.721l-0.394,0.714l-0.402,0.707l-0.414,0.707l-0.415,0.694l-0.429,0.693l-0.435,0.694l-0.449,0.68l-0.448,0.673l-0.456,0.673l-0.469,0.667l-0.476,0.659l-0.483,0.653l-0.49,0.646l-0.496,0.639l-0.503,0.633l-0.517,0.632l-0.517,0.619l-0.53,0.618l-0.531,0.612l-0.544,0.599l-0.55,0.598l-0.551,0.592l-0.565,0.578l-0.571,0.578l-0.578,0.571l-0.578,0.564l-0.591,0.551l-0.599,0.551l-0.598,0.544l-0.612,0.53l-0.619,0.531l-0.619,0.517l-0.632,0.516l-0.632,0.504l-0.64,0.496l-0.646,0.49l-0.652,0.482l-0.66,0.476l-0.666,0.47l-0.674,0.455l-0.673,0.449l-0.68,0.449l-0.693,0.435l-0.694,0.428l-0.694,0.415l-0.707,0.415l-0.707,0.401l-0.714,0.394l-0.721,0.381l-0.727,0.381l-0.728,0.367l-0.741,0.361l-0.734,0.353l-0.748,0.34l-0.748,0.333l-0.755,0.327l-0.762,0.319l-0.761,0.306l-0.769,0.3l-0.775,0.285l-0.782,0.279l-0.782,0.272l-0.782,0.258l-0.789,0.252l-0.795,0.245l-0.803,0.231l-0.802,0.224l-0.809,0.211l-0.809,0.204l-0.816,0.197l-0.816,0.184l-0.823,0.17l-0.823,0.163l-0.83,0.157l-0.836,0.142l-0.836,0.13l-0.837,0.129l-0.843,0.109l-0.843,0.102l-0.85,0.095l-0.857,0.081l-0.85,0.068l-0.863,0.062l-0.857,0.047l-0.864,0.041l-0.87,0.027l-0.871,0.014l-0.87,0.007l-0.87,-0.007l-0.871,-0.014l-0.87,-0.027l-0.864,-0.041l-0.857,-0.047l-0.863,-0.062l-0.85,-0.068l-0.857,-0.081l-0.85,-0.095l-0.843,-0.102l-0.843,-0.109l-0.837,-0.129l-0.836,-0.13l-0.836,-0.142l-0.83,-0.157l-0.823,-0.163l-0.823,-0.17l-0.816,-0.184l-0.816,-0.197l-0.809,-0.204l-0.809,-0.211l-0.802,-0.224l-0.803,-0.231l-0.795,-0.245l-0.789,-0.252l-0.782,-0.258l-0.782,-0.272l-0.782,-0.279l-0.775,-0.285l-0.769,-0.3l-0.761,-0.306l-0.762,-0.319l-0.755,-0.327l-0.748,-0.333l-0.748,-0.34l-0.734,-0.353l-0.741,-0.361l-0.728,-0.367l-0.727,-0.381l-0.721,-0.381l-0.714,-0.394l-0.707,-0.401l-0.707,-0.415l-0.694,-0.415l-0.694,-0.428l-0.693,-0.435l-0.68,-0.449l-0.673,-0.449l-0.674,-0.455l-0.666,-0.47l-0.66,-0.476l-0.652,-0.482l-0.646,-0.49l-0.64,-0.496l-0.632,-0.504l-0.632,-0.516l-0.619,-0.517l-0.619,-0.531l-0.612,-0.53l-0.598,-0.544l-0.599,-0.551l-0.591,-0.551l-0.578,-0.564l-0.578,-0.571l-0.571,-0.578l-0.565,-0.578l-0.551,-0.592l-0.55,-0.598l-0.544,-0.599l-0.531,-0.612l-0.53,-0.618l-0.517,-0.619l-0.517,-0.632l-0.503,-0.633l-0.496,-0.639l-0.49,-0.646l-0.483,-0.653l-0.476,-0.659l-0.469,-0.667l-0.456,-0.673l-0.448,-0.673l-0.449,-0.68l-0.435,-0.694l-0.429,-0.693l-0.415,-0.694l-0.414,-0.707l-0.402,-0.707l-0.394,-0.714l-0.381,-0.721l-0.381,-0.728l-0.367,-0.727l-0.36,-0.741l-0.354,-0.735l-0.34,-0.748l-0.333,-0.748l-0.326,-0.755l-0.32,-0.761l-0.306,-0.762l-0.299,-0.775l-0.286,-0.768l-0.278,-0.782l-0.272,-0.782l-0.259,-0.782l-0.251,-0.796l-0.245,-0.789l-0.231,-0.802l-0.225,-0.803l-0.211,-0.809l-0.204,-0.809l-0.197,-0.816l-0.183,-0.816l-0.17,-0.823l-0.164,-0.822l-0.156,-0.83l-0.143,-0.836l-0.129,-0.837l-0.129,-0.836l-0.109,-0.843l-0.102,-0.844l-0.095,-0.85l-0.082,-0.856l-0.068,-0.85l-0.061,-0.864l-0.048,-0.857l-0.04,-0.863l-0.028,-0.871l-0.013,-0.87l-0.007,-0.87l0.007,-0.871l0.013,-0.87l0.028,-0.871l0.04,-0.863l0.048,-0.857l0.061,-0.864l0.068,-0.85l0.082,-0.856l0.095,-0.85l0.102,-0.844l0.109,-0.843l0.129,-0.836l0.129,-0.836l0.143,-0.837l0.156,-0.829l0.164,-0.823l0.17,-0.823l0.183,-0.816l0.197,-0.816l0.204,-0.809l0.211,-0.809l0.225,-0.803l0.231,-0.802l0.245,-0.796l0.251,-0.789l0.259,-0.782l0.272,-0.782l0.278,-0.782l0.286,-0.775l0.299,-0.768l0.306,-0.762l0.32,-0.761l0.326,-0.755l0.333,-0.748l0.34,-0.748l0.354,-0.734l0.36,-0.742l0.367,-0.727l0.381,-0.728l0.381,-0.721l0.394,-0.714l0.402,-0.707l0.414,-0.707l0.415,-0.694l0.429,-0.693l0.435,-0.694l0.449,-0.68l0.448,-0.673l0.456,-0.673l0.469,-0.666l0.476,-0.66l0.483,-0.653l0.49,-0.646l0.496,-0.639l0.503,-0.632l0.517,-0.633l0.517,-0.619l0.53,-0.618l0.531,-0.612l0.544,-0.599l0.55,-0.598l0.551,-0.592l0.565,-0.578l0.571,-0.578l0.578,-0.571l0.578,-0.564l0.591,-0.551l0.599,-0.551l0.598,-0.544l0.612,-0.53l0.619,-0.531l0.619,-0.516l0.632,-0.517l0.632,-0.503l0.64,-0.497l0.646,-0.489l0.652,-0.483l0.66,-0.476l0.666,-0.469l0.674,-0.456l0.673,-0.456l0.68,-0.442l0.693,-0.435l0.694,-0.428l0.694,-0.415l0.707,-0.415l0.707,-0.401l0.714,-0.394l0.721,-0.381l0.727,-0.381l0.728,-0.367l0.741,-0.361l0.734,-0.353l0.748,-0.34l0.748,-0.333l0.755,-0.327l0.762,-0.319l0.761,-0.306l0.769,-0.3l0.775,-0.285l0.782,-0.279l0.782,-0.272l0.782,-0.258l0.789,-0.252l0.795,-0.245l0.803,-0.231l0.802,-0.224l0.809,-0.211l0.809,-0.204l0.816,-0.197l0.816,-0.184l0.823,-0.17l0.823,-0.163l0.83,-0.156l0.836,-0.143l0.836,-0.129l0.837,-0.13l0.843,-0.108l0.843,-0.102l0.85,-0.096l0.857,-0.081l0.85,-0.068l0.863,-0.061l0.857,-0.048l0.864,-0.041l0.87,-0.027l0.871,-0.014l0.87,-0.006l0.87,0.006Zm-1.7,3.4l-0.822,0.014l-0.83,0.027l-0.816,0.034l-0.823,0.048l-0.816,0.061l-0.809,0.061l-0.809,0.082l-0.809,0.088l-0.803,0.095l-0.802,0.109l-0.796,0.116l-0.795,0.122l-0.789,0.136l-0.789,0.15l-0.782,0.156l-0.782,0.163l-0.775,0.177l-0.775,0.184l-0.769,0.19l-0.761,0.204l-0.762,0.211l-0.761,0.217l-0.755,0.232l-0.748,0.238l-0.748,0.251l-0.741,0.252l-0.741,0.265l-0.735,0.279l-0.727,0.279l-0.728,0.292l-0.721,0.299l-0.714,0.313l-0.714,0.313l-0.707,0.326l-0.7,0.333l-0.701,0.347l-0.693,0.347l-0.687,0.36l-0.687,0.368l-0.68,0.374l-0.673,0.38l-0.667,0.388l-0.666,0.401l-0.659,0.401l-0.653,0.415l-0.646,0.422l-0.639,0.428l-0.64,0.435l-0.632,0.442l-0.626,0.456l-0.618,0.456l-0.619,0.469l-0.605,0.469l-0.606,0.483l-0.598,0.482l-0.592,0.497l-0.584,0.503l-0.578,0.503l-0.572,0.517l-0.564,0.524l-0.564,0.523l-0.551,0.537l-0.551,0.538l-0.537,0.55l-0.537,0.551l-0.524,0.565l-0.523,0.564l-0.517,0.571l-0.503,0.578l-0.504,0.585l-0.496,0.592l-0.483,0.598l-0.483,0.605l-0.469,0.605l-0.469,0.619l-0.456,0.619l-0.448,0.625l-0.449,0.633l-0.435,0.639l-0.429,0.639l-0.421,0.646l-0.415,0.653l-0.401,0.66l-0.402,0.666l-0.387,0.666l-0.381,0.674l-0.374,0.68l-0.367,0.686l-0.361,0.687l-0.346,0.694l-0.347,0.7l-0.333,0.701l-0.327,0.707l-0.313,0.714l-0.312,0.714l-0.3,0.721l-0.292,0.727l-0.279,0.728l-0.279,0.734l-0.265,0.741l-0.251,0.741l-0.252,0.748l-0.238,0.748l-0.231,0.755l-0.218,0.762l-0.21,0.761l-0.204,0.762l-0.191,0.768l-0.183,0.776l-0.177,0.775l-0.163,0.782l-0.157,0.782l-0.149,0.789l-0.136,0.788l-0.123,0.796l-0.115,0.795l-0.109,0.803l-0.095,0.802l-0.089,0.809l-0.081,0.81l-0.062,0.809l-0.061,0.816l-0.047,0.823l-0.034,0.816l-0.028,0.829l-0.013,0.823l-0.007,0.83l0.007,0.829l0.013,0.823l0.028,0.83l0.034,0.815l0.047,0.823l0.061,0.816l0.062,0.809l0.081,0.81l0.089,0.809l0.095,0.802l0.109,0.803l0.115,0.795l0.123,0.796l0.136,0.789l0.149,0.788l0.157,0.782l0.163,0.782l0.177,0.776l0.183,0.775l0.191,0.768l0.204,0.762l0.21,0.761l0.218,0.762l0.231,0.755l0.238,0.748l0.252,0.748l0.251,0.741l0.265,0.741l0.279,0.734l0.279,0.728l0.292,0.728l0.3,0.72l0.625,1.428l0.327,0.708l0.333,0.7l0.347,0.7l0.346,0.694l0.361,0.687l0.367,0.686l0.374,0.68l0.381,0.674l0.387,0.666l0.402,0.666l0.401,0.66l0.415,0.653l0.421,0.646l0.429,0.639l0.435,0.639l0.449,0.633l0.448,0.625l0.456,0.619l0.469,0.619l0.469,0.605l0.483,0.605l0.483,0.599l0.496,0.584l0.504,0.592l0.503,0.578l0.517,0.571l1.047,1.129l1.074,1.101l1.102,1.075l1.128,1.047l0.572,0.517l0.578,0.503l0.591,0.503l0.585,0.497l0.598,0.482l0.606,0.483l0.605,0.469l0.619,0.47l0.618,0.455l0.626,0.449l0.632,0.449l0.64,0.435l0.639,0.428l0.646,0.422l0.653,0.415l0.659,0.401l0.666,0.401l0.667,0.388l0.673,0.381l0.68,0.374l0.687,0.367l0.687,0.36l0.693,0.347l0.701,0.347l0.7,0.333l0.707,0.326l1.428,0.626l0.721,0.299l0.728,0.292l0.727,0.279l0.735,0.279l0.741,0.265l0.741,0.252l0.748,0.251l0.748,0.238l0.755,0.232l0.761,0.217l0.762,0.211l0.761,0.204l0.769,0.19l0.775,0.184l0.775,0.177l0.782,0.163l0.782,0.156l0.789,0.15l0.789,0.136l0.795,0.122l0.796,0.116l0.802,0.109l0.803,0.095l0.809,0.088l0.809,0.082l0.809,0.061l0.816,0.061l0.823,0.048l0.816,0.034l0.83,0.027l0.822,0.014l0.83,0.007l0.83,-0.007l0.822,-0.014l0.83,-0.027l0.816,-0.034l0.823,-0.048l0.816,-0.061l0.809,-0.061l0.809,-0.082l0.809,-0.088l0.803,-0.095l0.802,-0.109l0.796,-0.116l0.795,-0.122l0.789,-0.136l0.789,-0.15l0.782,-0.156l0.782,-0.163l0.775,-0.177l0.775,-0.184l0.769,-0.19l0.761,-0.204l0.762,-0.211l0.761,-0.217l0.755,-0.232l0.748,-0.238l0.748,-0.251l0.741,-0.252l0.741,-0.265l0.735,-0.279l0.727,-0.279l0.728,-0.292l0.721,-0.299l0.714,-0.313l0.714,-0.313l0.707,-0.326l0.7,-0.333l0.701,-0.347l0.693,-0.347l0.687,-0.36l0.687,-0.367l0.68,-0.374l0.673,-0.381l0.667,-0.388l0.666,-0.401l0.659,-0.401l0.653,-0.415l0.646,-0.422l0.639,-0.428l0.64,-0.435l0.632,-0.449l0.626,-0.449l0.618,-0.455l0.619,-0.47l0.605,-0.469l0.606,-0.483l0.598,-0.482l0.585,-0.497l0.591,-0.503l0.578,-0.503l0.572,-0.517l0.564,-0.524l0.564,-0.523l0.551,-0.537l0.551,-0.538l0.537,-0.55l0.537,-0.551l0.524,-0.565l0.523,-0.564l0.517,-0.571l0.503,-0.578l0.504,-0.592l0.496,-0.584l0.483,-0.599l0.483,-0.605l0.469,-0.605l0.469,-0.619l0.456,-0.619l0.448,-0.625l0.449,-0.633l0.435,-0.639l0.429,-0.639l0.421,-0.646l0.415,-0.653l0.401,-0.66l0.402,-0.666l0.387,-0.666l0.381,-0.674l0.374,-0.68l0.367,-0.686l0.361,-0.687l0.346,-0.694l0.347,-0.7l0.333,-0.7l0.327,-0.708l0.313,-0.714l0.312,-0.714l0.3,-0.72l0.292,-0.728l0.279,-0.728l0.279,-0.734l0.265,-0.741l0.251,-0.741l0.252,-0.748l0.238,-0.748l0.231,-0.755l0.218,-0.762l0.21,-0.761l0.204,-0.762l0.191,-0.768l0.183,-0.775l0.177,-0.776l0.163,-0.782l0.157,-0.782l0.149,-0.788l0.136,-0.789l0.123,-0.796l0.115,-0.795l0.109,-0.803l0.095,-0.802l0.089,-0.809l0.081,-0.81l0.062,-0.809l0.061,-0.816l0.047,-0.823l0.034,-0.815l0.028,-0.83l0.013,-0.823l0.007,-0.829l-0.007,-0.83l-0.013,-0.823l-0.028,-0.829l-0.034,-0.816l-0.047,-0.823l-0.061,-0.816l-0.062,-0.809l-0.081,-0.81l-0.089,-0.809l-0.095,-0.802l-0.109,-0.803l-0.115,-0.795l-0.123,-0.796l-0.136,-0.788l-0.149,-0.789l-0.157,-0.782l-0.163,-0.782l-0.177,-0.775l-0.183,-0.776l-0.191,-0.768l-0.204,-0.762l-0.21,-0.761l-0.218,-0.762l-0.231,-0.755l-0.238,-0.748l-0.252,-0.748l-0.251,-0.741l-0.265,-0.741l-0.279,-0.734l-0.279,-0.728l-0.292,-0.727l-0.3,-0.721l-0.312,-0.714l-0.313,-0.714l-0.327,-0.707l-0.333,-0.701l-0.347,-0.7l-0.346,-0.694l-0.361,-0.687l-0.367,-0.686l-0.374,-0.68l-0.381,-0.674l-0.387,-0.666l-0.402,-0.666l-0.401,-0.66l-0.415,-0.653l-0.421,-0.646l-0.429,-0.639l-0.435,-0.639l-0.449,-0.633l-0.448,-0.625l-0.456,-0.619l-0.469,-0.619l-0.469,-0.605l-0.483,-0.605l-0.483,-0.598l-0.496,-0.592l-0.504,-0.585l-0.503,-0.578l-0.517,-0.571l-0.523,-0.564l-0.524,-0.565l-0.537,-0.551l-0.537,-0.55l-0.551,-0.538l-0.551,-0.537l-0.564,-0.523l-0.564,-0.524l-0.572,-0.517l-0.578,-0.503l-0.584,-0.503l-0.592,-0.497l-0.598,-0.482l-0.606,-0.483l-0.605,-0.469l-0.619,-0.469l-0.618,-0.456l-0.626,-0.456l-0.632,-0.442l-0.64,-0.435l-0.639,-0.428l-0.646,-0.422l-0.653,-0.415l-0.659,-0.401l-0.666,-0.401l-0.667,-0.388l-0.673,-0.38l-0.68,-0.374l-0.687,-0.368l-0.687,-0.36l-0.693,-0.347l-0.701,-0.347l-0.7,-0.333l-0.707,-0.326l-0.714,-0.313l-0.714,-0.313l-0.721,-0.299l-0.728,-0.292l-0.727,-0.279l-0.735,-0.279l-0.741,-0.265l-0.741,-0.252l-0.748,-0.251l-0.748,-0.238l-0.755,-0.232l-0.761,-0.217l-0.762,-0.211l-0.761,-0.204l-0.769,-0.19l-0.775,-0.184l-0.775,-0.177l-0.782,-0.163l-0.782,-0.156l-0.789,-0.15l-0.789,-0.136l-0.795,-0.122l-0.796,-0.116l-0.802,-0.109l-0.803,-0.095l-0.809,-0.088l-0.809,-0.082l-0.809,-0.061l-0.816,-0.061l-0.823,-0.048l-0.816,-0.034l-0.83,-0.027l-0.822,-0.014l-0.83,-0.006l-0.83,0.006Z"
-    fill="#444" />
-    <path d="M200.041,77.715l-10.968,33.042l-32.901,0l27.418,22.028l-10.967,33.041l27.418,-22.027l27.417,22.027l-10.967,-33.041l27.418,-22.028l-32.901,0l-10.967,-33.042Z"
-    fill="#444" fillRule="nonzero" />
-</svg>;
-icons.infoLeft = <svg viewBox="0 0 400 400" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M390.993,128.492c0,-6.336 -5.144,-11.479 -11.479,-11.479l-359.028,0c-6.335,0 -11.479,5.143 -11.479,11.479l0,143.016c0,6.336 5.144,11.479 11.479,11.479l359.028,0c6.335,0 11.479,-5.143 11.479,-11.479l0,-143.016Z"
-    fill="#fff" />
-    <path d="M379.694,114.514l0.179,0.003l0.179,0.006l0.179,0.008l0.178,0.01l0.177,0.012l0.177,0.015l0.177,0.017l0.175,0.019l0.176,0.021l0.174,0.023l0.174,0.026l0.173,0.027l0.173,0.03l0.172,0.032l0.171,0.034l0.17,0.036l0.17,0.038l0.169,0.04l0.168,0.042l0.167,0.045l0.166,0.046l0.166,0.048l0.165,0.05l0.164,0.052l0.163,0.054l0.162,0.056l0.161,0.058l0.16,0.06l0.16,0.062l0.158,0.064l0.158,0.065l0.156,0.067l0.156,0.07l0.154,0.071l0.153,0.072l0.153,0.075l0.151,0.076l0.15,0.079l0.149,0.079l0.148,0.082l0.147,0.083l0.146,0.086l0.145,0.086l0.143,0.089l0.143,0.09l0.141,0.091l0.14,0.094l0.139,0.095l0.137,0.097l0.137,0.098l0.135,0.1l0.133,0.101l0.133,0.103l0.131,0.104l0.13,0.106l0.129,0.108l0.127,0.109l0.126,0.11l0.125,0.112l0.123,0.114l0.122,0.115l0.121,0.116l0.119,0.118l0.118,0.119l0.116,0.12l0.115,0.122l0.113,0.124l0.112,0.124l0.111,0.126l0.109,0.128l0.107,0.128l0.106,0.13l0.105,0.132l0.103,0.132l0.101,0.134l0.1,0.135l0.098,0.136l0.097,0.138l0.095,0.139l0.093,0.14l0.092,0.141l0.09,0.142l0.088,0.144l0.087,0.144l0.085,0.146l0.083,0.147l0.082,0.148l0.08,0.149l0.078,0.151l0.076,0.151l0.075,0.152l0.073,0.154l0.071,0.154l0.069,0.155l0.067,0.157l0.066,0.157l0.063,0.159l0.062,0.159l0.06,0.161l0.058,0.161l0.056,0.162l0.054,0.163l0.052,0.163l0.05,0.165l0.049,0.166l0.046,0.167l0.044,0.167l0.042,0.168l0.04,0.169l0.039,0.17l0.035,0.17l0.034,0.171l0.032,0.172l0.03,0.172l0.028,0.173l0.025,0.174l0.024,0.175l0.021,0.175l0.019,0.176l0.016,0.176l0.015,0.177l0.012,0.178l0.011,0.178l0.008,0.178l0.005,0.179l0.004,0.18l0.001,0.172l0,143.032l-0.001,0.172l-0.004,0.18l-0.005,0.179l-0.008,0.178l-0.011,0.178l-0.012,0.178l-0.015,0.177l-0.016,0.176l-0.019,0.176l-0.021,0.175l-0.024,0.175l-0.025,0.174l-0.028,0.173l-0.03,0.172l-0.032,0.172l-0.034,0.171l-0.035,0.17l-0.039,0.17l-0.04,0.169l-0.042,0.168l-0.044,0.167l-0.046,0.167l-0.049,0.166l-0.05,0.165l-0.052,0.163l-0.054,0.163l-0.056,0.162l-0.058,0.161l-0.06,0.161l-0.061,0.159l-0.064,0.159l-0.066,0.157l-0.067,0.156l-0.069,0.156l-0.071,0.154l-0.073,0.154l-0.075,0.152l-0.076,0.151l-0.078,0.151l-0.08,0.149l-0.082,0.148l-0.083,0.147l-0.085,0.146l-0.087,0.144l-0.088,0.144l-0.09,0.142l-0.092,0.141l-0.093,0.14l-0.096,0.139l-0.096,0.138l-0.098,0.136l-0.1,0.135l-0.101,0.134l-0.103,0.132l-0.105,0.131l-0.106,0.131l-0.107,0.129l-0.109,0.127l-0.111,0.126l-0.112,0.124l-0.113,0.124l-0.115,0.122l-0.117,0.12l-0.117,0.119l-0.119,0.118l-0.121,0.116l-0.122,0.115l-0.123,0.114l-0.125,0.112l-0.126,0.11l-0.127,0.109l-0.129,0.108l-0.13,0.106l-0.131,0.104l-0.133,0.103l-0.134,0.101l-0.134,0.1l-0.137,0.098l-0.137,0.097l-0.139,0.095l-0.14,0.093l-0.141,0.092l-0.143,0.09l-0.143,0.089l-0.145,0.086l-0.146,0.086l-0.147,0.083l-0.148,0.082l-0.149,0.079l-0.15,0.079l-0.151,0.076l-0.153,0.074l-0.153,0.073l-0.154,0.071l-0.156,0.07l-0.156,0.067l-0.158,0.065l-0.158,0.064l-0.16,0.062l-0.16,0.06l-0.161,0.058l-0.162,0.056l-0.163,0.054l-0.164,0.052l-0.165,0.05l-0.166,0.048l-0.166,0.046l-0.167,0.045l-0.168,0.042l-0.169,0.04l-0.17,0.038l-0.17,0.036l-0.171,0.034l-0.172,0.032l-0.173,0.03l-0.173,0.027l-0.174,0.026l-0.174,0.023l-0.176,0.021l-0.175,0.019l-0.177,0.017l-0.177,0.015l-0.177,0.012l-0.178,0.01l-0.179,0.008l-0.179,0.006l-0.179,0.003l-0.172,0.001l-359.044,0l-0.172,-0.001l-0.179,-0.003l-0.179,-0.006l-0.179,-0.008l-0.178,-0.01l-0.177,-0.012l-0.177,-0.015l-0.177,-0.017l-0.175,-0.019l-0.176,-0.021l-0.174,-0.023l-0.174,-0.026l-0.173,-0.027l-0.173,-0.03l-0.172,-0.032l-0.171,-0.034l-0.17,-0.036l-0.17,-0.038l-0.169,-0.04l-0.168,-0.042l-0.167,-0.045l-0.166,-0.046l-0.166,-0.048l-0.165,-0.05l-0.164,-0.052l-0.163,-0.054l-0.162,-0.056l-0.161,-0.058l-0.16,-0.06l-0.16,-0.062l-0.158,-0.064l-0.158,-0.065l-0.156,-0.067l-0.156,-0.07l-0.154,-0.071l-0.153,-0.073l-0.153,-0.074l-0.151,-0.076l-0.15,-0.079l-0.149,-0.079l-0.149,-0.082l-0.146,-0.083l-0.146,-0.086l-0.145,-0.086l-0.143,-0.089l-0.143,-0.09l-0.141,-0.092l-0.14,-0.093l-0.139,-0.095l-0.137,-0.097l-0.136,-0.098l-0.135,-0.1l-0.134,-0.101l-0.133,-0.103l-0.131,-0.104l-0.13,-0.106l-0.129,-0.108l-0.127,-0.109l-0.126,-0.11l-0.125,-0.112l-0.123,-0.114l-0.122,-0.115l-0.121,-0.116l-0.119,-0.118l-0.118,-0.119l-0.116,-0.12l-0.115,-0.122l-0.113,-0.124l-0.112,-0.124l-0.111,-0.126l-0.109,-0.128l-0.107,-0.128l-0.106,-0.13l-0.105,-0.132l-0.103,-0.132l-0.101,-0.134l-0.1,-0.135l-0.098,-0.136l-0.096,-0.138l-0.096,-0.139l-0.093,-0.14l-0.092,-0.141l-0.09,-0.142l-0.088,-0.144l-0.087,-0.144l-0.085,-0.146l-0.083,-0.147l-0.082,-0.148l-0.08,-0.149l-0.078,-0.151l-0.076,-0.151l-0.075,-0.152l-0.073,-0.154l-0.071,-0.154l-0.069,-0.155l-0.067,-0.157l-0.066,-0.157l-0.064,-0.159l-0.061,-0.159l-0.06,-0.16l-0.058,-0.162l-0.056,-0.162l-0.054,-0.163l-0.052,-0.164l-0.051,-0.164l-0.048,-0.166l-0.046,-0.167l-0.044,-0.167l-0.042,-0.168l-0.04,-0.169l-0.039,-0.169l-0.036,-0.171l-0.033,-0.171l-0.032,-0.172l-0.03,-0.172l-0.028,-0.174l-0.025,-0.173l-0.023,-0.175l-0.022,-0.175l-0.019,-0.176l-0.016,-0.176l-0.015,-0.177l-0.012,-0.178l-0.011,-0.178l-0.008,-0.178l-0.005,-0.179l-0.004,-0.18l-0.001,-0.172l0,-143.032l0.001,-0.172l0.004,-0.18l0.005,-0.179l0.008,-0.178l0.011,-0.178l0.012,-0.178l0.015,-0.177l0.016,-0.176l0.019,-0.176l0.022,-0.175l0.023,-0.175l0.025,-0.174l0.028,-0.173l0.03,-0.172l0.032,-0.172l0.033,-0.171l0.036,-0.171l0.039,-0.169l0.04,-0.169l0.042,-0.168l0.044,-0.167l0.046,-0.167l0.048,-0.166l0.051,-0.164l0.052,-0.164l0.054,-0.163l0.056,-0.162l0.058,-0.162l0.06,-0.16l0.061,-0.159l0.064,-0.159l0.066,-0.157l0.067,-0.157l0.069,-0.155l0.071,-0.154l0.073,-0.154l0.075,-0.152l0.076,-0.151l0.078,-0.151l0.08,-0.149l0.082,-0.148l0.083,-0.147l0.085,-0.146l0.087,-0.144l0.088,-0.144l0.09,-0.142l0.092,-0.141l0.093,-0.14l0.095,-0.139l0.097,-0.138l0.098,-0.136l0.1,-0.135l0.101,-0.134l0.103,-0.132l0.105,-0.132l0.106,-0.13l0.107,-0.128l0.109,-0.128l0.111,-0.126l0.112,-0.124l0.113,-0.124l0.115,-0.122l0.116,-0.12l0.118,-0.119l0.119,-0.118l0.121,-0.116l0.122,-0.115l0.123,-0.114l0.125,-0.112l0.126,-0.11l0.127,-0.109l0.129,-0.108l0.13,-0.106l0.131,-0.104l0.133,-0.103l0.133,-0.101l0.136,-0.1l0.136,-0.098l0.137,-0.097l0.139,-0.095l0.14,-0.093l0.141,-0.092l0.143,-0.09l0.143,-0.089l0.145,-0.086l0.146,-0.086l0.147,-0.083l0.148,-0.082l0.149,-0.079l0.15,-0.079l0.151,-0.076l0.153,-0.075l0.153,-0.072l0.154,-0.071l0.156,-0.07l0.156,-0.067l0.158,-0.065l0.158,-0.064l0.16,-0.062l0.16,-0.06l0.161,-0.058l0.162,-0.056l0.163,-0.054l0.164,-0.052l0.165,-0.05l0.166,-0.048l0.166,-0.046l0.167,-0.045l0.168,-0.042l0.169,-0.04l0.17,-0.038l0.17,-0.036l0.171,-0.034l0.172,-0.032l0.173,-0.03l0.173,-0.027l0.174,-0.026l0.174,-0.023l0.176,-0.021l0.175,-0.019l0.177,-0.017l0.177,-0.015l0.177,-0.012l0.178,-0.01l0.179,-0.008l0.179,-0.006l0.179,-0.003l0.172,-0.001l359.044,0l0.172,0.001Zm-359.208,2.499l-0.148,0.001l-0.148,0.003l-0.147,0.004l-0.147,0.007l-0.147,0.008l-0.145,0.01l-0.146,0.012l-0.145,0.014l-0.144,0.016l-0.144,0.017l-0.143,0.019l-0.143,0.021l-0.142,0.023l-0.142,0.024l-0.141,0.026l-0.14,0.028l-0.14,0.03l-0.139,0.031l-0.138,0.033l-0.138,0.034l-0.137,0.037l-0.137,0.038l-0.136,0.039l-0.135,0.041l-0.134,0.043l-0.134,0.044l-0.133,0.046l-0.132,0.048l-0.131,0.049l-0.131,0.051l-0.13,0.052l-0.129,0.053l-0.128,0.056l-0.128,0.056l-0.127,0.059l-0.126,0.059l-0.125,0.062l-0.124,0.062l-0.123,0.064l-0.122,0.066l-0.122,0.067l-0.121,0.069l-0.119,0.069l-0.119,0.072l-0.118,0.072l-0.117,0.074l-0.116,0.076l-0.115,0.076l-0.114,0.078l-0.113,0.08l-0.112,0.08l-0.111,0.082l-0.11,0.084l-0.109,0.084l-0.108,0.086l-0.107,0.087l-0.105,0.089l-0.105,0.089l-0.104,0.091l-0.102,0.092l-0.102,0.094l-0.1,0.094l-0.099,0.096l-0.098,0.096l-0.097,0.098l-0.096,0.1l-0.094,0.1l-0.093,0.101l-0.092,0.103l-0.091,0.103l-0.09,0.105l-0.088,0.106l-0.087,0.107l-0.086,0.108l-0.085,0.108l-0.083,0.11l-0.082,0.111l-0.081,0.113l-0.079,0.113l-0.078,0.114l-0.077,0.115l-0.075,0.116l-0.074,0.116l-0.073,0.118l-0.071,0.119l-0.07,0.12l-0.068,0.12l-0.067,0.122l-0.066,0.122l-0.064,0.124l-0.063,0.124l-0.061,0.125l-0.06,0.126l-0.058,0.127l-0.057,0.127l-0.055,0.128l-0.054,0.129l-0.052,0.13l-0.051,0.131l-0.049,0.132l-0.047,0.132l-0.046,0.133l-0.045,0.133l-0.042,0.135l-0.041,0.135l-0.04,0.136l-0.038,0.136l-0.036,0.137l-0.035,0.138l-0.033,0.139l-0.031,0.139l-0.029,0.14l-0.028,0.14l-0.026,0.141l-0.025,0.141l-0.022,0.143l-0.021,0.142l-0.019,0.144l-0.018,0.143l-0.015,0.145l-0.014,0.145l-0.012,0.145l-0.01,0.146l-0.009,0.146l-0.006,0.147l-0.005,0.148l-0.003,0.147l-0.001,0.149l0,143.016l0.001,0.149l0.003,0.147l0.005,0.148l0.006,0.147l0.009,0.146l0.01,0.146l0.012,0.145l0.014,0.145l0.015,0.145l0.018,0.143l0.019,0.144l0.021,0.142l0.022,0.143l0.025,0.141l0.026,0.141l0.028,0.14l0.029,0.14l0.031,0.139l0.033,0.139l0.035,0.138l0.036,0.137l0.038,0.136l0.04,0.136l0.041,0.135l0.042,0.135l0.045,0.133l0.046,0.133l0.047,0.132l0.049,0.132l0.051,0.131l0.052,0.13l0.054,0.129l0.055,0.128l0.057,0.127l0.058,0.127l0.06,0.126l0.061,0.125l0.063,0.124l0.064,0.124l0.066,0.122l0.067,0.122l0.068,0.12l0.07,0.12l0.071,0.119l0.073,0.118l0.074,0.116l0.075,0.116l0.077,0.115l0.078,0.114l0.079,0.113l0.081,0.113l0.082,0.111l0.083,0.11l0.085,0.108l0.086,0.108l0.087,0.107l0.088,0.106l0.09,0.105l0.091,0.103l0.092,0.103l0.093,0.101l0.094,0.1l0.096,0.1l0.097,0.098l0.098,0.096l0.099,0.096l0.1,0.094l0.102,0.094l0.102,0.092l0.104,0.091l0.105,0.089l0.105,0.089l0.107,0.087l0.108,0.086l0.109,0.084l0.11,0.084l0.111,0.082l0.112,0.08l0.113,0.08l0.114,0.078l0.115,0.076l0.116,0.076l0.117,0.074l0.118,0.072l0.119,0.072l0.119,0.069l0.121,0.069l0.122,0.067l0.122,0.066l0.123,0.064l0.124,0.062l0.125,0.062l0.126,0.059l0.127,0.059l0.128,0.056l0.128,0.056l0.129,0.053l0.13,0.052l0.131,0.051l0.131,0.049l0.132,0.048l0.133,0.046l0.134,0.044l0.134,0.043l0.135,0.041l0.136,0.039l0.137,0.038l0.137,0.037l0.138,0.034l0.138,0.033l0.139,0.031l0.14,0.03l0.14,0.028l0.141,0.026l0.142,0.024l0.142,0.023l0.143,0.021l0.143,0.019l0.144,0.017l0.144,0.016l0.145,0.014l0.146,0.012l0.145,0.01l0.147,0.008l0.147,0.007l0.147,0.004l0.148,0.003l0.148,0.001l359.028,0l0.148,-0.001l0.148,-0.003l0.147,-0.004l0.147,-0.007l0.147,-0.008l0.145,-0.01l0.146,-0.012l0.145,-0.014l0.144,-0.016l0.144,-0.017l0.143,-0.019l0.143,-0.021l0.142,-0.023l0.142,-0.024l0.141,-0.026l0.14,-0.028l0.14,-0.03l0.139,-0.031l0.138,-0.033l0.138,-0.034l0.137,-0.037l0.137,-0.038l0.136,-0.039l0.135,-0.041l0.134,-0.043l0.134,-0.044l0.133,-0.046l0.132,-0.048l0.131,-0.049l0.131,-0.051l0.13,-0.052l0.129,-0.053l0.128,-0.056l0.128,-0.056l0.127,-0.059l0.126,-0.059l0.125,-0.062l0.124,-0.062l0.123,-0.064l0.122,-0.066l0.122,-0.067l0.121,-0.069l0.119,-0.069l0.119,-0.072l0.118,-0.072l0.117,-0.074l0.116,-0.076l0.115,-0.076l0.114,-0.078l0.113,-0.08l0.112,-0.08l0.111,-0.082l0.11,-0.084l0.109,-0.084l0.108,-0.086l0.107,-0.087l0.105,-0.089l0.105,-0.089l0.104,-0.091l0.102,-0.092l0.102,-0.094l0.1,-0.094l0.099,-0.096l0.098,-0.096l0.097,-0.098l0.096,-0.1l0.094,-0.1l0.093,-0.101l0.092,-0.103l0.091,-0.103l0.09,-0.105l0.088,-0.106l0.087,-0.107l0.086,-0.108l0.085,-0.108l0.083,-0.11l0.082,-0.111l0.081,-0.113l0.079,-0.113l0.078,-0.114l0.077,-0.115l0.075,-0.116l0.074,-0.116l0.073,-0.118l0.071,-0.119l0.07,-0.12l0.068,-0.12l0.067,-0.122l0.066,-0.122l0.064,-0.124l0.063,-0.124l0.061,-0.125l0.06,-0.126l0.058,-0.127l0.057,-0.127l0.055,-0.128l0.054,-0.129l0.052,-0.13l0.051,-0.131l0.049,-0.132l0.047,-0.132l0.046,-0.133l0.045,-0.133l0.042,-0.135l0.041,-0.135l0.04,-0.136l0.038,-0.136l0.036,-0.137l0.035,-0.138l0.033,-0.139l0.031,-0.139l0.029,-0.14l0.028,-0.14l0.026,-0.141l0.025,-0.141l0.022,-0.143l0.021,-0.142l0.019,-0.144l0.018,-0.143l0.015,-0.145l0.014,-0.145l0.012,-0.145l0.01,-0.146l0.009,-0.146l0.006,-0.147l0.005,-0.148l0.003,-0.147l0.001,-0.149l0,-143.016l-0.001,-0.149l-0.003,-0.147l-0.005,-0.148l-0.006,-0.147l-0.009,-0.146l-0.01,-0.146l-0.012,-0.145l-0.014,-0.145l-0.015,-0.145l-0.018,-0.143l-0.019,-0.144l-0.021,-0.142l-0.022,-0.143l-0.025,-0.141l-0.026,-0.141l-0.028,-0.14l-0.029,-0.14l-0.031,-0.139l-0.033,-0.139l-0.035,-0.138l-0.036,-0.137l-0.038,-0.136l-0.04,-0.136l-0.041,-0.135l-0.042,-0.135l-0.045,-0.133l-0.046,-0.133l-0.047,-0.132l-0.049,-0.132l-0.051,-0.131l-0.052,-0.13l-0.054,-0.129l-0.055,-0.128l-0.057,-0.127l-0.058,-0.127l-0.06,-0.126l-0.061,-0.125l-0.063,-0.124l-0.064,-0.124l-0.066,-0.122l-0.067,-0.122l-0.068,-0.12l-0.07,-0.12l-0.071,-0.119l-0.073,-0.118l-0.074,-0.116l-0.075,-0.116l-0.077,-0.115l-0.078,-0.114l-0.079,-0.113l-0.081,-0.113l-0.082,-0.111l-0.083,-0.11l-0.085,-0.108l-0.086,-0.108l-0.087,-0.107l-0.088,-0.106l-0.09,-0.105l-0.091,-0.103l-0.092,-0.103l-0.093,-0.101l-0.094,-0.1l-0.096,-0.1l-0.097,-0.098l-0.098,-0.096l-0.099,-0.096l-0.1,-0.094l-0.102,-0.094l-0.102,-0.092l-0.104,-0.091l-0.105,-0.089l-0.105,-0.089l-0.107,-0.087l-0.108,-0.086l-0.109,-0.084l-0.11,-0.084l-0.111,-0.082l-0.112,-0.08l-0.113,-0.08l-0.114,-0.078l-0.115,-0.076l-0.116,-0.076l-0.117,-0.074l-0.118,-0.072l-0.119,-0.072l-0.119,-0.069l-0.121,-0.069l-0.122,-0.067l-0.122,-0.066l-0.123,-0.064l-0.124,-0.062l-0.125,-0.062l-0.126,-0.059l-0.127,-0.059l-0.128,-0.056l-0.128,-0.056l-0.129,-0.053l-0.13,-0.052l-0.131,-0.051l-0.131,-0.049l-0.132,-0.048l-0.133,-0.046l-0.134,-0.044l-0.134,-0.043l-0.135,-0.041l-0.136,-0.039l-0.137,-0.038l-0.137,-0.037l-0.138,-0.034l-0.138,-0.033l-0.139,-0.031l-0.14,-0.03l-0.14,-0.028l-0.141,-0.026l-0.142,-0.024l-0.142,-0.023l-0.143,-0.021l-0.143,-0.019l-0.144,-0.017l-0.144,-0.016l-0.145,-0.014l-0.146,-0.012l-0.145,-0.01l-0.147,-0.008l-0.147,-0.007l-0.147,-0.004l-0.148,-0.003l-0.148,-0.001l-359.028,0Z"
-    fill="#888" />
-    <circle cx="78.187" cy="200" r="59.757" fill="#444" />
-    <path d="M335.948,154.253c0,-4.187 -3.399,-7.586 -7.585,-7.586l-164.464,0c-4.187,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.398,7.585 7.585,7.585l164.464,0c4.186,0 7.585,-3.398 7.585,-7.585l0,-3.834Z"
-    fill="#444" />
-    <path d="M368.797,217.727c0,-3.137 -2.547,-5.683 -5.683,-5.683l-201.255,0c-3.137,0 -5.683,2.546 -5.683,5.683l0,2.674c0,3.137 2.546,5.684 5.683,5.684l201.255,0c3.136,0 5.683,-2.547 5.683,-5.684l0,-2.674Z"
-    fill="#888" />
-    <path d="M303.84,244.975c0,-3.137 -2.547,-5.683 -5.683,-5.683l-136.298,0c-3.137,0 -5.683,2.546 -5.683,5.683l0,2.675c0,3.136 2.546,5.683 5.683,5.683l136.298,0c3.136,0 5.683,-2.547 5.683,-5.683l0,-2.675Z"
-    fill="#888" />
-    <path d="M368.797,189.852c0,-3.137 -2.547,-5.683 -5.683,-5.683l-201.255,0c-3.137,0 -5.683,2.546 -5.683,5.683l0,2.674c0,3.137 2.546,5.684 5.683,5.684l201.255,0c3.136,0 5.683,-2.547 5.683,-5.684l0,-2.674Z"
-    fill="#888" />
-    <path d="M78.187,163.826l-8.793,26.38l-26.381,0l21.984,17.588l-8.793,26.38l21.983,-17.587l21.984,17.587l-8.793,-26.38l21.984,-17.588l-26.381,0l-8.794,-26.38Z"
-    fill="#eee" fillRule="nonzero" />
-</svg>;
-icons.infoImage = <svg viewBox="0 0 400 400" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <path d="M350.827,47.314c0,-11.515 -9.348,-20.863 -20.863,-20.863l-259.928,0c-11.515,0 -20.863,9.348 -20.863,20.863l0,305.372c0,11.515 9.348,20.863 20.863,20.863l259.928,0c11.515,0 20.863,-9.348 20.863,-20.863l0,-305.372Z"
-    fill="#fff" />
-    <path d="M330.137,20.452l0.173,0.001l0.172,0.003l0.172,0.004l0.172,0.005l0.172,0.006l0.172,0.007l0.171,0.008l0.171,0.009l0.171,0.011l0.171,0.011l0.17,0.012l0.17,0.014l0.17,0.014l0.17,0.016l0.169,0.017l0.169,0.018l0.169,0.018l0.168,0.02l0.168,0.021l0.168,0.022l0.168,0.023l0.167,0.024l0.167,0.025l0.167,0.026l0.166,0.027l0.166,0.028l0.165,0.029l0.166,0.03l0.165,0.031l0.164,0.032l0.165,0.033l0.163,0.035l0.164,0.035l0.163,0.036l0.163,0.037l0.163,0.038l0.162,0.039l0.162,0.04l0.161,0.041l0.161,0.042l0.16,0.043l0.161,0.044l0.159,0.045l0.16,0.046l0.159,0.047l0.158,0.047l0.159,0.049l0.157,0.05l0.158,0.05l0.157,0.052l0.156,0.052l0.156,0.054l0.156,0.054l0.155,0.055l0.155,0.056l0.154,0.057l0.154,0.058l0.153,0.059l0.153,0.06l0.153,0.061l0.152,0.061l0.151,0.063l0.151,0.063l0.151,0.065l0.15,0.065l0.149,0.066l0.15,0.067l0.149,0.068l0.148,0.068l0.147,0.07l0.148,0.07l0.146,0.071l0.146,0.073l0.146,0.073l0.145,0.073l0.145,0.075l0.144,0.076l0.143,0.076l0.143,0.077l0.143,0.078l0.142,0.079l0.141,0.08l0.141,0.08l0.141,0.082l0.139,0.082l0.14,0.083l0.138,0.083l0.138,0.085l0.138,0.085l0.137,0.086l0.137,0.087l0.136,0.088l0.135,0.089l0.135,0.089l0.134,0.09l0.134,0.091l0.132,0.092l0.133,0.092l0.132,0.093l0.131,0.094l0.131,0.095l0.13,0.096l0.129,0.096l0.129,0.097l0.136,0.103l0.247,0.192l0.252,0.201l0.25,0.203l0.247,0.207l0.245,0.209l0.242,0.213l0.239,0.215l0.237,0.218l0.235,0.22l0.231,0.224l0.229,0.226l0.226,0.229l0.224,0.232l0.22,0.234l0.218,0.236l0.215,0.24l0.213,0.242l0.209,0.245l0.207,0.247l0.203,0.25l0.201,0.252l0.192,0.248l0.103,0.135l0.097,0.129l0.096,0.129l0.096,0.13l0.095,0.131l0.094,0.131l0.093,0.132l0.092,0.133l0.092,0.132l0.091,0.134l0.09,0.134l0.089,0.135l0.089,0.136l0.088,0.135l0.087,0.137l0.086,0.137l0.085,0.138l0.085,0.138l0.083,0.138l0.083,0.14l0.082,0.139l0.082,0.141l0.08,0.141l0.08,0.141l0.079,0.142l0.078,0.142l0.077,0.144l0.076,0.143l0.076,0.144l0.075,0.145l0.073,0.145l0.073,0.146l0.072,0.146l0.072,0.147l0.07,0.147l0.07,0.147l0.068,0.149l0.068,0.148l0.067,0.149l0.066,0.15l0.065,0.15l0.065,0.151l0.063,0.151l0.062,0.151l0.062,0.152l0.061,0.153l0.059,0.152l0.06,0.154l0.058,0.154l0.057,0.154l0.056,0.154l0.055,0.156l0.054,0.156l0.054,0.156l0.052,0.156l0.052,0.157l0.05,0.158l0.05,0.158l0.049,0.158l0.047,0.158l0.047,0.159l0.046,0.159l0.045,0.16l0.044,0.161l0.043,0.16l0.042,0.161l0.041,0.162l0.04,0.161l0.039,0.162l0.038,0.163l0.037,0.163l0.036,0.163l0.035,0.163l0.035,0.165l0.033,0.164l0.032,0.164l0.031,0.166l0.03,0.164l0.029,0.166l0.028,0.166l0.027,0.166l0.026,0.167l0.025,0.167l0.024,0.167l0.023,0.168l0.022,0.168l0.021,0.168l0.02,0.168l0.018,0.168l0.018,0.17l0.017,0.17l0.015,0.169l0.015,0.17l0.014,0.17l0.012,0.17l0.011,0.171l0.011,0.171l0.009,0.17l0.008,0.172l0.007,0.172l0.006,0.172l0.005,0.172l0.004,0.171l0.003,0.173l0.001,0.173l0.001,0.163l0,305.392l-0.001,0.163l-0.001,0.173l-0.003,0.173l-0.004,0.172l-0.005,0.171l-0.006,0.172l-0.007,0.172l-0.008,0.172l-0.009,0.17l-0.011,0.171l-0.011,0.171l-0.012,0.17l-0.014,0.17l-0.015,0.17l-0.015,0.169l-0.017,0.17l-0.018,0.17l-0.018,0.167l-0.02,0.169l-0.021,0.168l-0.022,0.168l-0.023,0.168l-0.024,0.167l-0.025,0.167l-0.026,0.167l-0.027,0.166l-0.028,0.166l-0.029,0.166l-0.03,0.164l-0.031,0.166l-0.032,0.164l-0.033,0.164l-0.035,0.165l-0.035,0.163l-0.036,0.163l-0.037,0.163l-0.038,0.163l-0.039,0.162l-0.04,0.161l-0.041,0.162l-0.042,0.161l-0.043,0.16l-0.044,0.161l-0.045,0.16l-0.046,0.159l-0.047,0.159l-0.047,0.158l-0.049,0.158l-0.05,0.158l-0.05,0.158l-0.052,0.157l-0.052,0.156l-0.054,0.156l-0.054,0.156l-0.055,0.155l-0.056,0.155l-0.057,0.154l-0.058,0.154l-0.059,0.154l-0.06,0.152l-0.061,0.153l-0.061,0.152l-0.063,0.151l-0.063,0.151l-0.065,0.151l-0.065,0.15l-0.066,0.15l-0.067,0.149l-0.068,0.148l-0.068,0.149l-0.07,0.147l-0.07,0.148l-0.071,0.146l-0.073,0.146l-0.073,0.146l-0.073,0.145l-0.075,0.145l-0.076,0.144l-0.076,0.143l-0.077,0.144l-0.078,0.142l-0.079,0.142l-0.079,0.141l-0.081,0.142l-0.082,0.14l-0.082,0.139l-0.083,0.14l-0.083,0.138l-0.085,0.138l-0.085,0.138l-0.086,0.137l-0.087,0.137l-0.088,0.136l-0.089,0.135l-0.089,0.135l-0.09,0.134l-0.091,0.134l-0.092,0.132l-0.092,0.133l-0.093,0.132l-0.094,0.131l-0.095,0.131l-0.096,0.13l-0.096,0.129l-0.097,0.129l-0.103,0.136l-0.192,0.247l-0.201,0.252l-0.203,0.25l-0.207,0.247l-0.209,0.245l-0.213,0.242l-0.215,0.24l-0.217,0.236l-0.221,0.235l-0.224,0.231l-0.226,0.229l-0.229,0.226l-0.232,0.224l-0.234,0.22l-0.236,0.218l-0.24,0.215l-0.242,0.213l-0.245,0.209l-0.247,0.207l-0.25,0.203l-0.252,0.201l-0.248,0.192l-0.135,0.103l-0.129,0.097l-0.129,0.096l-0.13,0.096l-0.131,0.095l-0.131,0.094l-0.132,0.093l-0.133,0.093l-0.132,0.091l-0.134,0.091l-0.134,0.09l-0.135,0.089l-0.135,0.089l-0.136,0.088l-0.137,0.087l-0.137,0.086l-0.138,0.085l-0.138,0.085l-0.138,0.083l-0.14,0.083l-0.139,0.082l-0.14,0.082l-0.142,0.081l-0.141,0.079l-0.142,0.079l-0.142,0.078l-0.144,0.077l-0.143,0.076l-0.144,0.076l-0.145,0.075l-0.145,0.074l-0.146,0.072l-0.146,0.073l-0.146,0.071l-0.148,0.07l-0.147,0.07l-0.149,0.068l-0.148,0.068l-0.15,0.067l-0.149,0.066l-0.149,0.065l-0.152,0.065l-0.151,0.063l-0.151,0.063l-0.152,0.061l-0.153,0.061l-0.152,0.06l-0.154,0.059l-0.154,0.058l-0.154,0.057l-0.155,0.056l-0.155,0.055l-0.156,0.054l-0.156,0.054l-0.156,0.052l-0.157,0.052l-0.158,0.05l-0.158,0.05l-0.158,0.049l-0.158,0.047l-0.159,0.047l-0.159,0.046l-0.16,0.045l-0.161,0.044l-0.16,0.043l-0.161,0.042l-0.162,0.041l-0.161,0.04l-0.162,0.039l-0.163,0.038l-0.163,0.037l-0.163,0.036l-0.163,0.035l-0.165,0.035l-0.164,0.033l-0.164,0.032l-0.166,0.031l-0.164,0.03l-0.166,0.029l-0.167,0.028l-0.165,0.027l-0.167,0.026l-0.167,0.025l-0.167,0.024l-0.168,0.023l-0.168,0.022l-0.168,0.021l-0.169,0.02l-0.167,0.018l-0.17,0.018l-0.17,0.017l-0.169,0.015l-0.17,0.015l-0.17,0.014l-0.17,0.012l-0.171,0.011l-0.171,0.011l-0.17,0.009l-0.172,0.008l-0.172,0.007l-0.172,0.006l-0.172,0.005l-0.171,0.004l-0.173,0.003l-0.173,0.001l-0.163,0.001l-259.948,0l-0.163,-0.001l-0.173,-0.001l-0.173,-0.003l-0.171,-0.004l-0.172,-0.005l-0.172,-0.006l-0.172,-0.007l-0.172,-0.008l-0.17,-0.009l-0.171,-0.011l-0.171,-0.011l-0.17,-0.012l-0.17,-0.014l-0.17,-0.015l-0.169,-0.015l-0.17,-0.017l-0.17,-0.018l-0.167,-0.018l-0.169,-0.02l-0.168,-0.021l-0.168,-0.022l-0.168,-0.023l-0.167,-0.024l-0.167,-0.025l-0.167,-0.026l-0.165,-0.027l-0.167,-0.028l-0.166,-0.029l-0.164,-0.03l-0.166,-0.031l-0.164,-0.032l-0.164,-0.033l-0.165,-0.035l-0.163,-0.035l-0.163,-0.036l-0.163,-0.037l-0.163,-0.038l-0.162,-0.039l-0.161,-0.04l-0.162,-0.041l-0.161,-0.042l-0.16,-0.043l-0.161,-0.044l-0.159,-0.045l-0.16,-0.046l-0.159,-0.047l-0.158,-0.047l-0.158,-0.049l-0.158,-0.05l-0.158,-0.05l-0.157,-0.052l-0.156,-0.052l-0.156,-0.054l-0.156,-0.054l-0.155,-0.055l-0.155,-0.056l-0.154,-0.057l-0.154,-0.058l-0.154,-0.06l-0.152,-0.059l-0.153,-0.061l-0.152,-0.061l-0.151,-0.063l-0.151,-0.063l-0.151,-0.065l-0.15,-0.065l-0.149,-0.066l-0.15,-0.067l-0.148,-0.067l-0.149,-0.069l-0.147,-0.07l-0.147,-0.07l-0.147,-0.071l-0.146,-0.073l-0.146,-0.072l-0.145,-0.074l-0.145,-0.075l-0.144,-0.076l-0.143,-0.076l-0.144,-0.077l-0.142,-0.078l-0.142,-0.079l-0.141,-0.079l-0.141,-0.081l-0.141,-0.082l-0.139,-0.082l-0.14,-0.083l-0.138,-0.083l-0.138,-0.085l-0.138,-0.085l-0.137,-0.086l-0.137,-0.087l-0.135,-0.088l-0.136,-0.089l-0.135,-0.089l-0.134,-0.09l-0.134,-0.091l-0.132,-0.091l-0.133,-0.093l-0.132,-0.093l-0.131,-0.094l-0.131,-0.095l-0.13,-0.096l-0.129,-0.096l-0.129,-0.096l-0.135,-0.104l-0.248,-0.192l-0.252,-0.201l-0.25,-0.203l-0.247,-0.207l-0.245,-0.209l-0.242,-0.213l-0.24,-0.215l-0.236,-0.218l-0.234,-0.22l-0.232,-0.224l-0.229,-0.226l-0.226,-0.229l-0.224,-0.231l-0.221,-0.235l-0.217,-0.237l-0.215,-0.239l-0.213,-0.242l-0.209,-0.245l-0.207,-0.247l-0.203,-0.25l-0.201,-0.252l-0.192,-0.247l-0.103,-0.136l-0.097,-0.129l-0.097,-0.129l-0.095,-0.13l-0.095,-0.131l-0.094,-0.131l-0.093,-0.132l-0.092,-0.132l-0.092,-0.133l-0.091,-0.134l-0.09,-0.134l-0.089,-0.135l-0.089,-0.135l-0.088,-0.136l-0.086,-0.137l-0.087,-0.137l-0.085,-0.138l-0.085,-0.138l-0.083,-0.138l-0.083,-0.14l-0.082,-0.139l-0.082,-0.141l-0.08,-0.141l-0.08,-0.141l-0.079,-0.142l-0.078,-0.143l-0.077,-0.143l-0.076,-0.143l-0.076,-0.144l-0.075,-0.145l-0.073,-0.145l-0.073,-0.146l-0.073,-0.146l-0.071,-0.146l-0.07,-0.148l-0.07,-0.147l-0.068,-0.148l-0.068,-0.149l-0.067,-0.15l-0.066,-0.149l-0.065,-0.15l-0.065,-0.151l-0.063,-0.151l-0.063,-0.151l-0.061,-0.152l-0.061,-0.153l-0.06,-0.153l-0.059,-0.153l-0.057,-0.154l-0.058,-0.154l-0.056,-0.155l-0.055,-0.155l-0.054,-0.156l-0.054,-0.156l-0.052,-0.156l-0.052,-0.157l-0.05,-0.158l-0.05,-0.157l-0.049,-0.158l-0.047,-0.159l-0.047,-0.159l-0.046,-0.16l-0.045,-0.159l-0.044,-0.161l-0.043,-0.16l-0.042,-0.161l-0.041,-0.161l-0.04,-0.162l-0.039,-0.162l-0.038,-0.163l-0.037,-0.163l-0.036,-0.163l-0.035,-0.163l-0.035,-0.164l-0.033,-0.165l-0.032,-0.164l-0.031,-0.165l-0.03,-0.166l-0.029,-0.165l-0.028,-0.166l-0.027,-0.166l-0.026,-0.167l-0.025,-0.167l-0.024,-0.167l-0.023,-0.168l-0.022,-0.168l-0.021,-0.168l-0.02,-0.168l-0.018,-0.169l-0.018,-0.169l-0.017,-0.169l-0.015,-0.17l-0.015,-0.17l-0.014,-0.17l-0.012,-0.17l-0.011,-0.171l-0.011,-0.171l-0.009,-0.171l-0.008,-0.171l-0.007,-0.172l-0.006,-0.172l-0.005,-0.171l-0.004,-0.173l-0.003,-0.172l-0.001,-0.173l-0.001,-0.163l0,-305.392l0.001,-0.163l0.001,-0.173l0.003,-0.172l0.004,-0.172l0.005,-0.172l0.006,-0.172l0.007,-0.172l0.008,-0.171l0.009,-0.171l0.011,-0.171l0.011,-0.171l0.012,-0.17l0.014,-0.17l0.015,-0.17l0.015,-0.17l0.017,-0.169l0.018,-0.169l0.018,-0.169l0.02,-0.168l0.021,-0.168l0.022,-0.168l0.023,-0.168l0.024,-0.167l0.025,-0.167l0.026,-0.167l0.027,-0.166l0.028,-0.166l0.029,-0.165l0.03,-0.166l0.031,-0.165l0.032,-0.164l0.033,-0.165l0.035,-0.164l0.035,-0.163l0.036,-0.163l0.037,-0.163l0.038,-0.163l0.039,-0.162l0.04,-0.162l0.041,-0.161l0.042,-0.161l0.043,-0.16l0.044,-0.161l0.045,-0.159l0.046,-0.16l0.047,-0.159l0.047,-0.158l0.049,-0.159l0.05,-0.157l0.05,-0.158l0.052,-0.157l0.052,-0.156l0.054,-0.156l0.054,-0.156l0.055,-0.155l0.056,-0.155l0.057,-0.154l0.058,-0.154l0.059,-0.153l0.06,-0.153l0.061,-0.153l0.061,-0.152l0.063,-0.151l0.063,-0.151l0.065,-0.151l0.065,-0.15l0.066,-0.15l0.067,-0.149l0.068,-0.148l0.068,-0.149l0.07,-0.147l0.07,-0.147l0.072,-0.147l0.072,-0.146l0.073,-0.146l0.073,-0.145l0.075,-0.145l0.076,-0.144l0.076,-0.143l0.077,-0.143l0.078,-0.143l0.079,-0.142l0.08,-0.141l0.08,-0.141l0.082,-0.141l0.082,-0.139l0.083,-0.14l0.083,-0.138l0.085,-0.139l0.085,-0.137l0.086,-0.137l0.087,-0.137l0.088,-0.136l0.089,-0.135l0.089,-0.135l0.09,-0.134l0.091,-0.133l0.092,-0.134l0.092,-0.132l0.093,-0.132l0.094,-0.131l0.095,-0.131l0.095,-0.13l0.097,-0.129l0.097,-0.129l0.103,-0.136l0.192,-0.247l0.201,-0.252l0.203,-0.25l0.207,-0.247l0.209,-0.245l0.213,-0.242l0.215,-0.239l0.218,-0.237l0.22,-0.235l0.224,-0.231l0.226,-0.229l0.229,-0.226l0.231,-0.224l0.235,-0.22l0.237,-0.218l0.239,-0.215l0.242,-0.213l0.245,-0.209l0.247,-0.207l0.25,-0.203l0.252,-0.201l0.248,-0.192l0.135,-0.103l0.129,-0.097l0.129,-0.097l0.13,-0.095l0.131,-0.095l0.131,-0.094l0.132,-0.093l0.132,-0.092l0.134,-0.092l0.133,-0.091l0.134,-0.09l0.135,-0.089l0.135,-0.089l0.136,-0.088l0.137,-0.087l0.137,-0.086l0.137,-0.085l0.139,-0.085l0.138,-0.083l0.14,-0.083l0.139,-0.083l0.141,-0.081l0.141,-0.08l0.141,-0.08l0.142,-0.079l0.143,-0.078l0.143,-0.077l0.143,-0.076l0.144,-0.076l0.145,-0.075l0.145,-0.073l0.146,-0.073l0.146,-0.073l0.147,-0.071l0.147,-0.07l0.147,-0.07l0.149,-0.068l0.148,-0.068l0.149,-0.067l0.15,-0.066l0.15,-0.065l0.151,-0.065l0.151,-0.063l0.151,-0.063l0.152,-0.061l0.153,-0.061l0.153,-0.06l0.153,-0.059l0.154,-0.058l0.154,-0.057l0.155,-0.056l0.155,-0.055l0.156,-0.054l0.156,-0.054l0.156,-0.052l0.157,-0.052l0.158,-0.05l0.157,-0.05l0.159,-0.049l0.158,-0.047l0.159,-0.047l0.16,-0.046l0.159,-0.045l0.161,-0.044l0.16,-0.043l0.161,-0.042l0.162,-0.041l0.161,-0.04l0.162,-0.039l0.163,-0.038l0.163,-0.037l0.163,-0.036l0.163,-0.035l0.164,-0.035l0.165,-0.033l0.164,-0.032l0.165,-0.031l0.166,-0.03l0.165,-0.029l0.166,-0.028l0.167,-0.027l0.166,-0.026l0.167,-0.025l0.167,-0.024l0.168,-0.023l0.168,-0.022l0.168,-0.021l0.168,-0.02l0.169,-0.018l0.169,-0.018l0.169,-0.017l0.17,-0.016l0.17,-0.014l0.17,-0.014l0.17,-0.012l0.171,-0.011l0.171,-0.011l0.171,-0.009l0.171,-0.008l0.172,-0.007l0.172,-0.006l0.172,-0.005l0.172,-0.004l0.172,-0.003l0.173,-0.001l0.163,-0.001l259.948,0l0.163,0.001Zm-260.101,5.999l-0.135,0.001l-0.135,0.001l-0.134,0.002l-0.134,0.003l-0.134,0.004l-0.134,0.005l-0.134,0.005l-0.133,0.006l-0.133,0.008l-0.133,0.008l-0.133,0.008l-0.132,0.01l-0.133,0.011l-0.132,0.011l-0.131,0.012l-0.132,0.013l-0.131,0.014l-0.131,0.014l-0.131,0.016l-0.131,0.016l-0.13,0.017l-0.13,0.018l-0.13,0.018l-0.13,0.02l-0.129,0.02l-0.129,0.021l-0.129,0.022l-0.128,0.022l-0.128,0.024l-0.128,0.024l-0.128,0.025l-0.127,0.025l-0.128,0.027l-0.126,0.027l-0.127,0.028l-0.126,0.029l-0.126,0.029l-0.126,0.031l-0.125,0.031l-0.125,0.032l-0.125,0.032l-0.125,0.033l-0.124,0.034l-0.124,0.035l-0.123,0.036l-0.124,0.036l-0.123,0.037l-0.122,0.038l-0.122,0.038l-0.122,0.04l-0.122,0.039l-0.121,0.041l-0.121,0.042l-0.121,0.042l-0.12,0.042l-0.12,0.044l-0.12,0.044l-0.119,0.045l-0.119,0.046l-0.119,0.046l-0.118,0.047l-0.118,0.048l-0.117,0.049l-0.118,0.049l-0.116,0.05l-0.117,0.05l-0.116,0.051l-0.116,0.052l-0.115,0.053l-0.115,0.053l-0.114,0.054l-0.115,0.055l-0.113,0.055l-0.114,0.056l-0.113,0.057l-0.112,0.057l-0.113,0.058l-0.112,0.059l-0.111,0.059l-0.111,0.06l-0.111,0.06l-0.11,0.062l-0.11,0.061l-0.109,0.063l-0.109,0.063l-0.109,0.064l-0.108,0.064l-0.108,0.065l-0.107,0.066l-0.107,0.066l-0.106,0.067l-0.106,0.068l-0.106,0.068l-0.105,0.069l-0.105,0.069l-0.104,0.07l-0.104,0.071l-0.103,0.071l-0.103,0.072l-0.103,0.072l-0.102,0.073l-0.101,0.074l-0.102,0.074l-0.1,0.075l-0.1,0.076l-0.1,0.076l-0.198,0.153l-0.196,0.156l-0.194,0.159l-0.193,0.16l-0.19,0.163l-0.188,0.165l-0.187,0.168l-0.184,0.169l-0.182,0.172l-0.18,0.174l-0.178,0.176l-0.176,0.178l-0.174,0.18l-0.172,0.182l-0.169,0.184l-0.168,0.187l-0.165,0.188l-0.163,0.19l-0.16,0.192l-0.159,0.195l-0.156,0.196l-0.153,0.198l-0.076,0.1l-0.076,0.1l-0.075,0.1l-0.074,0.101l-0.073,0.102l-0.074,0.102l-0.072,0.103l-0.072,0.103l-0.071,0.103l-0.071,0.104l-0.07,0.104l-0.069,0.105l-0.069,0.105l-0.068,0.106l-0.068,0.106l-0.067,0.106l-0.066,0.107l-0.066,0.107l-0.065,0.108l-0.064,0.108l-0.064,0.109l-0.063,0.109l-0.063,0.109l-0.061,0.11l-0.062,0.11l-0.06,0.111l-0.06,0.111l-0.059,0.111l-0.059,0.112l-0.058,0.112l-0.057,0.113l-0.057,0.113l-0.056,0.114l-0.055,0.113l-0.055,0.115l-0.054,0.114l-0.053,0.115l-0.053,0.115l-0.052,0.116l-0.051,0.116l-0.05,0.117l-0.05,0.116l-0.049,0.118l-0.049,0.117l-0.048,0.118l-0.047,0.118l-0.046,0.119l-0.046,0.119l-0.045,0.119l-0.044,0.12l-0.043,0.12l-0.043,0.12l-0.042,0.121l-0.042,0.121l-0.04,0.121l-0.04,0.122l-0.04,0.121l-0.038,0.123l-0.038,0.122l-0.037,0.123l-0.036,0.124l-0.036,0.123l-0.035,0.124l-0.034,0.124l-0.033,0.125l-0.032,0.125l-0.032,0.125l-0.031,0.125l-0.031,0.126l-0.029,0.126l-0.029,0.126l-0.028,0.127l-0.027,0.126l-0.027,0.128l-0.025,0.127l-0.025,0.128l-0.024,0.128l-0.024,0.128l-0.022,0.128l-0.022,0.129l-0.021,0.129l-0.02,0.129l-0.02,0.13l-0.018,0.13l-0.018,0.13l-0.017,0.13l-0.016,0.131l-0.016,0.131l-0.014,0.131l-0.014,0.131l-0.013,0.132l-0.012,0.131l-0.011,0.132l-0.011,0.133l-0.01,0.132l-0.008,0.133l-0.008,0.133l-0.008,0.133l-0.006,0.133l-0.005,0.134l-0.005,0.134l-0.004,0.134l-0.003,0.134l-0.002,0.134l-0.001,0.135l-0.001,0.135l0,305.372l0.001,0.135l0.001,0.135l0.002,0.134l0.003,0.134l0.004,0.134l0.01,0.268l0.014,0.266l0.016,0.266l0.021,0.265l0.023,0.263l0.027,0.263l0.03,0.262l0.033,0.261l0.036,0.26l0.04,0.259l0.043,0.258l0.046,0.256l0.049,0.256l0.052,0.255l0.055,0.253l0.058,0.252l0.062,0.251l0.064,0.25l0.067,0.249l0.071,0.247l0.073,0.247l0.076,0.245l0.08,0.243l0.082,0.242l0.085,0.241l0.087,0.24l0.091,0.238l0.093,0.237l0.097,0.235l0.099,0.234l0.101,0.233l0.105,0.231l0.107,0.229l0.11,0.228l0.113,0.227l0.115,0.225l0.118,0.223l0.12,0.222l0.123,0.22l0.126,0.218l0.128,0.217l0.131,0.215l0.133,0.213l0.136,0.212l0.138,0.21l0.141,0.208l0.143,0.206l0.146,0.205l0.147,0.203l0.151,0.2l0.076,0.1l0.153,0.198l0.156,0.196l0.159,0.195l0.16,0.192l0.163,0.19l0.165,0.188l0.168,0.187l0.169,0.184l0.172,0.182l0.174,0.18l0.176,0.178l0.178,0.176l0.18,0.174l0.182,0.172l0.184,0.169l0.187,0.168l0.188,0.165l0.19,0.163l0.193,0.16l0.194,0.159l0.196,0.156l0.198,0.153l0.2,0.152l0.202,0.149l0.203,0.147l0.206,0.144l0.207,0.142l0.209,0.139l0.211,0.137l0.212,0.135l0.214,0.132l0.216,0.129l0.218,0.127l0.219,0.124l0.221,0.122l0.222,0.119l0.225,0.117l0.225,0.114l0.227,0.111l0.229,0.109l0.23,0.106l0.232,0.103l0.233,0.1l0.235,0.098l0.236,0.095l0.238,0.092l0.239,0.089l0.24,0.086l0.242,0.084l0.243,0.08l0.244,0.078l0.245,0.075l0.247,0.072l0.248,0.069l0.25,0.065l0.25,0.063l0.252,0.06l0.253,0.057l0.254,0.054l0.255,0.05l0.256,0.048l0.257,0.044l0.258,0.041l0.26,0.038l0.26,0.035l0.262,0.032l0.262,0.028l0.263,0.025l0.265,0.022l0.265,0.018l0.266,0.016l0.267,0.011l0.134,0.005l0.134,0.004l0.134,0.003l0.134,0.002l0.135,0.001l0.135,0.001l259.928,0l0.135,-0.001l0.135,-0.001l0.134,-0.002l0.134,-0.003l0.134,-0.004l0.134,-0.005l0.134,-0.005l0.133,-0.006l0.133,-0.008l0.133,-0.008l0.133,-0.008l0.132,-0.01l0.133,-0.011l0.132,-0.011l0.131,-0.012l0.132,-0.013l0.131,-0.014l0.131,-0.014l0.131,-0.016l0.131,-0.016l0.13,-0.017l0.13,-0.018l0.13,-0.018l0.13,-0.02l0.129,-0.02l0.129,-0.021l0.129,-0.022l0.128,-0.022l0.128,-0.024l0.128,-0.024l0.128,-0.025l0.127,-0.025l0.128,-0.027l0.126,-0.027l0.127,-0.028l0.126,-0.029l0.126,-0.029l0.126,-0.031l0.125,-0.031l0.125,-0.032l0.125,-0.032l0.125,-0.033l0.124,-0.034l0.124,-0.035l0.123,-0.036l0.124,-0.036l0.123,-0.037l0.122,-0.038l0.122,-0.038l0.122,-0.04l0.122,-0.039l0.121,-0.041l0.121,-0.042l0.121,-0.042l0.12,-0.042l0.12,-0.044l0.12,-0.044l0.119,-0.045l0.119,-0.046l0.119,-0.046l0.118,-0.047l0.118,-0.048l0.117,-0.049l0.118,-0.049l0.116,-0.05l0.117,-0.05l0.116,-0.051l0.116,-0.052l0.115,-0.053l0.115,-0.053l0.114,-0.054l0.115,-0.055l0.113,-0.055l0.114,-0.056l0.113,-0.057l0.112,-0.057l0.113,-0.058l0.112,-0.059l0.111,-0.059l0.111,-0.06l0.111,-0.06l0.11,-0.062l0.11,-0.061l0.109,-0.063l0.109,-0.063l0.109,-0.064l0.108,-0.064l0.108,-0.065l0.107,-0.066l0.107,-0.066l0.106,-0.067l0.106,-0.068l0.106,-0.068l0.105,-0.069l0.105,-0.069l0.104,-0.07l0.104,-0.071l0.103,-0.071l0.103,-0.072l0.103,-0.072l0.102,-0.073l0.101,-0.074l0.102,-0.074l0.1,-0.075l0.1,-0.076l0.1,-0.076l0.198,-0.153l0.196,-0.156l0.194,-0.159l0.193,-0.16l0.19,-0.163l0.188,-0.165l0.187,-0.168l0.184,-0.169l0.182,-0.172l0.18,-0.174l0.178,-0.176l0.176,-0.178l0.174,-0.18l0.172,-0.182l0.169,-0.184l0.168,-0.187l0.165,-0.188l0.163,-0.19l0.16,-0.192l0.159,-0.195l0.156,-0.196l0.153,-0.198l0.076,-0.1l0.076,-0.1l0.075,-0.1l0.074,-0.101l0.073,-0.102l0.074,-0.102l0.072,-0.103l0.072,-0.103l0.071,-0.103l0.071,-0.104l0.07,-0.104l0.069,-0.105l0.069,-0.105l0.068,-0.106l0.068,-0.106l0.067,-0.106l0.066,-0.107l0.066,-0.107l0.065,-0.108l0.064,-0.108l0.064,-0.109l0.063,-0.109l0.063,-0.109l0.061,-0.11l0.062,-0.11l0.06,-0.111l0.06,-0.111l0.059,-0.111l0.059,-0.112l0.058,-0.112l0.057,-0.113l0.057,-0.113l0.056,-0.114l0.055,-0.113l0.055,-0.115l0.054,-0.114l0.053,-0.115l0.053,-0.115l0.052,-0.116l0.051,-0.116l0.05,-0.117l0.05,-0.116l0.049,-0.118l0.049,-0.117l0.048,-0.118l0.047,-0.118l0.046,-0.119l0.046,-0.119l0.045,-0.119l0.044,-0.12l0.043,-0.12l0.043,-0.12l0.042,-0.121l0.042,-0.121l0.04,-0.121l0.04,-0.122l0.04,-0.121l0.038,-0.123l0.038,-0.122l0.037,-0.123l0.036,-0.124l0.036,-0.123l0.035,-0.124l0.034,-0.124l0.033,-0.125l0.032,-0.125l0.032,-0.125l0.031,-0.125l0.031,-0.126l0.029,-0.126l0.029,-0.126l0.028,-0.127l0.027,-0.126l0.027,-0.128l0.025,-0.127l0.025,-0.128l0.024,-0.128l0.024,-0.128l0.022,-0.128l0.022,-0.129l0.021,-0.129l0.02,-0.129l0.02,-0.13l0.018,-0.13l0.018,-0.13l0.017,-0.13l0.016,-0.131l0.016,-0.131l0.014,-0.131l0.014,-0.131l0.013,-0.132l0.012,-0.131l0.011,-0.132l0.011,-0.133l0.01,-0.132l0.008,-0.133l0.008,-0.133l0.008,-0.133l0.006,-0.133l0.005,-0.134l0.005,-0.134l0.004,-0.134l0.003,-0.134l0.002,-0.134l0.001,-0.135l0.001,-0.135l0,-305.372l-0.001,-0.135l-0.001,-0.135l-0.002,-0.134l-0.003,-0.134l-0.004,-0.134l-0.005,-0.134l-0.005,-0.134l-0.006,-0.133l-0.008,-0.133l-0.008,-0.133l-0.008,-0.133l-0.01,-0.132l-0.011,-0.133l-0.011,-0.132l-0.012,-0.131l-0.013,-0.132l-0.014,-0.131l-0.014,-0.131l-0.016,-0.131l-0.016,-0.131l-0.017,-0.13l-0.018,-0.13l-0.018,-0.13l-0.02,-0.13l-0.02,-0.129l-0.021,-0.129l-0.022,-0.129l-0.022,-0.128l-0.024,-0.128l-0.024,-0.128l-0.025,-0.128l-0.025,-0.127l-0.027,-0.128l-0.027,-0.126l-0.028,-0.127l-0.029,-0.126l-0.029,-0.126l-0.031,-0.126l-0.031,-0.125l-0.032,-0.125l-0.032,-0.125l-0.033,-0.125l-0.034,-0.124l-0.035,-0.124l-0.036,-0.123l-0.036,-0.124l-0.037,-0.123l-0.038,-0.122l-0.038,-0.123l-0.04,-0.121l-0.04,-0.122l-0.04,-0.121l-0.042,-0.121l-0.042,-0.121l-0.043,-0.12l-0.043,-0.12l-0.044,-0.12l-0.045,-0.119l-0.046,-0.119l-0.046,-0.119l-0.047,-0.118l-0.048,-0.118l-0.049,-0.117l-0.049,-0.118l-0.05,-0.116l-0.05,-0.117l-0.051,-0.116l-0.052,-0.116l-0.053,-0.115l-0.053,-0.115l-0.054,-0.114l-0.055,-0.115l-0.055,-0.113l-0.056,-0.114l-0.057,-0.113l-0.057,-0.113l-0.058,-0.112l-0.059,-0.112l-0.059,-0.111l-0.06,-0.111l-0.06,-0.111l-0.062,-0.11l-0.061,-0.11l-0.063,-0.109l-0.063,-0.109l-0.064,-0.109l-0.064,-0.108l-0.065,-0.108l-0.066,-0.107l-0.066,-0.107l-0.067,-0.106l-0.068,-0.106l-0.068,-0.106l-0.069,-0.105l-0.069,-0.105l-0.07,-0.104l-0.071,-0.104l-0.071,-0.103l-0.072,-0.103l-0.072,-0.103l-0.074,-0.102l-0.073,-0.102l-0.074,-0.101l-0.075,-0.1l-0.076,-0.1l-0.076,-0.1l-0.153,-0.198l-0.156,-0.196l-0.159,-0.195l-0.16,-0.192l-0.163,-0.19l-0.165,-0.188l-0.168,-0.187l-0.169,-0.184l-0.172,-0.182l-0.174,-0.18l-0.176,-0.178l-0.178,-0.176l-0.18,-0.174l-0.182,-0.172l-0.184,-0.169l-0.187,-0.168l-0.188,-0.165l-0.19,-0.163l-0.193,-0.16l-0.194,-0.159l-0.196,-0.156l-0.198,-0.153l-0.1,-0.076l-0.1,-0.076l-0.1,-0.075l-0.102,-0.074l-0.101,-0.074l-0.102,-0.073l-0.103,-0.072l-0.103,-0.072l-0.103,-0.071l-0.104,-0.071l-0.104,-0.07l-0.105,-0.069l-0.105,-0.069l-0.106,-0.068l-0.106,-0.068l-0.106,-0.067l-0.107,-0.066l-0.107,-0.066l-0.108,-0.065l-0.108,-0.064l-0.109,-0.064l-0.109,-0.063l-0.109,-0.063l-0.11,-0.061l-0.11,-0.062l-0.111,-0.06l-0.111,-0.06l-0.111,-0.059l-0.112,-0.059l-0.113,-0.058l-0.112,-0.057l-0.113,-0.057l-0.114,-0.056l-0.113,-0.055l-0.115,-0.055l-0.114,-0.054l-0.115,-0.053l-0.115,-0.053l-0.116,-0.052l-0.116,-0.051l-0.117,-0.05l-0.116,-0.05l-0.118,-0.049l-0.117,-0.049l-0.118,-0.048l-0.118,-0.047l-0.119,-0.046l-0.119,-0.046l-0.119,-0.045l-0.12,-0.044l-0.12,-0.044l-0.12,-0.042l-0.121,-0.042l-0.121,-0.042l-0.121,-0.041l-0.122,-0.039l-0.122,-0.04l-0.122,-0.038l-0.122,-0.038l-0.123,-0.037l-0.124,-0.036l-0.123,-0.036l-0.124,-0.035l-0.124,-0.034l-0.125,-0.033l-0.125,-0.032l-0.125,-0.032l-0.125,-0.031l-0.126,-0.031l-0.126,-0.029l-0.126,-0.029l-0.127,-0.028l-0.126,-0.027l-0.128,-0.027l-0.127,-0.025l-0.128,-0.025l-0.128,-0.024l-0.128,-0.024l-0.128,-0.022l-0.129,-0.022l-0.129,-0.021l-0.129,-0.02l-0.13,-0.02l-0.13,-0.018l-0.13,-0.018l-0.13,-0.017l-0.131,-0.016l-0.131,-0.016l-0.131,-0.014l-0.131,-0.014l-0.132,-0.013l-0.131,-0.012l-0.132,-0.011l-0.133,-0.011l-0.132,-0.01l-0.133,-0.008l-0.133,-0.008l-0.133,-0.008l-0.133,-0.006l-0.134,-0.005l-0.134,-0.005l-0.134,-0.004l-0.134,-0.003l-0.134,-0.002l-0.135,-0.001l-0.135,-0.001l-259.928,0Z"
-    fill="#eee" />
-    <circle cx="200" cy="119.069" r="70.019" fill="#eee" />
-    <path d="M289.817,213.579c0,-4.187 -3.399,-7.586 -7.585,-7.586l-164.464,0c-4.186,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.399,7.586 7.585,7.586l164.464,0c4.186,0 7.585,-3.399 7.585,-7.586l0,-3.834Z"
-    fill="#444" />
-    <path d="M245.908,342.578c0,-3.085 -2.505,-5.59 -5.59,-5.59l-80.636,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l80.636,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z"
-    fill="#888" />
-    <path d="M306.311,292.784c0,-3.137 -2.547,-5.683 -5.684,-5.683l-201.254,0c-3.137,0 -5.684,2.546 -5.684,5.683l0,2.674c0,3.137 2.547,5.684 5.684,5.684l201.254,0c3.137,0 5.684,-2.547 5.684,-5.684l0,-2.674Z"
-    fill="#888" />
-    <path d="M326.619,270.022c0,-3.136 -2.547,-5.683 -5.683,-5.683l-248.595,0c-3.136,0 -5.683,2.547 -5.683,5.683l0,2.675c0,3.136 2.547,5.683 5.683,5.683l248.595,0c3.136,0 5.683,-2.547 5.683,-5.683l0,-2.675Z"
-    fill="#888" />
-    <path d="M326.619,247.261c0,-3.137 -2.547,-5.684 -5.683,-5.684l-248.595,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l248.595,0c3.136,0 5.683,-2.546 5.683,-5.683l0,-2.674Z"
-    fill="#888" />
-    <path d="M250.396,319.84c0,-0.453 -0.367,-0.82 -0.82,-0.82l-99.152,0c-0.453,0 -0.82,0.367 -0.82,0.82l0,0.41c0,0.452 0.367,0.82 0.82,0.82l99.152,0c0.453,0 0.82,-0.368 0.82,-0.82l0,-0.41Z"
-    fill="#676767" />
-    <path d="M176.939,94.182l46.122,0c2.053,0 3.719,1.667 3.719,3.72l0,46.121c0,2.053 -1.666,3.719 -3.719,3.719l-46.122,0c-2.053,0 -3.719,-1.666 -3.719,-3.719l0,-46.121c0,-2.053 1.666,-3.72 3.719,-3.72Zm43.89,47.609l0,-41.658l-41.658,0l0,41.658l41.658,0Zm-20.829,-32.731c0,-3.273 -2.678,-5.951 -5.951,-5.951c-3.273,0 -5.951,2.678 -5.951,5.951c0,3.273 2.678,5.951 5.951,5.951c3.273,0 5.951,-2.678 5.951,-5.951Zm8.927,14.878c0,0 0,-17.854 8.926,-17.854l0,29.756c0,1.637 -1.339,2.976 -2.975,2.976l-29.756,0c-1.636,0 -2.975,-1.339 -2.975,-2.976l0,-20.829c5.951,0 8.926,11.902 8.926,11.902c0,0 2.976,-8.926 8.927,-8.926c5.951,0 8.927,5.951 8.927,5.951Z"
-    fill="#888" fillRule="nonzero" />
-</svg>;
-icons.infoBackground = <svg viewBox="0 0 400 400" xmlns="http://www.w3.org/2000/svg" fillRule="evenodd"
-clipRule="evenodd" strokeLinejoin="round" strokeMiterlimit="1.414">
-    <rect x="49.173" y="68.261" width="301.653" height="263.478" fill="#0a6689"
-    />
-    <g fill="#fff">
-        <path d="M289.818,251.349c0,-4.195 -3.406,-7.601 -7.6,-7.601l-164.435,0c-4.194,0 -7.6,3.406 -7.6,7.601l0,3.8c0,4.195 3.406,7.601 7.6,7.601l164.435,0c4.194,0 7.6,-3.406 7.6,-7.601l0,-3.8Z"
-        />
-        <path d="M238.631,288.345c0,-3.12 -2.533,-5.652 -5.652,-5.652l-65.958,0c-3.119,0 -5.652,2.532 -5.652,5.652l0,2.825c0,3.119 2.533,5.652 5.652,5.652l65.958,0c3.119,0 5.652,-2.533 5.652,-5.652l0,-2.825Z"
-        />
-    </g>
-    <path d="M199.486,97.833l-13.692,41.077l-41.077,0l34.231,27.385l-13.693,41.077l34.231,-27.384l34.231,27.384l-13.692,-41.077l34.231,-27.385l-41.077,0l-13.693,-41.077Z"
-    fill="#fff" fillRule="nonzero" />
-</svg>;
-icons.advancedbtn = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fill='#0058b0' fillRule='nonzero'>
-        <path d='M42.82,13.431l-37.64,0c-2.742,0.003 -4.998,2.259 -5.001,5.001l0,11.136c0.003,2.742 2.259,4.998 5.001,5.001l37.64,0c2.742,-0.003 4.998,-2.259 5.001,-5.001l0,-11.136c-0.003,-2.742 -2.259,-4.998 -5.001,-5.001Zm3.204,5.001l0,8.687c-0.003,1.757 -1.448,3.201 -3.204,3.204l-37.64,0c-1.756,-0.003 -3.201,-1.447 -3.204,-3.204l0,-8.687c0.003,-1.757 1.448,-3.201 3.204,-3.204l37.64,0c1.756,0.003 3.201,1.447 3.204,3.204Z'
-        />
-        <path d='M40.471,26.364l-32.942,0c-0.493,0 -0.898,0.406 -0.898,0.899c0,0.493 0.405,0.898 0.898,0.898l32.942,0c0.493,0 0.898,-0.405 0.898,-0.898c0,-0.493 -0.405,-0.899 -0.898,-0.899Z'
-        />
-    </g>
-</svg>;
-icons.opacity = <svg width="20px" height="20px" viewBox='0 0 20 20' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fillRule='nonzero'>
-        <path d='M14.043,7.612c2.23,2.178 2.23,5.726 0,7.904c-1.081,1.055 -2.515,1.638 -4.042,1.638c-1.527,0 -2.964,-0.583 -4.042,-1.638c-1.08,-1.056 -1.674,-2.46 -1.674,-3.952c0,-1.492 0.594,-2.896 1.676,-3.956l4.04,-4.026l4.038,4.026c0.002,0.002 0.004,0.002 0.004,0.004Z'
-        fill='url(#_Linear1)' />
-        <path d='M15.595,6.066c3.082,3.013 3.082,7.917 -0.002,10.931c-1.493,1.461 -3.479,2.265 -5.591,2.265c-2.112,0 -4.097,-0.804 -5.593,-2.265c-1.493,-1.46 -2.315,-3.402 -2.315,-5.467c0,-2.065 0.822,-4.007 2.315,-5.466l5.176,-5.162c0.221,-0.219 0.614,-0.219 0.833,0l5.177,5.164Zm-0.826,10.124c2.629,-2.568 2.629,-6.751 0,-9.32c0,-0.002 -0.003,-0.002 -0.005,-0.004l-4.762,-4.749l-4.764,4.749c-1.276,1.249 -1.976,2.905 -1.976,4.664c0,1.76 0.7,3.416 1.974,4.66c1.272,1.245 2.966,1.932 4.766,1.932c1.801,0 3.493,-0.687 4.767,-1.932Z'
-        />
-    </g>
-    <defs>
-        <linearGradient id='_Linear1' x2='1' gradientUnits='userSpaceOnUse' gradientTransform='matrix(11.4305 0 0 13.5721 4.285 10.368)'>
-            <stop offset='0' />
-            <stop offset='0.5' stopOpacity='0.749' />
-            <stop offset='1' stopOpacity='0' />
-        </linearGradient>
-    </defs>
-</svg>;
-icons.iconlistBlock = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <path d='M5.945,18.719c2.916,0 5.284,2.368 5.284,5.284c0,2.916 -2.368,5.283 -5.284,5.283c-2.916,0 -5.283,-2.367 -5.283,-5.283c0,-2.916 2.367,-5.284 5.283,-5.284Zm-1.157,8.089c0.133,0.136 0.352,0.138 0.488,0.005l4.068,-3.986c0.137,-0.134 0.139,-0.352 0.006,-0.488l-0.574,-0.595c-0.133,-0.137 -0.351,-0.139 -0.488,-0.005l-3.271,3.203l-1.495,-1.527c-0.134,-0.136 -0.352,-0.138 -0.488,-0.005l-0.551,0.541c-0.136,0.133 -0.139,0.352 -0.005,0.488l2.31,2.369Z'
-    fill='#0058b0' />
-    <path d='M5.895,2.016c2.916,0 5.283,2.367 5.283,5.283c0,2.916 -2.367,5.284 -5.283,5.284c-2.916,0 -5.283,-2.368 -5.283,-5.284c0,-2.916 2.367,-5.283 5.283,-5.283Zm-1.157,8.088c0.133,0.136 0.352,0.139 0.488,0.005l4.068,-3.986c0.136,-0.133 0.139,-0.352 0.005,-0.488l-0.573,-0.595c-0.133,-0.136 -0.352,-0.138 -0.488,-0.005l-3.271,3.204l-1.495,-1.527c-0.134,-0.136 -0.352,-0.139 -0.489,-0.005l-0.55,0.54c-0.137,0.134 -0.139,0.352 -0.006,0.489l2.311,2.368Z'
-    fill='#0058b0' />
-    <path d='M5.895,35.427c2.916,0 5.283,2.367 5.283,5.283c0,2.916 -2.367,5.284 -5.283,5.284c-2.916,0 -5.283,-2.368 -5.283,-5.284c0,-2.916 2.367,-5.283 5.283,-5.283Zm-1.157,8.088c0.133,0.136 0.352,0.139 0.488,0.005l4.068,-3.986c0.136,-0.133 0.139,-0.351 0.005,-0.488l-0.573,-0.595c-0.133,-0.136 -0.352,-0.138 -0.488,-0.005l-3.271,3.204l-1.495,-1.527c-0.134,-0.136 -0.352,-0.139 -0.489,-0.005l-0.55,0.54c-0.137,0.134 -0.139,0.352 -0.006,0.489l2.311,2.368Z'
-    fill='#0058b0' />
-    <g fill='#0058b0'>
-        <path d='M47.451,6.488c0,-0.611 -0.496,-1.107 -1.107,-1.107l-30.283,0c-0.611,0 -1.107,0.496 -1.107,1.107l0,2.214c0,0.611 0.496,1.107 1.107,1.107l30.283,0c0.611,0 1.107,-0.496 1.107,-1.107l0,-2.214Z'
-        />
-        <path d='M47.451,23.074c0,-0.607 -0.493,-1.1 -1.1,-1.1l-30.297,0c-0.607,0 -1.1,0.493 -1.1,1.1l0,2.2c0,0.607 0.493,1.1 1.1,1.1l30.297,0c0.607,0 1.1,-0.493 1.1,-1.1l0,-2.2Z'
-        />
-        <path d='M47.451,39.902c0,-0.607 -0.493,-1.1 -1.1,-1.1l-30.297,0c-0.607,0 -1.1,0.493 -1.1,1.1l0,2.2c0,0.607 0.493,1.1 1.1,1.1l30.297,0c0.607,0 1.1,-0.493 1.1,-1.1l0,-2.2Z'
-        />
-    </g>
-</svg>;
-icons.testimonialBlock = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <path d='M46.053,6.607l0.053,0.004l0.052,0.007l0.051,0.009l0.05,0.011l0.049,0.014l0.048,0.017l0.047,0.018l0.046,0.021l0.045,0.023l0.043,0.025l0.042,0.027l0.041,0.029l0.039,0.031l0.038,0.032l0.036,0.034l0.034,0.036l0.033,0.038l0.031,0.039l0.028,0.041l0.027,0.042l0.025,0.043l0.023,0.045l0.021,0.046l0.019,0.047l0.016,0.048l0.014,0.049l0.011,0.05l0.01,0.051l0.006,0.052l0.004,0.053l0.002,0.053l0,26.999l-0.002,0.053l-0.004,0.053l-0.006,0.051l-0.01,0.051l-0.011,0.05l-0.014,0.05l-0.016,0.048l-0.019,0.047l-0.021,0.046l-0.023,0.044l-0.025,0.044l-0.027,0.042l-0.028,0.041l-0.031,0.039l-0.033,0.037l-0.034,0.036l-0.036,0.035l-0.038,0.032l-0.039,0.031l-0.041,0.029l-0.042,0.027l-0.043,0.025l-0.045,0.023l-0.046,0.021l-0.047,0.018l-0.048,0.016l-0.049,0.014l-0.05,0.012l-0.051,0.009l-0.052,0.007l-0.053,0.004l-0.053,0.001l-15.751,0l-6.366,5.717l-6.397,-5.717l-15.486,0l-0.053,-0.001l-0.053,-0.004l-0.052,-0.007l-0.051,-0.009l-0.05,-0.012l-0.049,-0.014l-0.048,-0.016l-0.047,-0.018l-0.046,-0.021l-0.045,-0.023l-0.043,-0.025l-0.042,-0.027l-0.041,-0.029l-0.039,-0.031l-0.038,-0.032l-0.036,-0.035l-0.034,-0.036l-0.033,-0.037l-0.031,-0.039l-0.028,-0.041l-0.027,-0.042l-0.025,-0.044l-0.023,-0.044l-0.021,-0.046l-0.019,-0.047l-0.016,-0.048l-0.014,-0.05l-0.011,-0.05l-0.01,-0.051l-0.006,-0.051l-0.004,-0.053l-0.002,-0.053l0,-26.999l0.002,-0.053l0.004,-0.053l0.006,-0.052l0.01,-0.051l0.011,-0.05l0.014,-0.049l0.016,-0.048l0.019,-0.047l0.021,-0.046l0.023,-0.045l0.025,-0.043l0.027,-0.042l0.028,-0.041l0.031,-0.039l0.033,-0.038l0.034,-0.036l0.036,-0.034l0.038,-0.032l0.039,-0.031l0.041,-0.029l0.042,-0.027l0.043,-0.025l0.045,-0.023l0.046,-0.021l0.047,-0.018l0.048,-0.017l0.049,-0.014l0.05,-0.011l0.051,-0.009l0.052,-0.007l0.053,-0.004l0.053,-0.001l44,0l0.053,0.001Zm-43.09,1.999l0,25.071l15.287,0l5.63,5.032l5.603,-5.032l15.554,0l0,-25.071l-42.074,0Z'
-    fill='#0058b0' />
-    <path d='M5.395,19.137l2.412,-0.349c0.257,-0.042 0.477,-0.205 0.592,-0.438l1.078,-2.184c0.134,-0.272 0.413,-0.445 0.716,-0.445c0.304,0 0.582,0.173 0.716,0.445l1.078,2.184c0.115,0.233 0.336,0.396 0.592,0.438l2.412,0.349c0.391,0.058 0.684,0.396 0.684,0.791c0,0.215 -0.086,0.421 -0.24,0.571l-1.746,1.702c-0.187,0.184 -0.272,0.449 -0.228,0.708l0.411,2.403c0.008,0.044 0.012,0.089 0.012,0.134c0,0.438 -0.361,0.799 -0.8,0.799c-0.129,0 -0.257,-0.032 -0.372,-0.092l-2.155,-1.134c-0.233,-0.121 -0.512,-0.121 -0.746,0l-2.155,1.134c-0.114,0.06 -0.242,0.092 -0.372,0.092c-0.438,0 -0.799,-0.361 -0.799,-0.799c0,-0.045 0.004,-0.09 0.011,-0.134l0.411,-2.403c0.044,-0.259 -0.041,-0.524 -0.227,-0.708l-1.747,-1.702c-0.152,-0.15 -0.237,-0.355 -0.237,-0.569c0,-0.4 0.301,-0.742 0.699,-0.793Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M19.21,19.137l2.412,-0.349c0.257,-0.042 0.477,-0.205 0.591,-0.438l1.078,-2.184c0.134,-0.272 0.413,-0.445 0.716,-0.445c0.304,0 0.582,0.173 0.716,0.445l1.078,2.184c0.115,0.233 0.336,0.396 0.592,0.438l2.412,0.349c0.391,0.058 0.684,0.396 0.684,0.791c0,0.215 -0.085,0.421 -0.239,0.571l-1.746,1.702c-0.188,0.184 -0.272,0.449 -0.229,0.708l0.412,2.403c0.008,0.044 0.012,0.089 0.012,0.134c0,0.438 -0.361,0.799 -0.8,0.799c-0.13,0 -0.258,-0.032 -0.372,-0.092l-2.156,-1.134c-0.233,-0.121 -0.512,-0.121 -0.745,0l-2.156,1.134c-0.114,0.06 -0.242,0.092 -0.372,0.092c-0.438,0 -0.799,-0.361 -0.799,-0.799c0,-0.045 0.005,-0.09 0.011,-0.134l0.412,-2.403c0.044,-0.259 -0.041,-0.524 -0.227,-0.708l-1.747,-1.702c-0.152,-0.15 -0.237,-0.355 -0.237,-0.569c0,-0.4 0.3,-0.742 0.699,-0.793Z'
-    fill='#0058b0' fillRule='nonzero' />
-    <path d='M33.067,19.137l2.412,-0.349c0.257,-0.042 0.477,-0.205 0.592,-0.438l1.078,-2.184c0.134,-0.272 0.413,-0.445 0.716,-0.445c0.304,0 0.582,0.173 0.716,0.445l1.078,2.184c0.115,0.233 0.336,0.396 0.592,0.438l2.412,0.349c0.391,0.058 0.684,0.396 0.684,0.791c0,0.215 -0.086,0.421 -0.24,0.571l-1.746,1.702c-0.187,0.184 -0.272,0.449 -0.228,0.708l0.411,2.403c0.008,0.044 0.012,0.089 0.012,0.134c0,0.438 -0.361,0.799 -0.8,0.799c-0.129,0 -0.257,-0.032 -0.372,-0.092l-2.155,-1.134c-0.233,-0.121 -0.512,-0.121 -0.746,0l-2.155,1.134c-0.114,0.06 -0.242,0.092 -0.372,0.092c-0.438,0 -0.799,-0.361 -0.799,-0.799c0,-0.045 0.004,-0.09 0.011,-0.134l0.411,-2.403c0.044,-0.259 -0.041,-0.524 -0.227,-0.708l-1.747,-1.702c-0.152,-0.15 -0.237,-0.355 -0.237,-0.569c0,-0.4 0.301,-0.742 0.699,-0.793Z'
-    fill='#0058b0' fillRule='nonzero' />
-</svg>;
-icons.testimonialBasic = <svg viewBox='0 0 400 300' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect id='Artboard1' width='400' height='300' fill='none' />
-    <path d='M393.782,50.876c0,-11.515 -9.348,-20.863 -20.863,-20.863l-345.838,0c-11.515,0 -20.863,9.348 -20.863,20.863l0,198.248c0,11.515 9.348,20.863 20.863,20.863l345.838,0c11.515,0 20.863,-9.348 20.863,-20.863l0,-198.248Z'
-    fill='#fff' />
-    <circle cx='150.035' cy='214.556' r='28.108' fill='#eee' />
-    <path d='M277.897,201.379c0,-3.085 -2.505,-5.59 -5.59,-5.59l-80.636,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l80.636,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z'
-    fill='#888' />
-    <path d='M278.073,223.98c0,-3.085 -2.505,-5.59 -5.59,-5.59l-80.636,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l80.636,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z'
-    fill='#b0b0b0' />
-    <path d='M140.132,202.697l20.939,0c0.932,0 1.689,0.757 1.689,1.689l0,20.939c0,0.932 -0.757,1.688 -1.689,1.688l-20.939,0c-0.932,0 -1.688,-0.756 -1.688,-1.688l0,-20.939c0,-0.932 0.756,-1.689 1.688,-1.689Zm19.926,21.615l0,-18.913l-18.912,0l0,18.913l18.912,0Zm-9.456,-14.86c0,-1.486 -1.216,-2.702 -2.702,-2.702c-1.486,0 -2.702,1.216 -2.702,2.702c0,1.486 1.216,2.702 2.702,2.702c1.486,0 2.702,-1.216 2.702,-2.702Zm4.053,6.754c0,0 0,-8.105 4.052,-8.105l0,13.509c0,0.743 -0.608,1.351 -1.351,1.351l-13.509,0c-0.743,0 -1.35,-0.608 -1.35,-1.351l0,-9.456c2.701,0 4.052,5.403 4.052,5.403c0,0 1.351,-4.052 4.053,-4.052c2.701,0 4.053,2.701 4.053,2.701Z'
-    fill='#888' fillRule='nonzero' />
-    <path d='M333.379,65.119c0,-4.187 -3.399,-7.586 -7.585,-7.586l-251.588,0c-4.186,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.399,7.586 7.585,7.586l251.588,0c4.186,0 7.585,-3.399 7.585,-7.586l0,-3.834Z'
-    fill='#444' />
-    <path d='M347.747,153.688c0,-3.137 -2.547,-5.683 -5.684,-5.683l-284.126,0c-3.137,0 -5.684,2.546 -5.684,5.683l0,2.674c0,3.137 2.547,5.684 5.684,5.684l284.126,0c3.137,0 5.684,-2.547 5.684,-5.684l0,-2.674Z'
-    fill='#888' />
-    <path d='M371.417,126.188c0,-3.136 -2.547,-5.683 -5.683,-5.683l-331.468,0c-3.136,0 -5.683,2.547 -5.683,5.683l0,2.675c0,3.136 2.547,5.683 5.683,5.683l331.468,0c3.136,0 5.683,-2.547 5.683,-5.683l0,-2.675Z'
-    fill='#888' />
-    <path d='M371.417,98.54c0,-3.137 -2.547,-5.684 -5.683,-5.684l-331.468,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l331.468,0c3.136,0 5.683,-2.546 5.683,-5.683l0,-2.674Z'
-    fill='#888' />
-</svg>;
-icons.testimonialCard = <svg viewBox='0 0 400 300' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect id='Artboard1' width='400' height='300' fill='none' />
-    <rect x='38.715' y='5' width='324.316' height='290' fill='#fff' />
-    <rect x='50.019' y='5.616' width='299.963' height='111.996' fill='#eee'
-    />
-    <circle cx='189.601' cy='36.677' r='28.108' fill='#eee' />
-    <path d='M240.707,256.414c0,-2.676 -2.173,-4.85 -4.849,-4.85l-69.953,0c-2.676,0 -4.849,2.174 -4.849,4.85l0,2.44c0,2.676 2.173,4.85 4.849,4.85l69.953,0c2.676,0 4.849,-2.174 4.849,-4.85l0,-2.44Z'
-    fill='#888' />
-    <path d='M240.86,276.021c0,-2.676 -2.173,-4.85 -4.85,-4.85l-69.952,0c-2.677,0 -4.85,2.174 -4.85,4.85l0,2.44c0,2.676 2.173,4.85 4.85,4.85l69.952,0c2.677,0 4.85,-2.174 4.85,-4.85l0,-2.44Z'
-    fill='#b0b0b0' />
-    <path d='M176.538,43.322l39.493,0c1.758,0 3.185,1.428 3.185,3.186l0,39.492c0,1.758 -1.427,3.184 -3.185,3.184l-39.493,0c-1.758,0 -3.184,-1.426 -3.184,-3.184l0,-39.492c0,-1.758 1.426,-3.186 3.184,-3.186Zm37.582,40.766l0,-35.67l-35.67,0l0,35.67l35.67,0Zm-17.835,-28.026c0,-2.803 -2.293,-5.096 -5.096,-5.096c-2.802,0 -5.095,2.293 -5.095,5.096c0,2.802 2.293,5.096 5.095,5.096c2.803,0 5.096,-2.294 5.096,-5.096Zm7.644,12.74c0,0 0,-15.288 7.643,-15.288l0,25.479c0,1.402 -1.147,2.548 -2.548,2.548l-25.479,0c-1.401,0 -2.547,-1.146 -2.547,-2.548l0,-17.835c5.096,0 7.643,10.191 7.643,10.191c0,0 2.548,-7.643 7.644,-7.643c5.096,0 7.644,5.096 7.644,5.096Z'
-    fill='#888' fillRule='nonzero' />
-    <path d='M317.212,141.565c0,-4.187 -3.399,-7.586 -7.585,-7.586l-219.254,0c-4.186,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.399,7.586 7.585,7.586l219.254,0c4.186,0 7.585,-3.399 7.585,-7.586l0,-3.834Z'
-    fill='#444' />
-    <path d='M291.313,223.205c0,-2.939 -2.387,-5.324 -5.326,-5.324l-170.966,0c-2.939,0 -5.325,2.385 -5.325,5.324l0,2.506c0,2.939 2.386,5.325 5.325,5.325l170.966,0c2.939,0 5.326,-2.386 5.326,-5.325l0,-2.506Z'
-    fill='#888' />
-    <path d='M313.488,197.441c0,-2.938 -2.386,-5.324 -5.324,-5.324l-215.319,0c-2.938,0 -5.324,2.386 -5.324,5.324l0,2.506c0,2.938 2.386,5.325 5.324,5.325l215.319,0c2.938,0 5.324,-2.387 5.324,-5.325l0,-2.506Z'
-    fill='#888' />
-    <path d='M313.488,171.538c0,-2.939 -2.386,-5.325 -5.324,-5.325l-215.319,0c-2.938,0 -5.324,2.386 -5.324,5.325l0,2.505c0,2.939 2.386,5.325 5.324,5.325l215.319,0c2.938,0 5.324,-2.386 5.324,-5.325l0,-2.505Z'
-    fill='#888' />
-</svg>;
-icons.testimonialBubble = <svg viewBox='0 0 400 300' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect id='Artboard1' width='400' height='300' fill='none' />
-    <path d='M393.782,50.876c0,-11.515 -9.348,-20.863 -20.863,-20.863l-345.838,0c-11.515,0 -20.863,9.348 -20.863,20.863l0,114.463c0,11.515 9.348,20.863 20.863,20.863l345.838,0c11.515,0 20.863,-9.348 20.863,-20.863l0,-114.463Z'
-    fill='#fff' stroke='#888' strokeWidth='4' />
-    <circle cx='150.035' cy='240.744' r='28.108' fill='#eee' />
-    <path d='M277.897,227.567c0,-3.085 -2.505,-5.59 -5.59,-5.59l-80.636,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l80.636,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z'
-    fill='#888' />
-    <path d='M278.073,250.168c0,-3.085 -2.505,-5.59 -5.59,-5.59l-80.636,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l80.636,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z'
-    fill='#b0b0b0' />
-    <path d='M140.132,228.885l20.939,0c0.932,0 1.689,0.757 1.689,1.689l0,20.939c0,0.932 -0.757,1.688 -1.689,1.688l-20.939,0c-0.932,0 -1.688,-0.756 -1.688,-1.688l0,-20.939c0,-0.932 0.756,-1.689 1.688,-1.689Zm19.926,21.615l0,-18.913l-18.912,0l0,18.913l18.912,0Zm-9.456,-14.86c0,-1.486 -1.216,-2.702 -2.702,-2.702c-1.486,0 -2.702,1.216 -2.702,2.702c0,1.486 1.216,2.702 2.702,2.702c1.486,0 2.702,-1.216 2.702,-2.702Zm4.053,6.754c0,0 0,-8.105 4.052,-8.105l0,13.509c0,0.743 -0.608,1.351 -1.351,1.351l-13.509,0c-0.743,0 -1.35,-0.608 -1.35,-1.351l0,-9.456c2.701,0 4.052,5.403 4.052,5.403c0,0 1.351,-4.052 4.053,-4.052c2.701,0 4.053,2.701 4.053,2.701Z'
-    fill='#888' fillRule='nonzero' />
-    <path d='M333.379,63.119c0,-4.187 -3.399,-7.586 -7.585,-7.586l-251.588,0c-4.186,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.399,7.586 7.585,7.586l251.588,0c4.186,0 7.585,-3.399 7.585,-7.586l0,-3.834Z'
-    fill='#444' />
-    <path d='M347.747,151.688c0,-3.137 -2.547,-5.683 -5.684,-5.683l-284.126,0c-3.137,0 -5.684,2.546 -5.684,5.683l0,2.674c0,3.137 2.547,5.684 5.684,5.684l284.126,0c3.137,0 5.684,-2.547 5.684,-5.684l0,-2.674Z'
-    fill='#888' />
-    <path d='M371.417,124.188c0,-3.136 -2.547,-5.683 -5.683,-5.683l-331.468,0c-3.136,0 -5.683,2.547 -5.683,5.683l0,2.675c0,3.136 2.547,5.683 5.683,5.683l331.468,0c3.136,0 5.683,-2.547 5.683,-5.683l0,-2.675Z'
-    fill='#888' />
-    <path d='M371.417,96.54c0,-3.137 -2.547,-5.684 -5.683,-5.684l-331.468,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l331.468,0c3.136,0 5.683,-2.546 5.683,-5.683l0,-2.674Z'
-    fill='#888' />
-    <path d='M200,202.637l16.507,-14.995l-33.014,0l16.507,14.995Z' fill='#888'
-    />
-</svg>;
-icons.testimonialInline = <svg viewBox='0 0 400 300' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect id='Artboard1' width='400' height='300' fill='none' />
-    <path d='M393.782,80.651c0,-11.515 -9.348,-20.863 -20.863,-20.863l-345.838,0c-11.515,0 -20.863,9.348 -20.863,20.863l0,114.463c0,11.515 9.348,20.863 20.863,20.863l345.838,0c11.515,0 20.863,-9.348 20.863,-20.863l0,-114.463Z'
-    fill='#fff' stroke='#888' strokeWidth='4' />
-    <circle cx='58.029' cy='105.034' r='28.108' fill='#eee' />
-    <path d='M323.808,231.809c0,-3.085 -2.505,-5.59 -5.59,-5.59l-111.083,0c-3.085,0 -5.59,2.505 -5.59,5.59l0,2.813c0,3.085 2.505,5.59 5.59,5.59l111.083,0c3.085,0 5.59,-2.505 5.59,-5.59l0,-2.813Z'
-    fill='#888' />
-    <path d='M48.126,93.175l20.939,0c0.932,0 1.689,0.757 1.689,1.689l0,20.939c0,0.932 -0.757,1.688 -1.689,1.688l-20.939,0c-0.932,0 -1.688,-0.756 -1.688,-1.688l0,-20.939c0,-0.932 0.756,-1.689 1.688,-1.689Zm19.926,21.615l0,-18.913l-18.912,0l0,18.913l18.912,0Zm-9.456,-14.86c0,-1.486 -1.216,-2.702 -2.702,-2.702c-1.486,0 -2.702,1.216 -2.702,2.702c0,1.486 1.216,2.702 2.702,2.702c1.486,0 2.702,-1.216 2.702,-2.702Zm4.053,6.754c0,0 0,-8.105 4.052,-8.105l0,13.509c0,0.743 -0.608,1.351 -1.351,1.351l-13.509,0c-0.743,0 -1.35,-0.608 -1.35,-1.351l0,-9.456c2.701,0 4.052,5.403 4.052,5.403c0,0 1.351,-4.052 4.053,-4.052c2.701,0 4.053,2.701 4.053,2.701Z'
-    fill='#888' fillRule='nonzero' />
-    <path d='M360.415,93.146c0,-4.187 -3.399,-7.586 -7.585,-7.586l-241.001,0c-4.186,0 -7.585,3.399 -7.585,7.586l0,3.834c0,4.187 3.399,7.586 7.585,7.586l241.001,0c4.186,0 7.585,-3.399 7.585,-7.586l0,-3.834Z'
-    fill='#444' />
-    <path d='M323.405,181.463c0,-3.137 -2.547,-5.683 -5.684,-5.683l-284.126,0c-3.137,0 -5.684,2.546 -5.684,5.683l0,2.674c0,3.137 2.547,5.684 5.684,5.684l284.126,0c3.137,0 5.684,-2.547 5.684,-5.684l0,-2.674Z'
-    fill='#888' />
-    <path d='M371.417,153.963c0,-3.136 -2.547,-5.683 -5.683,-5.683l-331.468,0c-3.136,0 -5.683,2.547 -5.683,5.683l0,2.675c0,3.136 2.547,5.683 5.683,5.683l331.468,0c3.136,0 5.683,-2.547 5.683,-5.683l0,-2.675Z'
-    fill='#888' />
-    <path d='M371.417,126.315c0,-3.137 -2.547,-5.684 -5.683,-5.684l-255.823,0c-3.136,0 -5.683,2.547 -5.683,5.684l0,2.674c0,3.137 2.547,5.683 5.683,5.683l255.823,0c3.136,0 5.683,-2.546 5.683,-5.683l0,-2.674Z'
-    fill='#888' />
-    <path d='M348.967,232.413l16.507,-14.996l-33.014,0l16.507,14.996Z' fill='#888'
-    />
-</svg>;
-icons.accord01 = <svg viewBox='0 0 400 240' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='4' y='30.098' width='390' height='120.406' fill='#fff' />
-    <path d='M394,150.504l-390,0l0,-120.406l390,0l0,120.406Zm-389,-119.406l0,118.406l388,0l0,-118.406l-388,0Z'
-    fill='#eee' />
-    <rect x='4' y='30.125' width='390' height='30' fill='#444' />
-    <rect x='4' y='149.759' width='390' height='30' fill='#eee' />
-    <rect x='4' y='179.902' width='390' height='30' fill='#eee' />
-    <rect x='366.177' y='162.759' width='16' height='4' fill='#444' />
-    <rect x='366.177' y='43.125' width='16' height='4' fill='#fff' />
-    <rect x='366.177' y='192.902' width='16' height='4' fill='#444' />
-    <path d='M376.177,172.759l0,-16l-4,0l0,16l4,0Z' fill='#444' />
-    <path d='M376.177,202.902l0,-16l-4,0l0,16l4,0Z' fill='#444' />
-    <path d='M190.685,42.625c0,-1.38 -1.12,-2.5 -2.5,-2.5l-170,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l170,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#fff' fillOpacity='0.349' />
-    <path d='M240.685,162.259c0,-1.38 -1.12,-2.5 -2.5,-2.5l-220,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.379 1.121,2.5 2.5,2.5l220,0c1.38,0 2.5,-1.121 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M160.685,192.402c0,-1.379 -1.12,-2.5 -2.5,-2.5l-140,0c-1.379,0 -2.5,1.121 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l140,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.349' />
-    <path d='M375.067,99.771c0,-0.827 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.673 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M375.067,114.124c0,-0.828 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M320.956,128.477c0,-0.828 -0.672,-1.5 -1.5,-1.5l-302.314,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l302.314,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M219.025,79.795c0,-1.104 -0.896,-2 -2,-2l-199.383,0c-1.103,0 -2,0.896 -2,2l0,4c0,1.104 0.897,2 2,2l199.383,0c1.104,0 2,-0.896 2,-2l0,-4Z'
-    fill='#444' fillOpacity='0.353' />
-</svg>;
-icons.accord02 = <svg viewBox='0 0 400 240' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='5' y='13.744' width='390' height='212.512' fill='#fff' />
-    <path d='M395,23.771c0,-5.519 -4.481,-10 -10,-10l-370,0c-5.519,0 -10,4.481 -10,10l0,20c0,5.519 4.481,10 10,10l370,0c5.519,0 10,-4.481 10,-10l0,-20Z'
-    fill='#f3690e' />
-    <path d='M395,153.164c0,-5.519 -4.481,-10 -10,-10l-370,0c-5.519,0 -10,4.481 -10,10l0,20c0,5.519 4.481,10 10,10l370,0c5.519,0 10,-4.481 10,-10l0,-20Z'
-    fill='#eee' />
-    <path d='M395,195.992c0,-5.519 -4.481,-10 -10,-10l-370,0c-5.519,0 -10,4.481 -10,10l0,20c0,5.519 4.481,10 10,10l370,0c5.519,0 10,-4.481 10,-10l0,-20Z'
-    fill='#eee' />
-    <rect x='367.177' y='161.163' width='16' height='4' fill='#444' />
-    <rect x='367.177' y='31.715' width='16' height='4' fill='#fff' />
-    <rect x='367.177' y='204.295' width='16' height='4' fill='#444' />
-    <path d='M377.177,171.163l0,-16l-4,0l0,16l4,0Z' fill='#444' />
-    <path d='M377.177,214.295l0,-16l-4,0l0,16l4,0Z' fill='#444' />
-    <path d='M191.685,31.271c0,-1.379 -1.12,-2.5 -2.5,-2.5l-170,0c-1.379,0 -2.5,1.121 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l170,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#fff' fillOpacity='0.349' />
-    <path d='M241.685,160.664c0,-1.38 -1.12,-2.5 -2.5,-2.5l-220,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l220,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M161.685,203.492c0,-1.38 -1.12,-2.5 -2.5,-2.5l-140,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l140,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.349' />
-    <path d='M375.067,91.903c0,-0.827 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.673 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M375.067,106.256c0,-0.828 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M320.956,120.609c0,-0.828 -0.672,-1.5 -1.5,-1.5l-302.314,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l302.314,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M219.025,71.927c0,-1.104 -0.896,-2 -2,-2l-199.383,0c-1.103,0 -2,0.896 -2,2l0,4c0,1.104 0.897,2 2,2l199.383,0c1.104,0 2,-0.896 2,-2l0,-4Z'
-    fill='#444' fillOpacity='0.353' />
-</svg>;
-icons.accord03 = <svg viewBox='0 0 400 240' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='5.009' y='9.637' width='390' height='131.134' fill='#fff' />
-    <path d='M395.009,140.771l-390,0l0,-131.134l390,0l0,131.134Zm-389,-130.134l0,129.134l388,0l0,-129.134l-388,0Z'
-    fill='#eee' />
-    <path d='M375.076,89.387c0,-0.827 -0.672,-1.5 -1.5,-1.5l-356.424,0c-0.828,0 -1.5,0.673 -1.5,1.5l0,3c0,0.828 0.672,1.5 1.5,1.5l356.424,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M375.076,103.74c0,-0.828 -0.672,-1.5 -1.5,-1.5l-356.424,0c-0.828,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.672,1.5 1.5,1.5l356.424,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M320.966,118.093c0,-0.828 -0.673,-1.5 -1.5,-1.5l-302.314,0c-0.828,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.672,1.5 1.5,1.5l302.314,0c0.827,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M219.035,69.411c0,-1.104 -0.897,-2 -2,-2l-199.383,0c-1.104,0 -2,0.896 -2,2l0,4c0,1.104 0.896,2 2,2l199.383,0c1.103,0 2,-0.896 2,-2l0,-4Z'
-    fill='#444' fillOpacity='0.353' />
-    <rect x='5.009' y='9.664' width='390' height='40' fill='#fff' />
-    <path d='M395.009,49.664l-390,0l0,-40l390,0l0,40Zm-389,-39l0,38l388,0l0,-38l-388,0Z'
-    fill='#eee' />
-    <rect x='5.009' y='145.401' width='390' height='40' fill='#fff' />
-    <path d='M395.009,185.401l-390,0l0,-40l390,0l0,40Zm-389,-39l0,38l388,0l0,-38l-388,0Z'
-    fill='#eee' />
-    <rect x='5.961' y='145.419' width='1.018' height='40.003' fill='#eee'
-    />
-    <rect x='4.991' y='9.615' width='2.022' height='40.003' fill='#0e9cd1'
-    />
-    <rect x='5.009' y='190.384' width='390' height='40' fill='#fff' stroke='#eee'
-    />
-    <rect x='5.429' y='190.382' width='1.018' height='40.003' fill='#eee'
-    />
-    <g fill='#444'>
-        <path d='M376.465,168.229l-7.071,-7.071l-1.414,1.415l7.071,7.071l1.414,-1.415Z'
-        />
-        <path d='M375.321,169.644l7.071,-7.071l-1.414,-1.415l-7.071,7.071l1.414,1.415Z'
-        />
-    </g>
-    <g fill='#444'>
-        <path d='M376.465,212.916l-7.071,-7.071l-1.414,1.414l7.071,7.071l1.414,-1.414Z'
-        />
-        <path d='M375.321,214.33l7.071,-7.071l-1.414,-1.414l-7.071,7.071l1.414,1.414Z'
-        />
-    </g>
-    <g fill='#444'>
-        <path d='M373.907,26.836l7.071,7.071l1.414,-1.414l-7.071,-7.071l-1.414,1.414Z'
-        />
-        <path d='M375.051,25.422l-7.071,7.071l1.414,1.414l7.071,-7.071l-1.414,-1.414Z'
-        />
-    </g>
-    <path d='M191.695,27.164c0,-1.379 -1.12,-2.5 -2.5,-2.5l-170,0c-1.38,0 -2.5,1.121 -2.5,2.5l0,5c0,1.38 1.12,2.5 2.5,2.5l170,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.553' />
-    <path d='M241.695,162.901c0,-1.38 -1.12,-2.5 -2.5,-2.5l-220,0c-1.38,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.12,2.5 2.5,2.5l220,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M161.695,207.587c0,-1.379 -1.12,-2.5 -2.5,-2.5l-140,0c-1.38,0 -2.5,1.121 -2.5,2.5l0,5c0,1.38 1.12,2.5 2.5,2.5l140,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.349' />
-</svg>;
-icons.accord04 = <svg viewBox='0 0 400 240' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <rect x='4.957' y='11.561' width='390' height='132.583' fill='#fff' />
-    <rect x='4.957' y='11.588' width='390' height='40' fill='#fff' />
-    <rect x='5.043' y='148.355' width='390' height='40' fill='#fff' />
-    <rect x='4.957' y='188.439' width='390' height='40' fill='#fff' />
-    <rect x='4.957' y='224.446' width='390' height='3.993' fill='#eee' />
-    <rect x='4.957' y='184.362' width='390' height='3.993' fill='#eee' />
-    <rect x='4.957' y='47.595' width='390' height='3.993' fill='#0e9cd1' />
-    <path d='M191.642,29.088c0,-1.38 -1.12,-2.5 -2.5,-2.5l-170,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l170,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.651' />
-    <path d='M241.642,165.597c0,-1.379 -1.12,-2.5 -2.5,-2.5l-220,0c-1.379,0 -2.5,1.121 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l220,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M376.067,92.996c0,-0.828 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M376.067,107.349c0,-0.828 -0.672,-1.5 -1.5,-1.5l-356.425,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.828 0.673,1.5 1.5,1.5l356.425,0c0.828,0 1.5,-0.672 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M321.956,121.702c0,-0.828 -0.672,-1.5 -1.5,-1.5l-302.314,0c-0.827,0 -1.5,0.672 -1.5,1.5l0,3c0,0.827 0.673,1.5 1.5,1.5l302.314,0c0.828,0 1.5,-0.673 1.5,-1.5l0,-3Z'
-    fill='#444' fillOpacity='0.353' />
-    <path d='M161.642,205.643c0,-1.38 -1.12,-2.5 -2.5,-2.5l-140,0c-1.379,0 -2.5,1.12 -2.5,2.5l0,5c0,1.38 1.121,2.5 2.5,2.5l140,0c1.38,0 2.5,-1.12 2.5,-2.5l0,-5Z'
-    fill='#444' fillOpacity='0.349' />
-    <path d='M220.025,73.019c0,-1.104 -0.896,-2 -2,-2l-199.383,0c-1.103,0 -2,0.896 -2,2l0,4c0,1.104 0.897,2 2,2l199.383,0c1.104,0 2,-0.896 2,-2l0,-4Z'
-    fill='#444' fillOpacity='0.353' />
-</svg>;
-icons.headingBlock = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='1.414'>
-    <g fill='#0058b0' fillRule='nonzero'>
-        <path d='M45.013,46.298c0.705,0 1.285,-0.58 1.285,-1.285l0,-42.026c0,-0.705 -0.58,-1.285 -1.285,-1.285l-42.026,0c-0.705,0 -1.285,0.58 -1.285,1.285l0,42.026c0,0.705 0.58,1.285 1.285,1.285l42.026,0Zm-40.741,-42.026l39.456,0l0,39.456l-39.456,0l0,-39.456Z'
-        />
-        <path d='M11.221,8.736l7.36,0l0,12.117l10.838,0l0,-12.117l7.361,0l0,30.518l-7.361,0l0,-12.743l-10.838,0l0,12.743l-7.36,0l0,-30.518Z'
-        />
-    </g>
-</svg>;
-icons.gallery = <svg width="20px" height="20px" viewBox='0 0 48 48' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <g fill='#0058b0' fillRule='nonzero'>
-        <path d='M46.652,11.743l-38.504,0c-0.414,0 -0.755,0.341 -0.755,0.755l0,28.023c0,0.414 0.341,0.755 0.755,0.755l38.504,0c0.414,0 0.755,-0.341 0.755,-0.755l0,-28.029c-0.003,-0.412 -0.344,-0.749 -0.755,-0.749Zm-0.754,28.017l-36.995,0l0,-26.508l36.995,0l0,26.508Z'
-        />
-        <path d='M1.53,5.125l38.504,0c0.412,0 0.752,0.337 0.755,0.749l0,4.421l-1.509,0l0,-3.66l-36.995,0l0,26.507l3.772,0l0,1.516l-4.527,0c-0.414,0 -0.755,-0.341 -0.755,-0.755l0,-28.023c0,-0.414 0.341,-0.755 0.755,-0.755Z'
-        />
-        <path d='M35,37.002l8.603,0l-9.256,-11.785l-4.302,5.476l-6.946,-8.846l-11.902,15.155l23.803,0Z'
-        />
-        <path d='M37.607,24.719c2.402,-0.003 4.378,-1.98 4.378,-4.382c0,-2.403 -1.978,-4.381 -4.382,-4.381c-2.403,0 -4.381,1.978 -4.381,4.381c0.003,2.404 1.981,4.381 4.385,4.382Zm0,-7.254c1.574,0.003 2.868,1.299 2.868,2.872c0,1.576 -1.297,2.872 -2.872,2.872c-1.576,0 -2.872,-1.296 -2.872,-2.872c0.001,-1.575 1.299,-2.872 2.874,-2.872l0.002,0Z'
-        />
-    </g>
-</svg>;
-icons.galCarousel = <svg viewBox='0 0 48 37' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <rect id='Artboard1' width='47.974' height='36.014' fill='none' />
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M22.456,10.263l-20.191,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,14.696c0,0.217 0.179,0.396 0.396,0.396l20.191,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-14.699c-0.002,-0.216 -0.18,-0.393 -0.396,-0.393Zm-0.395,14.692l-19.401,0l0,-13.9l19.401,0l0,13.9Z'
-        />
-        <path d='M16.346,23.509l4.512,0l-4.854,-6.18l-2.256,2.872l-3.643,-4.639l-6.242,7.947l12.483,0Z'
-        />
-        <path d='M17.713,17.068c1.26,-0.001 2.296,-1.038 2.296,-2.298c0,-1.26 -1.038,-2.297 -2.298,-2.297c-1.26,0 -2.298,1.037 -2.298,2.297c0.002,1.261 1.039,2.298 2.3,2.298Zm0,-3.804c0.825,0.002 1.504,0.681 1.504,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.827,0 -1.506,-0.68 -1.506,-1.506c0,-0.826 0.681,-1.506 1.507,-1.506l0.001,0Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M45.661,10.263l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,14.696c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-14.699c-0.002,-0.216 -0.18,-0.393 -0.396,-0.393Zm-0.396,14.692l-19.4,0l0,-13.9l19.4,0l0,13.9Z'
-        />
-        <path d='M39.551,23.509l4.511,0l-4.854,-6.18l-2.256,2.872l-3.642,-4.639l-6.242,7.947l12.483,0Z'
-        />
-        <path d='M40.918,17.068c1.259,-0.001 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.297,-2.297c-1.261,0 -2.298,1.037 -2.298,2.297c0.002,1.261 1.039,2.298 2.3,2.298Zm0,-3.804c0.825,0.002 1.504,0.681 1.504,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.827,0 -1.507,-0.68 -1.507,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-    </g>
-    <circle cx='21.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-    <circle cx='24.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-    <circle cx='27.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-</svg>;
-icons.galFluid = <svg viewBox='0 0 48 37' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <rect id='Artboard1' width='47.974' height='36.014' fill='none' />
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M10.287,17.068c1.259,-0.001 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.298 2.3,2.298Zm0,-3.804c0.825,0.002 1.503,0.681 1.503,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-        <path d='M0.871,17.863l1.807,-2.301l3.643,4.639l2.256,-2.872l4.854,6.18l-12.56,0l0,-5.646Z'
-        />
-        <path d='M15.03,10.263c0.215,0 0.394,0.177 0.396,0.393l0,14.699c0,0.217 -0.179,0.396 -0.396,0.396l-14.159,0l0,-0.796l13.763,0l0,-13.9l-13.763,0l0,-0.792l14.159,0Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M47.053,12.922l0,1.2c-0.095,0.197 -0.149,0.417 -0.149,0.648c0,0.232 0.053,0.452 0.149,0.649l0,1.199c-0.568,-0.42 -0.94,-1.094 -0.941,-1.848c0,-0.754 0.372,-1.428 0.941,-1.848Z'
-        />
-        <path d='M47.053,23.509l-12.491,0l6.242,-7.947l3.643,4.639l2.256,-2.872l0.35,0.446l0,5.734Z'
-        />
-        <path d='M32.964,10.263l14.089,0l0,0.792l-13.694,0l0,13.9l13.694,0l0,0.796l-14.089,0c-0.217,0 -0.396,-0.179 -0.396,-0.396l0,-14.696c0,-0.217 0.179,-0.396 0.396,-0.396Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M30.712,10.263l-13.451,0c-0.217,0 -0.395,0.179 -0.395,0.396l0,14.696c0,0.217 0.178,0.396 0.395,0.396l13.451,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-14.699c-0.002,-0.216 -0.18,-0.393 -0.396,-0.393Zm-0.395,14.692l-12.66,0l0,-13.9l12.66,0l0,13.9Z'
-        />
-        <path d='M26.263,23.458l2.443,0l-2.628,-3.347l-1.222,1.555l-1.973,-2.512l-3.38,4.304l6.76,0Z'
-        />
-        <path d='M26.836,17.145c0.837,-0.001 1.527,-0.691 1.527,-1.529c0,-0.838 -0.69,-1.528 -1.529,-1.528c-0.838,0 -1.528,0.69 -1.528,1.528c0.001,0.839 0.691,1.529 1.53,1.529Zm0,-2.531c0.548,0.001 1,0.453 1,1.002c0,0.55 -0.452,1.002 -1.002,1.002c-0.55,0 -1.002,-0.452 -1.002,-1.002c0,-0.549 0.453,-1.002 1.003,-1.002l0.001,0Z'
-        />
-    </g>
-    <circle cx='21.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-    <circle cx='24.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-    <circle cx='27.053' cy='27.427' r='0.518' fill='#3a3a3a' />
-</svg>;
-icons.galGrid = <svg viewBox='0 0 48 37' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <rect id='Artboard1' width='47.974' height='36.014' fill='none' />
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M45.54,1.456l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,14.696c0,0.217 0.179,0.395 0.396,0.395l20.192,0c0.217,0 0.396,-0.178 0.396,-0.395l0,-14.699c-0.002,-0.216 -0.181,-0.393 -0.396,-0.393Zm-0.396,14.692l-19.4,0l0,-13.9l19.4,0l0,13.9Z'
-        />
-        <path d='M39.43,14.702l4.511,0l-4.854,-6.18l-2.256,2.872l-3.643,-4.639l-6.241,7.947l12.483,0Z'
-        />
-        <path d='M40.797,8.261c1.259,-0.002 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.297 2.3,2.298Zm0,-3.804c0.825,0.001 1.503,0.681 1.503,1.506c0,0.826 -0.679,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-        <path d='M22.626,1.456l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,14.696c0,0.217 0.179,0.395 0.396,0.395l20.192,0c0.217,0 0.396,-0.178 0.396,-0.395l0,-14.699c-0.002,-0.216 -0.181,-0.393 -0.396,-0.393Zm-0.396,14.692l-19.4,0l0,-13.9l19.4,0l0,13.9Z'
-        />
-        <path d='M16.515,14.702l4.512,0l-4.854,-6.18l-2.256,2.872l-3.643,-4.639l-6.241,7.947l12.482,0Z'
-        />
-        <path d='M17.883,8.261c1.259,-0.002 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.297 2.3,2.298Zm0,-3.804c0.825,0.001 1.503,0.681 1.503,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-        <g>
-            <path d='M45.54,19.071l-20.192,0c-0.217,0 -0.396,0.178 -0.396,0.395l0,14.696c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-14.699c-0.002,-0.216 -0.181,-0.392 -0.396,-0.392Zm-0.396,14.692l-19.4,0l0,-13.901l19.4,0l0,13.901Z'
-            />
-            <path d='M39.43,32.317l4.511,0l-4.854,-6.181l-2.256,2.872l-3.643,-4.639l-6.241,7.948l12.483,0Z'
-            />
-            <path d='M40.797,25.875c1.259,-0.001 2.295,-1.038 2.295,-2.297c0,-1.261 -1.037,-2.298 -2.298,-2.298c-1.26,0 -2.297,1.037 -2.297,2.298c0.002,1.26 1.039,2.297 2.3,2.297Zm0,-3.803c0.825,0.001 1.503,0.68 1.503,1.506c0,0.826 -0.679,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.827 0.681,-1.506 1.507,-1.506l0.002,0Z'
-            />
-        </g>
-        <g>
-            <path d='M22.626,19.071l-20.192,0c-0.217,0 -0.396,0.178 -0.396,0.395l0,14.696c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-14.699c-0.002,-0.216 -0.181,-0.392 -0.396,-0.392Zm-0.396,14.692l-19.4,0l0,-13.901l19.4,0l0,13.901Z'
-            />
-            <path d='M16.515,32.317l4.512,0l-4.854,-6.181l-2.256,2.872l-3.643,-4.639l-6.241,7.948l12.482,0Z'
-            />
-            <path d='M17.883,25.875c1.259,-0.001 2.295,-1.038 2.295,-2.297c0,-1.261 -1.037,-2.298 -2.298,-2.298c-1.26,0 -2.297,1.037 -2.297,2.298c0.002,1.26 1.039,2.297 2.3,2.297Zm0,-3.803c0.825,0.001 1.503,0.68 1.503,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.827 0.681,-1.506 1.507,-1.506l0.002,0Z'
-            />
-        </g>
-    </g>
-</svg>;
-icons.galMasonry = <svg viewBox='0 0 48 37' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <rect id='Artboard1' width='47.974' height='36.014' fill='none' />
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M45.54,1.456l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,11.139c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-11.142c-0.002,-0.216 -0.181,-0.393 -0.396,-0.393Zm-0.396,11.136l-19.4,0l0,-10.344l19.4,0l0,10.344Z'
-        />
-        <path d='M36.728,11.438l3.535,0l-3.803,-4.843l-1.768,2.25l-2.854,-3.635l-4.891,6.228l9.781,0Z'
-        />
-        <path d='M40.797,8.261c1.259,-0.002 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.297 2.3,2.298Zm0,-3.804c0.825,0.001 1.503,0.681 1.503,1.506c0,0.826 -0.679,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M22.626,1.456l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,21.375c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-21.378c-0.002,-0.216 -0.181,-0.393 -0.396,-0.393Zm-0.396,21.372l-19.4,0l0,-20.58l19.4,0l0,20.58Z'
-        />
-        <path d='M16.515,19.869l4.512,0l-4.854,-6.18l-2.256,2.871l-3.643,-4.639l-6.241,7.948l12.482,0Z'
-        />
-        <path d='M17.883,9.509c1.259,-0.001 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.297 2.3,2.298Zm0,-3.804c0.825,0.002 1.503,0.681 1.503,1.506c0,0.826 -0.68,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M45.54,15.418l-20.192,0c-0.217,0 -0.396,0.179 -0.396,0.396l0,18.143c0,0.217 0.179,0.396 0.396,0.396l20.192,0c0.217,0 0.396,-0.179 0.396,-0.396l0,-18.146c-0.002,-0.216 -0.181,-0.393 -0.396,-0.393Zm-0.396,18.14l-19.4,0l0,-17.348l19.4,0l0,17.348Z'
-        />
-        <path d='M39.43,31.64l4.511,0l-4.854,-6.18l-2.256,2.872l-3.643,-4.639l-6.241,7.947l12.483,0Z'
-        />
-        <path d='M40.797,23.21c1.259,-0.002 2.295,-1.039 2.295,-2.298c0,-1.26 -1.037,-2.298 -2.298,-2.298c-1.26,0 -2.297,1.038 -2.297,2.298c0.002,1.26 1.039,2.297 2.3,2.298Zm0,-3.804c0.825,0.001 1.503,0.681 1.503,1.506c0,0.826 -0.679,1.506 -1.506,1.506c-0.826,0 -1.506,-0.68 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-    </g>
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M2.434,25.752l20.192,0c0.215,0.001 0.394,0.177 0.396,0.393l0,8.438l-0.792,0l0,-8.039l-19.4,0l0,8.039l-0.792,0l0,-8.435c0,-0.217 0.179,-0.395 0.396,-0.396Z'
-        />
-        <path d='M13.048,34.583l-5.548,0l2.774,-3.532l2.774,3.532Zm4.511,0l-2.772,0l1.386,-1.765l1.386,1.765Z'
-        />
-        <path d='M17.883,32.557c1.259,-0.001 2.295,-1.038 2.295,-2.298c0,-1.26 -1.037,-2.297 -2.298,-2.297c-1.26,0 -2.297,1.037 -2.297,2.297c0.002,1.261 1.039,2.298 2.3,2.298Zm0,-3.804c0.825,0.002 1.503,0.681 1.503,1.506c0,0.827 -0.68,1.506 -1.506,1.506c-0.826,0 -1.506,-0.679 -1.506,-1.506c0.001,-0.826 0.681,-1.506 1.507,-1.506l0.002,0Z'
-        />
-    </g>
-</svg>;
-icons.galSlider = <svg viewBox='0 0 48 37' xmlns='http://www.w3.org/2000/svg' fillRule='evenodd'
-clipRule='evenodd' strokeLinejoin='round' strokeMiterlimit='2'>
-    <rect id='Artboard1' width='47.974' height='36.014' fill='none' />
-    <g fill='#3a3a3a' fillRule='nonzero'>
-        <path d='M40.817,5.098l-33.661,0c-0.361,0 -0.659,0.298 -0.659,0.66l0,24.498c0,0.362 0.298,0.66 0.659,0.66l33.661,0c0.362,0 0.66,-0.298 0.66,-0.66l0,-24.503c-0.003,-0.361 -0.3,-0.655 -0.66,-0.655Zm-0.659,24.493l-32.342,0l0,-23.173l32.342,0l0,23.173Z'
-        />
-        <path d='M30.631,27.18l7.521,0l-8.092,-10.303l-3.76,4.788l-6.073,-7.734l-10.405,13.249l20.809,0Z'
-        />
-        <path d='M32.91,16.442c2.1,-0.003 3.827,-1.731 3.827,-3.831c0,-2.101 -1.729,-3.83 -3.831,-3.83c-2.101,0 -3.83,1.729 -3.83,3.83c0.003,2.102 1.732,3.83 3.834,3.831Zm0,-6.341c1.376,0.002 2.507,1.135 2.507,2.51c0,1.377 -1.133,2.511 -2.511,2.511c-1.377,0 -2.51,-1.134 -2.51,-2.511c0.001,-1.377 1.135,-2.51 2.512,-2.51l0.002,0Z'
-        />
-    </g>
-    <circle cx='21.053' cy='32.743' r='0.518' fill='#3a3a3a' />
-    <circle cx='24.053' cy='32.743' r='0.518' fill='#3a3a3a' />
-    <circle cx='27.053' cy='32.743' r='0.518' fill='#3a3a3a' />
-</svg>;
-icons.galtiles =  <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 48 37"
->
-<path fill="none" d="M0 0H47.974V36.014H0z"></path>
-<g fill="#3a3a3a" fillRule="nonzero">
-  <path d="M45.54 1.456H31.111a.398.398 0 00-.396.396v14.696c0 .217.179.395.396.395H45.54a.397.397 0 00.396-.395V1.849a.399.399 0 00-.396-.393zm-.396 14.692H31.507v-13.9h13.637v13.9z"></path>
-  <path d="M41.123 14.376h3.023l-3.253-4.143-1.513 1.926-2.442-3.11-4.183 5.327h8.368zM41.598 6.658a1.504 1.504 0 001.494-1.496c0-.821-.675-1.496-1.496-1.496S40.1 4.341 40.1 5.162a1.506 1.506 0 001.498 1.496zm0-2.477a.985.985 0 01.978.981c0 .537-.442.98-.98.98a.986.986 0 01-.981-.98.987.987 0 01.981-.981h.002zM28.89 1.456H2.434a.398.398 0 00-.396.396v14.696c0 .217.179.395.396.395H28.89a.397.397 0 00.396-.395V1.849a.399.399 0 00-.396-.393zm-.396 14.692H2.83v-13.9h25.664v13.9z"></path>
-  <path d="M19.501 14.702h4.512l-4.854-6.18-2.256 2.872-3.643-4.639-6.241 7.947h12.482zM21.245 8.261a2.31 2.31 0 002.295-2.298 2.308 2.308 0 00-2.298-2.297 2.308 2.308 0 00-2.297 2.297 2.312 2.312 0 002.3 2.298zm0-3.804a1.513 1.513 0 011.503 1.506c0 .826-.68 1.506-1.506 1.506s-1.506-.68-1.506-1.506a1.515 1.515 0 011.507-1.506h.002z"></path>
-  <g>
-	<path d="M45.54 19.071H16.57a.397.397 0 00-.396.395v14.696c0 .217.179.396.396.396h28.97a.398.398 0 00.396-.396V19.463a.398.398 0 00-.396-.392zm-.396 14.692H16.966V19.862h28.178v13.901z"></path>
-	<path d="M34.842 32.317h4.511l-4.854-6.181-2.256 2.872-3.643-4.639-6.241 7.948h12.483zM38.68 25.875a2.308 2.308 0 002.295-2.297 2.309 2.309 0 00-2.298-2.298 2.308 2.308 0 00-2.297 2.298 2.311 2.311 0 002.3 2.297zm0-3.803c.825.001 1.503.68 1.503 1.506s-.679 1.506-1.506 1.506c-.826 0-1.506-.68-1.506-1.506a1.514 1.514 0 011.507-1.506h.002z"></path>
-  </g>
-  <g>
-	<path d="M14.099 19.071H2.434a.397.397 0 00-.396.395v14.696c0 .217.179.396.396.396h11.665a.398.398 0 00.396-.396V19.463a.398.398 0 00-.396-.392zm-.396 14.692H2.83V19.862h10.873v13.901z"></path>
-	<path d="M10.271 32.317h2.255L10.1 29.228l-1.127 1.435-1.821-2.318-3.119 3.972h6.238zM10.674 24.896a1.288 1.288 0 001.28-1.281c0-.703-.579-1.282-1.282-1.282-.702 0-1.281.579-1.281 1.282a1.29 1.29 0 001.283 1.281zm0-2.121c.46.001.838.379.838.84 0 .461-.379.84-.84.84a.844.844 0 01-.84-.84.845.845 0 01.841-.84h.001z"></path>
-  </g>
-</g>
-</svg>;
-icons.thumbslider = <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fillRule="evenodd"
-      strokeLinejoin="round"
-      strokeMiterlimit="2"
-      clipRule="evenodd"
-      viewBox="0 0 48 37"
-    >
-      <g fill="#3a3a3a" fillRule="nonzero">
-        <path d="M40.856 2.066H7.195a.663.663 0 00-.659.66v24.498c0 .362.298.66.659.66h33.661c.362 0 .66-.298.66-.66V2.721a.663.663 0 00-.66-.655zm-.659 24.493H7.855V3.386h32.342v23.173z"></path>
-        <path d="M30.67 24.148h7.521l-8.092-10.303-3.76 4.788-6.073-7.734L9.861 24.148H30.67zM32.949 13.41c2.1-.003 3.827-1.731 3.827-3.831 0-2.101-1.729-3.83-3.831-3.83a3.848 3.848 0 00-3.83 3.83 3.854 3.854 0 003.834 3.831zm0-6.341a2.523 2.523 0 012.507 2.51 2.523 2.523 0 01-2.511 2.511 2.523 2.523 0 01-2.51-2.511 2.524 2.524 0 012.512-2.51h.002zM14.706 28.741H6.631a.16.16 0 00-.158.159v5.876c0 .087.072.158.158.158h8.075a.159.159 0 00.158-.158v-5.878a.16.16 0 00-.158-.157zm-.159 5.875H6.79v-5.558h7.757v5.558z"></path>
-        <path d="M12.262 34.038h1.804l-1.941-2.471-.902 1.148-1.456-1.855-2.496 3.178h4.991zM12.809 31.462a.923.923 0 00.918-.919.923.923 0 00-.919-.918.923.923 0 00-.919.918.924.924 0 00.92.919zm0-1.521c.33.001.601.273.601.602a.605.605 0 01-.602.603.605.605 0 01-.602-.603c0-.33.272-.602.602-.602h.001z"></path>
-        <g>
-          <path d="M23.535 28.741H15.46a.16.16 0 00-.158.159v5.876c0 .087.072.158.158.158h8.075a.159.159 0 00.158-.158v-5.878a.16.16 0 00-.158-.157zm-.158 5.875h-7.758v-5.558h7.758v5.558z"></path>
-          <path d="M21.091 34.038h1.804l-1.941-2.471-.902 1.148-1.456-1.855-2.496 3.178h4.991zM21.638 31.462a.923.923 0 00.918-.919.923.923 0 00-.919-.918.923.923 0 00-.919.918.924.924 0 00.92.919zm0-1.521c.33.001.601.273.601.602a.605.605 0 01-.602.603.605.605 0 01-.602-.603c0-.33.272-.602.602-.602h.001z"></path>
-        </g>
-        <g>
-          <path d="M32.452 28.741h-8.074a.16.16 0 00-.158.159v5.876c0 .087.071.158.158.158h8.074a.16.16 0 00.159-.158v-5.878a.16.16 0 00-.159-.157zm-.158 5.875h-7.758v-5.558h7.758v5.558z"></path>
-          <path d="M30.009 34.038h1.804l-1.941-2.471-.902 1.148-1.457-1.855-2.495 3.178h4.991zM30.556 31.462a.923.923 0 00.918-.919.923.923 0 00-.919-.918.923.923 0 00-.919.918.924.924 0 00.92.919zm0-1.521c.33.001.601.273.601.602a.605.605 0 01-.602.603.605.605 0 01-.602-.603c0-.33.272-.602.602-.602h.001z"></path>
-        </g>
-        <g>
-          <path d="M41.369 28.741h-8.075a.16.16 0 00-.158.159v5.876c0 .087.072.158.158.158h8.075a.159.159 0 00.158-.158v-5.878a.16.16 0 00-.158-.157zm-.159 5.875h-7.757v-5.558h7.757v5.558z"></path>
-          <path d="M38.925 34.038h1.804l-1.941-2.471-.902 1.148-1.456-1.855-2.496 3.178h4.991zM39.472 31.462a.923.923 0 00.918-.919.923.923 0 00-.919-.918.923.923 0 00-.919.918.924.924 0 00.92.919zm0-1.521c.33.001.601.273.601.602a.605.605 0 01-.602.603.605.605 0 01-.602-.603c0-.33.272-.602.602-.602h.001z"></path>
-        </g>
-      </g>
-	</svg>;
-	icons.formBlock = <svg
-	viewBox="0 0 48 48"
-	fillRule="evenodd"
-	clipRule="evenodd"
-	strokeLinejoin="round"
-	width="20px" height="20px"
-	strokeMiterlimit="2"
-  >
-	<path
-	  d="M45.073 15.23l.052.002.051.003.05.005.05.005.05.007.05.009.049.009.048.011.049.011.047.013.048.014.046.015.047.017.045.017.045.019.045.019.044.021.044.021.042.023.043.024.041.024.041.026.041.026.039.028.039.028.038.029.038.031.036.031.036.032.035.032.035.034.033.034.033.035.032.036.031.037.03.037.03.039.028.039.027.039.027.04.025.041.025.042.024.042.022.043.022.044.021.044.019.044.019.045.017.046.016.046.015.047.014.047.013.048.012.048.011.049.009.049.008.049.007.05.006.051.005.05.003.051.002.051v16.075l-.002.051-.003.051-.005.051-.006.05-.007.05-.008.05-.009.049-.011.048-.012.049-.013.047-.014.048-.015.046-.016.047-.017.045-.019.045-.019.045-.021.044-.022.044-.022.042-.024.043-.025.041-.025.041-.027.04-.027.04-.028.039-.03.038-.03.038-.031.036-.032.036-.033.035-.033.035-.035.033-.035.033-.036.032-.036.031-.038.03-.038.029-.039.029-.039.027-.041.027-.041.025-.041.025-.043.024-.042.022-.044.022-.044.02-.045.02-.045.018-.045.018-.047.016-.046.015-.048.014-.047.013-.049.012-.048.011-.049.009-.05.008-.05.007-.05.006-.05.005-.051.003-.052.002H2.975l-.052-.002-.051-.003-.05-.005-.051-.006-.049-.007-.05-.008-.049-.009-.049-.011-.048-.012-.048-.013-.047-.014-.047-.015-.046-.016-.046-.018-.045-.018-.044-.02-.044-.02-.044-.022-.043-.022-.042-.024-.042-.025-.04-.025-.041-.027-.039-.027-.039-.029-.038-.029-.038-.03-.037-.031-.036-.032-.035-.033-.034-.033-.034-.035-.032-.035-.032-.036-.031-.036-.03-.038-.03-.038-.028-.039-.028-.04-.026-.04-.026-.041-.024-.041-.024-.043-.023-.042-.021-.044-.021-.044-.019-.045-.019-.045-.017-.045-.016-.047-.016-.046-.014-.048-.013-.047-.011-.049-.011-.048-.009-.049-.009-.05-.007-.05-.005-.05-.005-.051-.003-.051-.002-.051-.001-.052V17.231l.001-.052.002-.051.003-.051.005-.05.005-.051.007-.05.009-.049.009-.049.011-.049.011-.048.013-.048.014-.047.016-.047.016-.046.017-.046.019-.045.019-.044.021-.044.021-.044.023-.043.024-.042.024-.042.026-.041.026-.04.028-.039.028-.039.03-.039.03-.037.031-.037.032-.036.032-.035.034-.034.034-.034.035-.032.036-.032.037-.031.038-.031.038-.029.039-.028.039-.028.041-.026.04-.026.042-.024.042-.024.043-.023.044-.021.044-.021.044-.019.045-.019.046-.017.046-.017.047-.015.047-.014.048-.013.048-.011.049-.011.049-.009.05-.009.049-.007.051-.005.05-.005.051-.003.052-.002.051-.001h41.996l.051.001zM3.524 17.729v14.975h40.999V17.729H3.524zM46.281 5.839l.04.003.039.005.038.007.038.008.037.011.037.012.035.014.035.016.034.017.032.019.032.021.031.021.03.024.028.024.027.026.026.027.025.029.023.029.022.031.02.032.019.033.017.033.016.035.014.036.012.036.011.037.009.038.006.038.005.04.003.039.001.041v6.244l-.001.041-.003.039-.005.04-.006.038-.009.038-.011.037-.012.036-.014.036-.016.035-.017.033-.019.033-.02.032-.022.031-.023.029-.025.029-.026.027-.027.026-.028.024-.03.024-.031.021-.032.021-.032.019-.034.017-.035.016-.035.014-.037.012-.037.01-.038.009-.038.007-.039.005-.04.003-.04.001H1.807l-.04-.001-.04-.003-.039-.005-.039-.007-.037-.009-.038-.01-.036-.012-.035-.014-.035-.016-.034-.017-.033-.019-.031-.021-.031-.021-.03-.024-.028-.024-.027-.026-.026-.027-.025-.029-.023-.029-.022-.031-.02-.032-.019-.033-.017-.033-.016-.035-.014-.036-.012-.036-.011-.037-.009-.038-.007-.038-.005-.04-.003-.039-.001-.041V6.621l.001-.041.003-.039.005-.04.007-.038.009-.038.011-.037.012-.036.014-.036.016-.035.017-.033.019-.033.02-.032.022-.031.023-.029.025-.029.026-.027.027-.026.028-.024.03-.024.031-.021.031-.021.033-.019.034-.017.035-.016.035-.014.036-.012.038-.011.037-.008.039-.007.039-.005.04-.003.04-.001h44.434l.04.001zM3.524 8.338v2.81h40.999v-2.81H3.524zM22.012 38.096c0-.748-.607-1.355-1.355-1.355H2.332c-.748 0-1.355.607-1.355 1.355v2.711c0 .748.607 1.355 1.355 1.355h18.325c.748 0 1.355-.607 1.355-1.355v-2.711z"
-	  fill="#0058b0"
-	/>
-  </svg>;
-icons.infoStart = <svg
-  xmlns="http://www.w3.org/2000/svg"
-  fillRule="evenodd"
-  strokeLinejoin="round"
-  strokeMiterlimit="2"
-  clipRule="evenodd"
-  viewBox="0 0 400 400"
->
-  <path fill="#eee" d="M46.942 44.98H353.057V355.019H46.942z"></path>
-  <path
-	fill="#565656"
-	d="M289.817 181.49a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-  ></path>
-  <path
-	fill="#888"
-	d="M245.908 310.489a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 260.695a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 237.933a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 215.172a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-  ></path>
-  <path
-	fill="#888"
-	fillRule="nonzero"
-	d="M176.481 80.65h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V84.37a3.721 3.721 0 013.719-3.72zm43.89 47.609V86.601h-41.658v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-  ></path>
-</svg>;
-  icons.infoBasic = <svg
-  xmlns="http://www.w3.org/2000/svg"
-  fillRule="evenodd"
-  strokeLinejoin="round"
-  strokeMiterlimit="2"
-  clipRule="evenodd"
-  viewBox="0 0 400 400"
->
-  <path
-	fill="#fff"
-	d="M350.827 47.314c0-11.515-9.348-20.863-20.863-20.863H70.036c-11.515 0-20.863 9.348-20.863 20.863v305.372c0 11.515 9.348 20.863 20.863 20.863h259.928c11.515 0 20.863-9.348 20.863-20.863V47.314z"
-  ></path>
-  <path
-	fill="#eee"
-	d="M330.137 20.452l.173.001.172.003.172.004.172.005.172.006.172.007.171.008.171.009.342.022.17.012.34.028.17.016.169.017.338.036.168.02.168.021.168.022.168.023.167.024.167.025.167.026.166.027.166.028.165.029.166.03.165.031.164.032.165.033.163.035.164.035.163.036.163.037.163.038.162.039.162.04.161.041.161.042.16.043.161.044.159.045.16.046.159.047.158.047.159.049.157.05.158.05.157.052.156.052.312.108.155.055.155.056.154.057.154.058.153.059.153.06.153.061.152.061.302.126.151.065.15.065.149.066.15.067.149.068.148.068.147.07.148.07.146.071.292.146.145.073.145.075.144.076.143.076.143.077.143.078.142.079.282.16.141.082.139.082.14.083.138.083.276.17.137.086.137.087.136.088.27.178.134.09.134.091.132.092.133.092.132.093.131.094.131.095.13.096.129.096.129.097.136.103.247.192.252.201.25.203.247.207.245.209.242.213.239.215.237.218.235.22.231.224.229.226.226.229.224.232.22.234.218.236.215.24.213.242.209.245.207.247.203.25.201.252.192.248.103.135.097.129.096.129.096.13.095.131.094.131.093.132.092.133.092.132.091.134.09.134.089.135.089.136.088.135.087.137.086.137.17.276.083.138.083.14.082.139.082.141.16.282.079.142.078.142.077.144.076.143.076.144.075.145.073.145.073.146.072.146.072.147.14.294.068.149.068.148.067.149.066.15.065.15.065.151.063.151.062.151.062.152.061.153.059.152.06.154.058.154.057.154.056.154.055.156.108.312.052.156.052.157.1.316.049.158.047.158.047.159.046.159.045.16.044.161.043.16.042.161.041.162.04.161.039.162.038.163.037.163.036.163.035.163.035.165.033.164.032.164.031.166.03.164.029.166.028.166.027.166.026.167.025.167.024.167.023.168.022.168.021.168.02.168.018.168.018.17.017.17.015.169.015.17.014.17.012.17.022.342.009.17.008.172.007.172.006.172.005.172.004.171.003.173.001.173.001.163v305.392l-.001.163-.001.173-.003.173-.004.172-.005.171-.006.172-.007.172-.008.172-.009.17-.011.171-.011.171-.012.17-.014.17-.015.17-.015.169-.017.17-.018.17-.018.167-.02.169-.021.168-.022.168-.023.168-.024.167-.025.167-.026.167-.027.166-.028.166-.029.166-.03.164-.031.166-.032.164-.033.164-.035.165-.035.163-.036.163-.037.163-.038.163-.039.162-.04.161-.041.162-.042.161-.043.16-.044.161-.045.16-.046.159-.047.159-.047.158-.049.158-.05.158-.05.158-.052.157-.052.156-.054.156-.054.156-.055.155-.056.155-.057.154-.058.154-.059.154-.06.152-.061.153-.061.152-.063.151-.063.151-.065.151-.065.15-.066.15-.067.149-.068.148-.068.149-.07.147-.07.148-.071.146-.073.146-.073.146-.073.145-.075.145-.076.144-.076.143-.077.144-.078.142-.079.142-.079.141-.081.142-.082.14-.082.139-.083.14-.083.138-.085.138-.085.138-.086.137-.087.137-.088.136-.089.135-.089.135-.09.134-.091.134-.092.132-.092.133-.093.132-.094.131-.095.131-.096.13-.096.129-.097.129-.103.136-.192.247-.201.252-.203.25-.207.247-.209.245-.213.242-.215.24-.217.236-.221.235-.224.231-.226.229-.229.226-.232.224-.234.22-.236.218-.24.215-.242.213-.245.209-.247.207-.25.203-.252.201-.248.192-.135.103-.129.097-.129.096-.13.096-.131.095-.131.094-.132.093-.133.093-.132.091-.134.091-.134.09-.135.089-.135.089-.136.088-.137.087-.137.086-.138.085-.138.085-.138.083-.14.083-.139.082-.14.082-.142.081-.141.079-.142.079-.142.078-.144.077-.143.076-.144.076-.145.075-.145.074-.146.072-.146.073-.146.071-.148.07-.147.07-.149.068-.148.068-.15.067-.149.066-.149.065-.152.065-.151.063-.151.063-.152.061-.153.061-.152.06-.154.059-.154.058-.154.057-.155.056-.155.055-.156.054-.156.054-.156.052-.157.052-.158.05-.158.05-.158.049-.158.047-.159.047-.159.046-.16.045-.161.044-.16.043-.161.042-.162.041-.161.04-.162.039-.163.038-.163.037-.163.036-.163.035-.165.035-.164.033-.164.032-.166.031-.164.03-.166.029-.167.028-.165.027-.167.026-.167.025-.167.024-.168.023-.168.022-.168.021-.169.02-.167.018-.17.018-.17.017-.169.015-.17.015-.17.014-.17.012-.171.011-.171.011-.17.009-.172.008-.172.007-.172.006-.172.005-.171.004-.173.003-.173.001-.163.001H70.026l-.163-.001-.173-.001-.173-.003-.171-.004-.172-.005-.172-.006-.172-.007-.172-.008-.17-.009-.171-.011-.171-.011-.17-.012-.17-.014-.17-.015-.169-.015-.17-.017-.17-.018-.167-.018-.169-.02-.168-.021-.168-.022-.168-.023-.167-.024-.167-.025-.167-.026-.165-.027-.167-.028-.166-.029-.164-.03-.166-.031-.164-.032-.164-.033-.165-.035-.163-.035-.163-.036-.163-.037-.163-.038-.162-.039-.161-.04-.162-.041-.161-.042-.16-.043-.161-.044-.159-.045-.16-.046-.159-.047-.158-.047-.158-.049-.158-.05-.158-.05-.157-.052-.156-.052-.156-.054-.156-.054-.155-.055-.155-.056-.154-.057-.154-.058-.154-.06-.152-.059-.153-.061-.152-.061-.151-.063-.151-.063-.151-.065-.15-.065-.149-.066-.15-.067-.148-.067-.149-.069-.147-.07-.147-.07-.147-.071-.146-.073-.146-.072-.145-.074-.145-.075-.144-.076-.143-.076-.144-.077-.142-.078-.142-.079-.141-.079-.141-.081-.141-.082-.139-.082-.14-.083-.138-.083-.138-.085-.138-.085-.137-.086-.137-.087-.135-.088-.136-.089-.135-.089-.134-.09-.134-.091-.132-.091-.133-.093-.132-.093-.131-.094-.131-.095-.13-.096-.129-.096-.129-.096-.135-.104-.248-.192-.252-.201-.25-.203-.247-.207-.245-.209-.242-.213-.24-.215-.236-.218-.234-.22-.232-.224-.229-.226-.226-.229-.224-.231-.221-.235-.217-.237-.215-.239-.213-.242-.209-.245-.207-.247-.203-.25-.201-.252-.192-.247-.103-.136-.097-.129-.097-.129-.095-.13-.095-.131-.094-.131-.093-.132-.092-.132-.092-.133-.091-.134-.09-.134-.089-.135-.089-.135-.088-.136-.086-.137-.087-.137-.085-.138-.085-.138-.083-.138-.083-.14-.082-.139-.082-.141-.08-.141-.08-.141-.079-.142-.078-.143-.077-.143-.076-.143-.076-.144-.075-.145-.073-.145-.073-.146-.073-.146-.071-.146-.07-.148-.07-.147-.068-.148-.068-.149-.067-.15-.066-.149-.065-.15-.065-.151-.063-.151-.063-.151-.061-.152-.061-.153-.06-.153-.059-.153-.057-.154-.058-.154-.056-.155-.055-.155-.054-.156-.054-.156-.052-.156-.052-.157-.05-.158-.05-.157-.049-.158-.047-.159-.047-.159-.046-.16-.045-.159-.044-.161-.043-.16-.042-.161-.041-.161-.04-.162-.039-.162-.038-.163-.037-.163-.036-.163-.035-.163-.035-.164-.033-.165-.032-.164-.031-.165-.03-.166-.029-.165-.028-.166-.027-.166-.026-.167-.025-.167-.024-.167-.023-.168-.022-.168-.021-.168-.02-.168-.018-.169-.018-.169-.017-.169-.015-.17-.015-.17-.014-.17-.012-.17-.011-.171-.011-.171-.009-.171-.008-.171-.007-.172-.006-.172-.005-.171-.004-.173-.003-.172-.001-.173-.001-.163V47.304l.001-.163.001-.173.003-.172.004-.172.005-.172.006-.172.007-.172.008-.171.009-.171.011-.171.011-.171.012-.17.014-.17.015-.17.015-.17.017-.169.018-.169.018-.169.02-.168.021-.168.022-.168.023-.168.024-.167.025-.167.026-.167.027-.166.028-.166.029-.165.03-.166.031-.165.032-.164.033-.165.035-.164.035-.163.036-.163.037-.163.038-.163.039-.162.04-.162.041-.161.042-.161.043-.16.044-.161.045-.159.046-.16.047-.159.047-.158.049-.159.05-.157.05-.158.052-.157.052-.156.054-.156.054-.156.055-.155.056-.155.057-.154.058-.154.059-.153.06-.153.061-.153.061-.152.063-.151.063-.151.065-.151.065-.15.066-.15.067-.149.068-.148.068-.149.07-.147.07-.147.072-.147.072-.146.073-.146.073-.145.075-.145.076-.144.076-.143.077-.143.078-.143.079-.142.08-.141.08-.141.082-.141.082-.139.083-.14.083-.138.085-.139.085-.137.086-.137.087-.137.088-.136.089-.135.089-.135.09-.134.091-.133.092-.134.092-.132.093-.132.094-.131.095-.131.095-.13.097-.129.097-.129.103-.136.192-.247.201-.252.203-.25.207-.247.209-.245.213-.242.215-.239.218-.237.22-.235.224-.231.226-.229.229-.226.231-.224.235-.22.237-.218.239-.215.242-.213.245-.209.247-.207.25-.203.252-.201.248-.192.135-.103.129-.097.129-.097.13-.095.131-.095.131-.094.132-.093.132-.092.134-.092.133-.091.134-.09.135-.089.135-.089.136-.088.137-.087.137-.086.137-.085.139-.085.138-.083.14-.083.139-.083.141-.081.141-.08.141-.08.142-.079.143-.078.143-.077.143-.076.144-.076.145-.075.145-.073.146-.073.146-.073.147-.071.147-.07.147-.07.149-.068.148-.068.149-.067.15-.066.15-.065.151-.065.151-.063.151-.063.152-.061.153-.061.153-.06.153-.059.154-.058.154-.057.155-.056.155-.055.156-.054.156-.054.156-.052.157-.052.158-.05.157-.05.159-.049.158-.047.159-.047.16-.046.159-.045.161-.044.16-.043.161-.042.162-.041.161-.04.162-.039.163-.038.163-.037.163-.036.163-.035.164-.035.165-.033.164-.032.165-.031.166-.03.165-.029.166-.028.167-.027.166-.026.167-.025.167-.024.168-.023.168-.022.168-.021.168-.02.169-.018.169-.018.169-.017.17-.016.17-.014.17-.014.17-.012.171-.011.171-.011.171-.009.171-.008.172-.007.172-.006.172-.005.172-.004.172-.003.173-.001.163-.001h259.948l.163.001zM70.036 26.451l-.135.001-.135.001-.134.002-.134.003-.134.004-.134.005-.134.005-.133.006-.133.008-.133.008-.133.008-.132.01-.133.011-.132.011-.131.012-.132.013-.131.014-.131.014-.131.016-.131.016-.13.017-.13.018-.13.018-.13.02-.129.02-.129.021-.129.022-.128.022-.128.024-.128.024-.128.025-.127.025-.128.027-.126.027-.127.028-.126.029-.126.029-.126.031-.125.031-.125.032-.125.032-.125.033-.124.034-.124.035-.123.036-.124.036-.123.037-.122.038-.122.038-.122.04-.122.039-.121.041-.121.042-.121.042-.12.042-.12.044-.12.044-.119.045-.119.046-.119.046-.118.047-.118.048-.117.049-.118.049-.116.05-.117.05-.116.051-.116.052-.115.053-.115.053-.114.054-.115.055-.113.055-.114.056-.113.057-.112.057-.113.058-.112.059-.111.059-.111.06-.111.06-.11.062-.11.061-.109.063-.109.063-.109.064-.108.064-.108.065-.107.066-.107.066-.106.067-.106.068-.106.068-.105.069-.105.069-.104.07-.104.071-.103.071-.103.072-.103.072-.102.073-.101.074-.102.074-.1.075-.1.076-.1.076-.198.153-.196.156-.194.159-.193.16-.19.163-.188.165-.187.168-.184.169-.182.172-.18.174-.178.176-.176.178-.174.18-.172.182-.169.184-.168.187-.165.188-.163.19-.16.192-.159.195-.156.196-.153.198-.076.1-.076.1-.075.1-.074.101-.073.102-.074.102-.072.103-.072.103-.071.103-.071.104-.07.104-.069.105-.069.105-.068.106-.068.106-.067.106-.066.107-.066.107-.065.108-.064.108-.064.109-.063.109-.063.109-.061.11-.062.11-.06.111-.06.111-.059.111-.059.112-.058.112-.057.113-.057.113-.056.114-.055.113-.055.115-.054.114-.053.115-.053.115-.052.116-.051.116-.05.117-.05.116-.049.118-.049.117-.048.118-.047.118-.046.119-.046.119-.045.119-.044.12-.043.12-.043.12-.042.121-.042.121-.04.121-.04.122-.04.121-.038.123-.038.122-.037.123-.036.124-.036.123-.035.124-.034.124-.033.125-.032.125-.032.125-.031.125-.031.126-.029.126-.029.126-.028.127-.027.126-.027.128-.025.127-.025.128-.024.128-.024.128-.022.128-.022.129-.021.129-.02.129-.02.13-.018.13-.018.13-.017.13-.016.131-.016.131-.014.131-.014.131-.013.132-.012.131-.011.132-.011.133-.01.132-.008.133-.008.133-.008.133-.006.133-.005.134-.005.134-.004.134-.003.134-.002.134-.001.135-.001.135v305.372l.002.27.002.134.003.134.004.134.01.268.014.266.016.266.021.265.023.263.027.263.03.262.033.261.036.26.04.259.043.258.046.256.049.256.052.255.055.253.058.252.062.251.064.25.067.249.071.247.073.247.076.245.08.243.082.242.085.241.087.24.091.238.093.237.097.235.099.234.101.233.105.231.107.229.11.228.113.227.115.225.118.223.12.222.123.22.126.218.128.217.131.215.133.213.136.212.138.21.141.208.143.206.146.205.147.203.151.2.076.1.153.198.156.196.159.195.16.192.163.19.165.188.168.187.169.184.172.182.174.18.176.178.178.176.18.174.182.172.184.169.187.168.188.165.19.163.193.16.194.159.196.156.198.153.2.152.202.149.203.147.206.144.207.142.209.139.211.137.212.135.214.132.216.129.218.127.219.124.221.122.222.119.225.117.225.114.227.111.229.109.23.106.232.103.233.1.235.098.236.095.238.092.239.089.24.086.242.084.243.08.244.078.245.075.247.072.248.069.25.065.25.063.252.06.253.057.254.054.255.05.256.048.257.044.258.041.26.038.26.035.262.032.262.028.263.025.265.022.265.018.266.016.267.011.134.005.134.004.134.003.134.002.27.002h259.928l.135-.001.135-.001.134-.002.134-.003.134-.004.134-.005.134-.005.133-.006.133-.008.133-.008.133-.008.132-.01.133-.011.132-.011.131-.012.132-.013.131-.014.131-.014.131-.016.131-.016.13-.017.13-.018.13-.018.13-.02.129-.02.129-.021.129-.022.128-.022.128-.024.128-.024.128-.025.127-.025.128-.027.126-.027.127-.028.126-.029.126-.029.126-.031.125-.031.125-.032.125-.032.125-.033.124-.034.124-.035.123-.036.124-.036.123-.037.122-.038.122-.038.122-.04.122-.039.121-.041.121-.042.121-.042.12-.042.12-.044.12-.044.119-.045.119-.046.119-.046.118-.047.118-.048.117-.049.118-.049.116-.05.117-.05.116-.051.116-.052.115-.053.115-.053.114-.054.115-.055.113-.055.114-.056.113-.057.112-.057.113-.058.112-.059.111-.059.111-.06.111-.06.11-.062.11-.061.109-.063.109-.063.109-.064.108-.064.108-.065.107-.066.107-.066.106-.067.106-.068.106-.068.105-.069.105-.069.104-.07.104-.071.103-.071.103-.072.103-.072.102-.073.101-.074.102-.074.1-.075.1-.076.1-.076.198-.153.196-.156.194-.159.193-.16.19-.163.188-.165.187-.168.184-.169.182-.172.18-.174.178-.176.176-.178.174-.18.172-.182.169-.184.168-.187.165-.188.163-.19.16-.192.159-.195.156-.196.153-.198.076-.1.076-.1.075-.1.074-.101.073-.102.074-.102.072-.103.072-.103.071-.103.071-.104.07-.104.069-.105.069-.105.068-.106.068-.106.067-.106.066-.107.066-.107.065-.108.064-.108.064-.109.063-.109.063-.109.061-.11.062-.11.06-.111.06-.111.059-.111.059-.112.058-.112.057-.113.057-.113.056-.114.055-.113.055-.115.054-.114.053-.115.053-.115.052-.116.051-.116.05-.117.05-.116.049-.118.049-.117.048-.118.047-.118.046-.119.046-.119.045-.119.044-.12.043-.12.043-.12.042-.121.042-.121.04-.121.04-.122.04-.121.038-.123.038-.122.037-.123.036-.124.036-.123.035-.124.034-.124.033-.125.032-.125.032-.125.031-.125.031-.126.029-.126.029-.126.028-.127.027-.126.027-.128.025-.127.025-.128.024-.128.024-.128.022-.128.022-.129.021-.129.02-.129.02-.13.018-.13.018-.13.017-.13.016-.131.016-.131.014-.131.014-.131.013-.132.012-.131.011-.132.011-.133.01-.132.008-.133.008-.133.008-.133.006-.133.005-.134.005-.134.004-.134.003-.134.002-.134.001-.135.001-.135V47.314l-.001-.135-.001-.135-.002-.134-.003-.134-.004-.134-.005-.134-.005-.134-.006-.133-.008-.133-.008-.133-.008-.133-.01-.132-.011-.133-.011-.132-.012-.131-.013-.132-.014-.131-.014-.131-.016-.131-.016-.131-.017-.13-.018-.13-.018-.13-.02-.13-.02-.129-.021-.129-.022-.129-.022-.128-.024-.128-.024-.128-.025-.128-.025-.127-.027-.128-.027-.126-.028-.127-.029-.126-.029-.126-.031-.126-.031-.125-.032-.125-.032-.125-.033-.125-.034-.124-.035-.124-.036-.123-.036-.124-.037-.123-.038-.122-.038-.123-.04-.121-.04-.122-.04-.121-.042-.121-.042-.121-.043-.12-.043-.12-.044-.12-.045-.119-.046-.119-.046-.119-.047-.118-.048-.118-.049-.117-.049-.118-.05-.116-.05-.117-.051-.116-.052-.116-.053-.115-.053-.115-.054-.114-.055-.115-.055-.113-.056-.114-.057-.113-.057-.113-.058-.112-.059-.112-.059-.111-.06-.111-.06-.111-.062-.11-.061-.11-.063-.109-.063-.109-.064-.109-.064-.108-.065-.108-.066-.107-.066-.107-.067-.106-.068-.106-.068-.106-.069-.105-.069-.105-.07-.104-.071-.104-.071-.103-.072-.103-.072-.103-.074-.102-.073-.102-.074-.101-.075-.1-.076-.1-.076-.1-.153-.198-.156-.196-.159-.195-.16-.192-.163-.19-.165-.188-.168-.187-.169-.184-.172-.182-.174-.18-.176-.178-.178-.176-.18-.174-.182-.172-.184-.169-.187-.168-.188-.165-.19-.163-.193-.16-.194-.159-.196-.156-.198-.153-.1-.076-.1-.076-.1-.075-.102-.074-.101-.074-.102-.073-.103-.072-.103-.072-.103-.071-.104-.071-.104-.07-.105-.069-.105-.069-.106-.068-.106-.068-.106-.067-.107-.066-.107-.066-.108-.065-.108-.064-.109-.064-.109-.063-.109-.063-.11-.061-.11-.062-.111-.06-.111-.06-.111-.059-.112-.059-.113-.058-.112-.057-.113-.057-.114-.056-.113-.055-.115-.055-.114-.054-.115-.053-.115-.053-.116-.052-.116-.051-.117-.05-.116-.05-.118-.049-.117-.049-.118-.048-.118-.047-.119-.046-.119-.046-.119-.045-.12-.044-.12-.044-.12-.042-.121-.042-.121-.042-.121-.041-.122-.039-.122-.04-.122-.038-.122-.038-.123-.037-.124-.036-.123-.036-.124-.035-.124-.034-.125-.033-.125-.032-.125-.032-.125-.031-.126-.031-.126-.029-.126-.029-.127-.028-.126-.027-.128-.027-.127-.025-.128-.025-.128-.024-.128-.024-.128-.022-.129-.022-.129-.021-.129-.02-.13-.02-.13-.018-.13-.018-.13-.017-.131-.016-.131-.016-.131-.014-.131-.014-.132-.013-.131-.012-.132-.011-.133-.011-.132-.01-.133-.008-.133-.008-.133-.008-.133-.006-.134-.005-.134-.005-.134-.004-.134-.003-.134-.002-.135-.001-.135-.001H70.036z"
-  ></path>
-  <circle cx="200" cy="119.069" r="70.019" fill="#eee"></circle>
-  <path
-	fill="#565656"
-	d="M289.817 213.579a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-  ></path>
-  <path
-	fill="#888"
-	d="M245.908 342.578a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 292.784a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 270.022a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 247.261a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-  ></path>
-  <path
-	fill="#888"
-	fillRule="nonzero"
-	d="M176.939 94.182h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V97.902a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658h-41.658v41.658h41.658zM200 109.06c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-  ></path>
-</svg>;
-icons.infoLeftAbove = <svg
-      xmlns="http://www.w3.org/2000/svg"
-      fillRule="evenodd"
-      strokeLinejoin="round"
-      strokeMiterlimit="2"
-      clipRule="evenodd"
-      viewBox="0 0 400 400"
-    >
-      <path
-        fill="#fff"
-        d="M7.03 60.607H392.61699999999996V345.77200000000005H7.03z"
-      ></path>
-      <path
-        fill="#eee"
-        d="M392.617 345.772H7.03V60.607h385.587v285.165zM12.03 65.607v275.165h375.587V65.607H12.03z"
-      ></path>
-      <path
-        fill="#eee"
-        d="M29.352 81.149H151.446V196.28300000000002H29.352z"
-      ></path>
-      <path
-        fill="#888"
-        fillRule="nonzero"
-        d="M67.338 111.82h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H67.338a3.72 3.72 0 01-3.719-3.719V115.54a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658H69.57v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H75.521a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-      ></path>
-      <path
-        fill="#565656"
-        d="M299.746 222.29a7.59 7.59 0 00-7.585-7.586H35.651a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h256.51a7.588 7.588 0 007.585-7.585v-3.834z"
-      ></path>
-      <path
-        fill="#888"
-        d="M372.406 285.764a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h333.112a5.687 5.687 0 005.683-5.684v-2.674zM266.41 313.012a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h227.116a5.686 5.686 0 005.683-5.683v-2.675zM370.188 257.889a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h330.894a5.687 5.687 0 005.683-5.684v-2.674z"
-      ></path>
-	</svg>;
-icons.infoLeft = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  stroke="#eee"
-  strokeWidth="5"
-  d="M390.993 128.492c0-6.336-5.144-11.479-11.479-11.479H20.486c-6.335 0-11.479 5.143-11.479 11.479v143.016c0 6.336 5.144 11.479 11.479 11.479h359.028c6.335 0 11.479-5.143 11.479-11.479V128.492z"
-></path>
-<circle cx="78.187" cy="200" r="59.757" fill="#eee"></circle>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M55.126 173.22h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H55.126a3.72 3.72 0 01-3.719-3.719V176.94a3.721 3.721 0 013.719-3.72zm43.89 47.609v-41.658H57.358v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H63.309a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-<path
-  fill="#565656"
-  d="M335.948 154.253a7.59 7.59 0 00-7.585-7.586H163.899a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h164.464a7.588 7.588 0 007.585-7.585v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M368.797 217.727a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h201.255a5.687 5.687 0 005.683-5.684v-2.674zM303.84 244.975a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h136.298a5.686 5.686 0 005.683-5.683v-2.675zM368.797 189.852a5.686 5.686 0 00-5.683-5.683H161.859a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h201.255a5.687 5.687 0 005.683-5.684v-2.674z"
-></path>
-</svg>;
-icons.infoTopOverlay = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  d="M350.827 132.614c0-11.515-9.348-20.863-20.863-20.863H70.036c-11.515 0-20.863 9.348-20.863 20.863v209.179c0 11.515 9.348 20.863 20.863 20.863h259.928c11.515 0 20.863-9.348 20.863-20.863V132.614z"
-></path>
-<path
-  fill="#eee"
-  d="M330.137 106.007l.173.001.172.003.172.004.172.005.172.006.172.007.171.008.171.009.342.022.17.012.34.028.17.016.169.017.338.036.168.02.168.021.168.022.168.023.167.024.167.025.167.026.166.027.166.028.165.029.166.03.165.031.164.032.165.033.163.035.164.035.163.036.163.037.163.038.162.039.162.04.161.041.161.042.16.043.161.044.159.045.16.046.159.047.158.047.159.049.157.05.158.05.157.052.156.052.312.108.155.055.155.056.154.057.154.058.153.059.153.06.153.061.152.061.302.126.151.065.15.065.149.066.15.067.149.068.148.068.147.07.148.07.146.071.292.146.145.073.145.075.144.076.143.076.143.077.143.078.142.079.282.16.141.082.139.082.14.083.138.083.276.17.137.086.137.087.136.088.27.178.134.09.134.091.132.092.133.092.132.093.131.094.131.095.13.096.129.096.129.097.136.103.247.192.252.201.25.203.247.207.245.209.242.213.239.215.237.218.235.22.231.224.229.226.226.229.224.232.22.234.218.236.215.24.213.242.209.245.207.247.203.25.201.252.192.248.103.135.097.129.096.129.096.13.095.131.094.131.093.132.092.133.092.132.091.134.09.134.089.135.089.136.088.135.087.137.086.137.17.276.083.138.083.14.082.139.082.141.16.282.079.142.078.142.077.144.076.143.076.144.075.145.073.145.073.146.072.146.072.147.14.294.068.149.068.148.067.149.066.15.065.15.065.151.063.151.062.151.062.152.061.153.059.152.06.154.058.154.057.154.056.154.055.156.108.312.052.156.052.157.1.316.049.158.047.158.047.159.046.159.045.16.044.161.043.16.042.161.041.162.04.161.039.162.038.163.037.163.036.163.035.163.035.165.033.164.032.164.031.166.03.164.029.166.028.166.027.166.026.167.025.167.024.167.023.168.022.168.021.168.02.168.018.168.018.17.017.17.015.169.015.17.014.17.012.17.022.342.009.17.008.172.007.172.006.172.005.172.004.171.003.173.001.173.001.163v208.944l-.001.163-.001.173-.003.173-.004.172-.005.171-.006.172-.007.172-.008.172-.009.17-.011.171-.011.171-.012.17-.014.17-.015.17-.015.169-.017.17-.018.17-.018.167-.02.169-.021.168-.022.168-.023.168-.024.167-.025.167-.026.167-.027.166-.028.166-.029.166-.03.164-.031.166-.032.164-.033.164-.035.165-.035.163-.036.163-.037.163-.038.163-.039.162-.04.161-.041.162-.042.161-.043.16-.044.161-.045.16-.046.159-.047.159-.047.158-.049.158-.05.158-.05.158-.052.157-.052.156-.054.156-.054.156-.055.155-.056.155-.057.154-.058.154-.059.154-.06.152-.061.153-.061.152-.063.151-.063.151-.065.151-.065.15-.066.15-.067.149-.068.148-.068.149-.07.147-.07.148-.071.146-.073.146-.073.146-.073.145-.075.145-.076.144-.076.143-.077.144-.078.142-.079.142-.079.141-.081.142-.082.14-.082.139-.083.14-.083.138-.085.138-.085.138-.086.137-.087.137-.088.136-.089.135-.089.135-.09.134-.091.134-.092.132-.092.133-.093.132-.094.131-.095.131-.096.13-.096.129-.097.129-.103.136-.192.247-.201.252-.203.25-.207.247-.209.245-.213.242-.215.24-.217.236-.221.235-.224.231-.226.229-.229.226-.232.224-.234.22-.236.218-.24.215-.242.213-.245.209-.247.207-.25.203-.252.201-.248.192-.135.103-.129.097-.129.096-.13.096-.131.095-.131.094-.132.093-.133.093-.132.091-.134.091-.134.09-.135.089-.135.089-.136.088-.137.087-.137.086-.138.085-.138.085-.138.083-.14.083-.139.082-.14.082-.142.081-.141.079-.142.079-.142.078-.144.077-.143.076-.144.076-.145.075-.145.074-.146.072-.146.073-.146.071-.148.07-.147.07-.149.068-.148.068-.15.067-.149.066-.149.065-.152.065-.151.063-.151.063-.152.061-.153.061-.152.06-.154.059-.154.058-.154.057-.155.056-.155.055-.156.054-.156.054-.156.052-.157.052-.158.05-.158.05-.158.049-.158.047-.159.047-.159.046-.16.045-.161.044-.16.043-.161.042-.162.041-.161.04-.162.039-.163.038-.163.037-.163.036-.163.035-.165.035-.164.033-.164.032-.166.031-.164.03-.166.029-.167.028-.165.027-.167.026-.167.025-.167.024-.168.023-.168.022-.168.021-.169.02-.167.018-.17.018-.17.017-.169.015-.17.015-.17.014-.17.012-.171.011-.171.011-.17.009-.172.008-.172.007-.172.006-.172.005-.171.004-.173.003-.173.001-.163.001H70.026l-.163-.001-.173-.001-.173-.003-.171-.004-.172-.005-.172-.006-.172-.007-.172-.008-.17-.009-.171-.011-.171-.011-.17-.012-.17-.014-.17-.015-.169-.015-.17-.017-.17-.018-.167-.018-.169-.02-.168-.021-.168-.022-.168-.023-.167-.024-.167-.025-.167-.026-.165-.027-.167-.028-.166-.029-.164-.03-.166-.031-.164-.032-.164-.033-.165-.035-.163-.035-.163-.036-.163-.037-.163-.038-.162-.039-.161-.04-.162-.041-.161-.042-.16-.043-.161-.044-.159-.045-.16-.046-.159-.047-.158-.047-.158-.049-.158-.05-.158-.05-.157-.052-.156-.052-.156-.054-.156-.054-.155-.055-.155-.056-.154-.057-.154-.058-.154-.06-.152-.059-.153-.061-.152-.061-.151-.063-.151-.063-.151-.065-.15-.065-.149-.066-.15-.067-.148-.067-.149-.069-.147-.07-.147-.07-.147-.071-.146-.073-.146-.072-.145-.074-.145-.075-.144-.076-.143-.076-.144-.077-.142-.078-.142-.079-.141-.079-.141-.081-.141-.082-.139-.082-.14-.083-.138-.083-.138-.085-.138-.085-.137-.086-.137-.087-.135-.088-.136-.089-.135-.089-.134-.09-.134-.091-.132-.091-.133-.093-.132-.093-.131-.094-.131-.095-.13-.096-.129-.096-.129-.096-.135-.104-.248-.192-.252-.201-.25-.203-.247-.207-.245-.209-.242-.213-.24-.215-.236-.218-.234-.22-.232-.224-.229-.226-.226-.229-.224-.231-.221-.235-.217-.237-.215-.239-.213-.242-.209-.245-.207-.247-.203-.25-.201-.252-.192-.247-.103-.136-.097-.129-.097-.129-.095-.13-.095-.131-.094-.131-.093-.132-.092-.132-.092-.133-.091-.134-.09-.134-.089-.135-.089-.135-.088-.136-.086-.137-.087-.137-.085-.138-.085-.138-.083-.138-.083-.14-.082-.139-.082-.141-.08-.141-.08-.141-.079-.142-.078-.143-.077-.143-.076-.143-.076-.144-.075-.145-.073-.145-.073-.146-.073-.146-.071-.146-.07-.148-.07-.147-.068-.148-.068-.149-.067-.15-.066-.149-.065-.15-.065-.151-.063-.151-.063-.151-.061-.152-.061-.153-.06-.153-.059-.153-.057-.154-.058-.154-.056-.155-.055-.155-.054-.156-.054-.156-.052-.156-.052-.157-.05-.158-.05-.157-.049-.158-.047-.159-.047-.159-.046-.16-.045-.159-.044-.161-.043-.16-.042-.161-.041-.161-.04-.162-.039-.162-.038-.163-.037-.163-.036-.163-.035-.163-.035-.164-.033-.165-.032-.164-.031-.165-.03-.166-.029-.165-.028-.166-.027-.166-.026-.167-.025-.167-.024-.167-.023-.168-.022-.168-.021-.168-.02-.168-.018-.169-.018-.169-.017-.169-.015-.17-.015-.17-.014-.17-.012-.17-.011-.171-.011-.171-.009-.171-.008-.171-.007-.172-.006-.172-.005-.171-.004-.173-.003-.172-.001-.173-.001-.163V132.859l.001-.163.001-.173.003-.172.004-.172.005-.172.006-.172.007-.172.008-.171.009-.171.011-.171.011-.171.012-.17.014-.17.015-.17.015-.17.017-.169.018-.169.018-.169.02-.168.021-.168.022-.168.023-.168.024-.167.025-.167.026-.167.027-.166.028-.166.029-.165.03-.166.031-.165.032-.164.033-.165.035-.164.035-.163.036-.163.037-.163.038-.163.039-.162.04-.162.041-.161.042-.161.043-.16.044-.161.045-.159.046-.16.047-.159.047-.158.049-.159.05-.157.05-.158.052-.157.052-.156.054-.156.054-.156.055-.155.056-.155.057-.154.058-.154.059-.153.06-.153.061-.153.061-.152.063-.151.063-.151.065-.151.065-.15.066-.15.067-.149.068-.148.068-.149.07-.147.07-.147.072-.147.072-.146.073-.146.073-.145.075-.145.076-.144.076-.143.077-.143.078-.143.079-.142.08-.141.08-.141.082-.141.082-.139.083-.14.083-.138.085-.139.085-.137.086-.137.087-.137.088-.136.089-.135.089-.135.09-.134.091-.133.092-.134.092-.132.093-.132.094-.131.095-.131.095-.13.097-.129.097-.129.103-.136.192-.247.201-.252.203-.25.207-.247.209-.245.213-.242.215-.239.218-.237.22-.235.224-.231.226-.229.229-.226.231-.224.235-.22.237-.218.239-.215.242-.213.245-.209.247-.207.25-.203.252-.201.248-.192.135-.103.129-.097.129-.097.13-.095.131-.095.131-.094.132-.093.132-.092.134-.092.133-.091.134-.09.135-.089.135-.089.136-.088.137-.087.137-.086.137-.085.139-.085.138-.083.14-.083.139-.083.141-.081.141-.08.141-.08.142-.079.143-.078.143-.077.143-.076.144-.076.145-.075.145-.073.146-.073.146-.073.147-.071.147-.07.147-.07.149-.068.148-.068.149-.067.15-.066.15-.065.151-.065.151-.063.151-.063.152-.061.153-.061.153-.06.153-.059.154-.058.154-.057.155-.056.155-.055.156-.054.156-.054.156-.052.157-.052.158-.05.157-.05.159-.049.158-.047.159-.047.16-.046.159-.045.161-.044.16-.043.161-.042.162-.041.161-.04.162-.039.163-.038.163-.037.163-.036.163-.035.164-.035.165-.033.164-.032.165-.031.166-.03.165-.029.166-.028.167-.027.166-.026.167-.025.167-.024.168-.023.168-.022.168-.021.168-.02.169-.018.169-.018.169-.017.17-.016.17-.014.17-.014.17-.012.171-.011.171-.011.171-.009.171-.008.172-.007.172-.006.172-.005.172-.004.172-.003.173-.001.163-.001h259.948l.163.001zm-260.101 5.999l-.135.001-.135.001-.134.002-.134.003-.134.004-.134.005-.134.005-.133.006-.133.008-.133.008-.133.008-.132.01-.133.011-.132.011-.131.012-.132.013-.131.014-.131.014-.131.016-.131.016-.13.017-.13.018-.13.018-.13.02-.129.02-.129.021-.129.022-.128.022-.128.024-.128.024-.128.025-.127.025-.128.027-.126.027-.127.028-.126.029-.126.029-.126.031-.125.031-.125.032-.125.032-.125.033-.124.034-.124.035-.123.036-.124.036-.123.037-.122.038-.122.038-.122.04-.122.039-.121.041-.121.042-.121.042-.12.042-.12.044-.12.044-.119.045-.119.046-.119.046-.118.047-.118.048-.117.049-.118.049-.116.05-.117.05-.116.051-.116.052-.115.053-.115.053-.114.054-.115.055-.113.055-.114.056-.113.057-.112.057-.113.058-.112.059-.111.059-.111.06-.111.06-.11.062-.11.061-.109.063-.109.063-.109.064-.108.064-.108.065-.107.066-.107.066-.106.067-.106.068-.106.068-.105.069-.105.069-.104.07-.104.071-.103.071-.103.072-.103.072-.102.073-.101.074-.102.074-.1.075-.1.076-.1.076-.198.153-.196.156-.194.159-.193.16-.19.163-.188.165-.187.168-.184.169-.182.172-.18.174-.178.176-.176.178-.174.18-.172.182-.169.184-.168.187-.165.188-.163.19-.16.192-.159.195-.156.196-.153.198-.076.1-.076.1-.075.1-.074.101-.073.102-.074.102-.072.103-.072.103-.071.103-.071.104-.07.104-.069.105-.069.105-.068.106-.068.106-.067.106-.066.107-.066.107-.065.108-.064.108-.064.109-.063.109-.063.109-.061.11-.062.11-.06.111-.06.111-.059.111-.059.112-.058.112-.057.113-.057.113-.056.114-.055.113-.055.115-.054.114-.053.115-.053.115-.052.116-.051.116-.05.117-.05.116-.049.118-.049.117-.048.118-.047.118-.046.119-.046.119-.045.119-.044.12-.043.12-.043.12-.042.121-.042.121-.04.121-.04.122-.04.121-.038.123-.038.122-.037.123-.036.124-.036.123-.035.124-.034.124-.033.125-.032.125-.032.125-.031.125-.031.126-.029.126-.029.126-.028.127-.027.126-.027.128-.025.127-.025.128-.024.128-.024.128-.022.128-.022.129-.021.129-.02.129-.02.13-.018.13-.018.13-.017.13-.016.131-.016.131-.014.131-.014.131-.013.132-.012.131-.011.132-.011.133-.01.132-.008.133-.008.133-.008.133-.006.133-.005.134-.005.134-.004.134-.003.134-.002.134-.001.135-.001.135v208.924l.002.27.002.134.003.134.004.134.01.268.014.266.016.266.021.265.023.263.027.263.03.262.033.261.036.26.04.259.043.258.046.256.049.256.052.255.055.253.058.252.062.251.064.25.067.249.071.247.073.247.076.245.08.243.082.242.085.241.087.24.091.238.093.237.097.235.099.234.101.233.105.231.107.229.11.228.113.227.115.225.118.223.12.222.123.22.126.218.128.217.131.215.133.213.136.212.138.21.141.208.143.206.146.205.147.203.151.2.076.1.153.198.156.196.159.195.16.192.163.19.165.188.168.187.169.184.172.182.174.18.176.178.178.176.18.174.182.172.184.169.187.168.188.165.19.163.193.16.194.159.196.156.198.153.2.152.202.149.203.147.206.144.207.142.209.139.211.137.212.135.214.132.216.129.218.127.219.124.221.122.222.119.225.117.225.114.227.111.229.109.23.106.232.103.233.1.235.098.236.095.238.092.239.089.24.086.242.084.243.08.244.078.245.075.247.072.248.069.25.065.25.063.252.06.253.057.254.054.255.05.256.048.257.044.258.041.26.038.26.035.262.032.262.028.263.025.265.022.265.018.266.016.267.011.134.005.134.004.134.003.134.002.27.002h259.928l.135-.001.135-.001.134-.002.134-.003.134-.004.134-.005.134-.005.133-.006.133-.008.133-.008.133-.008.132-.01.133-.011.132-.011.131-.012.132-.013.131-.014.131-.014.131-.016.131-.016.13-.017.13-.018.13-.018.13-.02.129-.02.129-.021.129-.022.128-.022.128-.024.128-.024.128-.025.127-.025.128-.027.126-.027.127-.028.126-.029.126-.029.126-.031.125-.031.125-.032.125-.032.125-.033.124-.034.124-.035.123-.036.124-.036.123-.037.122-.038.122-.038.122-.04.122-.039.121-.041.121-.042.121-.042.12-.042.12-.044.12-.044.119-.045.119-.046.119-.046.118-.047.118-.048.117-.049.118-.049.116-.05.117-.05.116-.051.116-.052.115-.053.115-.053.114-.054.115-.055.113-.055.114-.056.113-.057.112-.057.113-.058.112-.059.111-.059.111-.06.111-.06.11-.062.11-.061.109-.063.109-.063.109-.064.108-.064.108-.065.107-.066.107-.066.106-.067.106-.068.106-.068.105-.069.105-.069.104-.07.104-.071.103-.071.103-.072.103-.072.102-.073.101-.074.102-.074.1-.075.1-.076.1-.076.198-.153.196-.156.194-.159.193-.16.19-.163.188-.165.187-.168.184-.169.182-.172.18-.174.178-.176.176-.178.174-.18.172-.182.169-.184.168-.187.165-.188.163-.19.16-.192.159-.195.156-.196.153-.198.076-.1.076-.1.075-.1.074-.101.073-.102.074-.102.072-.103.072-.103.071-.103.071-.104.07-.104.069-.105.069-.105.068-.106.068-.106.067-.106.066-.107.066-.107.065-.108.064-.108.064-.109.063-.109.063-.109.061-.11.062-.11.06-.111.06-.111.059-.111.059-.112.058-.112.057-.113.057-.113.056-.114.055-.113.055-.115.054-.114.053-.115.053-.115.052-.116.051-.116.05-.117.05-.116.049-.118.049-.117.048-.118.047-.118.046-.119.046-.119.045-.119.044-.12.043-.12.043-.12.042-.121.042-.121.04-.121.04-.122.04-.121.038-.123.038-.122.037-.123.036-.124.036-.123.035-.124.034-.124.033-.125.032-.125.032-.125.031-.125.031-.126.029-.126.029-.126.028-.127.027-.126.027-.128.025-.127.025-.128.024-.128.024-.128.022-.128.022-.129.021-.129.02-.129.02-.13.018-.13.018-.13.017-.13.016-.131.016-.131.014-.131.014-.131.013-.132.012-.131.011-.132.011-.133.01-.132.008-.133.008-.133.008-.133.006-.133.005-.134.005-.134.004-.134.003-.134.002-.134.001-.135.001-.135V132.869l-.001-.135-.001-.135-.002-.134-.003-.134-.004-.134-.005-.134-.005-.134-.006-.133-.008-.133-.008-.133-.008-.133-.01-.132-.011-.133-.011-.132-.012-.131-.013-.132-.014-.131-.014-.131-.016-.131-.016-.131-.017-.13-.018-.13-.018-.13-.02-.13-.02-.129-.021-.129-.022-.129-.022-.128-.024-.128-.024-.128-.025-.128-.025-.127-.027-.128-.027-.126-.028-.127-.029-.126-.029-.126-.031-.126-.031-.125-.032-.125-.032-.125-.033-.125-.034-.124-.035-.124-.036-.123-.036-.124-.037-.123-.038-.122-.038-.123-.04-.121-.04-.122-.04-.121-.042-.121-.042-.121-.043-.12-.043-.12-.044-.12-.045-.119-.046-.119-.046-.119-.047-.118-.048-.118-.049-.117-.049-.118-.05-.116-.05-.117-.051-.116-.052-.116-.053-.115-.053-.115-.054-.114-.055-.115-.055-.113-.056-.114-.057-.113-.057-.113-.058-.112-.059-.112-.059-.111-.06-.111-.06-.111-.062-.11-.061-.11-.063-.109-.063-.109-.064-.109-.064-.108-.065-.108-.066-.107-.066-.107-.067-.106-.068-.106-.068-.106-.069-.105-.069-.105-.07-.104-.071-.104-.071-.103-.072-.103-.072-.103-.074-.102-.073-.102-.074-.101-.075-.1-.076-.1-.076-.1-.153-.198-.156-.196-.159-.195-.16-.192-.163-.19-.165-.188-.168-.187-.169-.184-.172-.182-.174-.18-.176-.178-.178-.176-.18-.174-.182-.172-.184-.169-.187-.168-.188-.165-.19-.163-.193-.16-.194-.159-.196-.156-.198-.153-.1-.076-.1-.076-.1-.075-.102-.074-.101-.074-.102-.073-.103-.072-.103-.072-.103-.071-.104-.071-.104-.07-.105-.069-.105-.069-.106-.068-.106-.068-.106-.067-.107-.066-.107-.066-.108-.065-.108-.064-.109-.064-.109-.063-.109-.063-.11-.061-.11-.062-.111-.06-.111-.06-.111-.059-.112-.059-.113-.058-.112-.057-.113-.057-.114-.056-.113-.055-.115-.055-.114-.054-.115-.053-.115-.053-.116-.052-.116-.051-.117-.05-.116-.05-.118-.049-.117-.049-.118-.048-.118-.047-.119-.046-.119-.046-.119-.045-.12-.044-.12-.044-.12-.042-.121-.042-.121-.042-.121-.041-.122-.039-.122-.04-.122-.038-.122-.038-.123-.037-.124-.036-.123-.036-.124-.035-.124-.034-.125-.033-.125-.032-.125-.032-.125-.031-.126-.031-.126-.029-.126-.029-.127-.028-.126-.027-.128-.027-.127-.025-.128-.025-.128-.024-.128-.024-.128-.022-.129-.022-.129-.021-.129-.02-.13-.02-.13-.018-.13-.018-.13-.017-.131-.016-.131-.016-.131-.014-.131-.014-.132-.013-.131-.012-.132-.011-.133-.011-.132-.01-.133-.008-.133-.008-.133-.008-.133-.006-.134-.005-.134-.005-.134-.004-.134-.003-.134-.002-.135-.001-.135-.001H70.036z"
-></path>
-<circle cx="200" cy="108.176" r="70.019" fill="#fff"></circle>
-<path
-  fill="#eee"
-  d="M200.453 38.158l.451.005.452.007.45.01.45.013.449.015.448.019.448.021.447.024.446.027.445.03.445.032.444.036.443.038.442.04.442.044.44.046.44.049.439.052.439.054.437.057.437.06.435.062.435.065.434.068.433.071.432.073.432.075.43.079.43.081.428.083.428.087.427.088.425.092.425.094.424.096.423.099.421.102.421.104.42.107.419.109.418.112.417.114.415.117.415.119.414.122.412.124.412.127.41.129.41.132.408.134.407.137.406.138.405.142.404.144.403.146.401.148.4.151.4.153.398.156.396.158.396.16.394.163.394.165.392.167.39.17.39.172.388.174.387.177.386.179.385.181.383.183.382.186.38.188.38.19.378.192.377.194.375.197.374.199.373.201.371.203.37.206.369.208.367.209.366.212.364.214.363.217.362.218.36.22.359.223.357.225.356.226.354.229.353.231.352.233.35.235.348.237.347.239.346.241.344.243.342.246.341.247.34.249.337.251.337.253.334.255.333.257.332.259.33.261.328.263.327.265.325.266.323.269.322.27.32.273.319.274.317.276.315.278.313.28.312.282.31.283.308.286.307.287.305.289.303.291.302.292.299.295.298.296.296.298.295.299.292.302.291.303.289.305.287.307.286.308.283.31.282.312.28.313.278.315.276.317.274.319.273.32.27.322.269.323.266.325.265.327.263.328.261.33.259.332.257.333.255.334.253.337.251.337.249.34.248.341.245.342.243.344.241.346.239.347.237.348.235.35.233.352.231.353.229.354.226.356.225.357.223.359.22.36.218.362.217.363.214.364.212.366.209.367.208.369.206.37.203.371.201.373.199.374.197.375.194.377.192.378.191.38.187.38.186.382.183.383.181.385.179.386.177.387.174.388.172.39.17.39.167.392.165.394.163.394.16.396.158.396.156.398.153.4.151.4.148.402.147.402.143.404.142.405.139.406.136.407.134.408.132.41.129.41.127.412.124.412.122.414.119.415.117.415.114.417.112.418.109.419.107.42.104.421.102.421.099.423.096.424.094.425.092.426.088.426.087.428.083.428.081.43.079.43.075.432.073.432.071.433.068.434.065.435.062.435.06.437.057.437.054.439.052.439.049.44.046.44.044.442.04.442.038.443.036.444.032.445.03.445.027.446.024.447.021.448.019.448.015.449.013.45.01.45.007.452.005.451.001.453-.001.453-.005.451-.007.452-.01.45-.013.45-.015.449-.019.448-.021.448-.024.447-.027.446-.03.445-.032.445-.036.444-.038.443-.04.442-.044.442-.046.44-.049.44-.052.439-.054.439-.057.437-.06.437-.062.435-.065.435-.068.434-.071.433-.073.433-.075.431-.079.43-.081.43-.083.428-.087.428-.088.427-.092.425-.094.425-.096.424-.099.423-.102.422-.104.42-.107.42-.109.419-.112.418-.114.417-.117.415-.119.415-.122.414-.124.412-.127.412-.129.41-.132.41-.134.408-.136.407-.139.406-.142.405-.143.404-.147.403-.148.401-.151.4-.153.4-.156.398-.158.396-.16.396-.163.394-.165.394-.167.392-.17.39-.172.39-.174.388-.177.387-.179.386-.181.385-.183.383-.186.382-.187.38-.191.38-.192.378-.194.377-.197.375-.199.374-.201.373-.203.371-.206.37-.208.369-.209.367-.212.366-.214.364-.217.363-.218.362-.22.36-.223.359-.225.357-.226.356-.229.355-.231.352-.233.352-.235.35-.237.348-.239.347-.241.346-.243.344-.245.342-.248.341-.249.34-.251.337-.253.337-.255.334-.257.333-.259.332-.261.33-.263.328-.265.327-.266.325-.269.323-.27.322-.273.32-.274.319-.276.317-.278.315-.28.313-.282.312-.283.31-.286.308-.287.307-.289.305-.291.303-.292.302-.295.299-.296.298-.298.296-.299.295-.302.292-.303.291-.305.289-.307.287-.308.286-.31.283-.312.282-.313.28-.315.278-.317.276-.319.274-.32.273-.322.27-.323.269-.325.266-.327.265-.328.263-.33.261-.332.259-.333.257-.334.255-.337.253-.337.251-.34.249-.341.248-.342.245-.344.243-.346.241-.347.239-.348.237-.35.235-.352.233-.353.231-.354.229-.356.226-.357.225-.359.223-.36.22-.362.218-.363.217-.364.214-.366.212-.367.209-.369.208-.37.206-.371.203-.373.201-.374.199-.375.197-.377.194-.378.192-.38.191-.38.187-.382.186-.383.183-.385.181-.386.179-.387.177-.388.174-.39.172-.39.17-.392.167-.394.165-.394.163-.396.16-.396.158-.398.156-.4.153-.4.151-.401.148-.403.147-.404.143-.405.142-.406.139-.407.136-.408.134-.41.132-.41.129-.412.127-.412.124-.414.122-.415.119-.415.117-.417.114-.418.112-.419.109-.42.107-.421.104-.421.102-.423.099-.424.096-.425.094-.425.092-.427.088-.428.087-.428.083-.43.081-.43.079-.432.075-.432.073-.433.071-.434.068-.435.065-.435.062-.437.06-.437.057-.439.054-.439.052-.44.049-.44.046-.442.044-.442.04-.443.038-.444.036-.445.032-.445.03-.446.027-.447.024-.448.021-.448.019-.449.015-.45.013-.45.01-.452.007-.451.005-.453.001-.453-.001-.451-.005-.452-.007-.45-.01-.45-.013-.449-.015-.448-.019-.448-.021-.447-.024-.446-.027-.445-.03-.445-.032-.444-.036-.443-.038-.442-.04-.442-.044-.44-.046-.44-.049-.439-.052-.439-.054-.437-.057-.437-.06-.435-.062-.435-.065-.434-.068-.433-.071-.432-.073-.432-.075-.43-.079-.43-.081-.428-.083-.428-.087-.427-.088-.425-.092-.425-.094-.424-.096-.423-.099-.421-.102-.421-.104-.42-.107-.419-.109-.418-.112-.417-.114-.415-.117-.415-.119-.414-.122-.412-.124-.412-.127-.41-.129-.41-.132-.408-.134-.407-.136-.406-.139-.405-.142-.404-.143-.403-.147-.401-.148-.4-.151-.4-.153-.398-.156-.396-.158-.396-.16-.394-.163-.394-.165-.392-.167-.39-.17-.39-.172-.388-.174-.387-.177-.386-.179-.385-.181-.383-.183-.382-.186-.38-.187-.38-.191-.378-.192-.377-.194-.375-.197-.374-.199-.373-.201-.371-.203-.37-.206-.369-.208-.367-.209-.366-.212-.364-.214-.363-.217-.362-.218-.36-.22-.359-.223-.357-.225-.356-.226-.354-.229-.353-.231-.352-.233-.35-.235-.348-.237-.347-.239-.346-.241-.344-.243-.342-.245-.341-.248-.34-.249-.337-.251-.337-.253-.334-.255-.333-.257-.332-.259-.33-.261-.328-.263-.327-.265-.325-.266-.323-.269-.322-.27-.32-.273-.319-.274-.317-.276-.315-.278-.313-.28-.312-.282-.31-.283-.308-.286-.307-.287-.305-.289-.303-.291-.302-.292-.299-.295-.298-.296-.296-.298-.295-.299-.292-.302-.291-.303-.289-.305-.287-.307-.286-.308-.283-.31-.282-.312-.28-.313-.278-.315-.276-.317-.274-.319-.273-.32-.27-.322-.269-.323-.266-.325-.265-.327-.263-.328-.261-.33-.259-.332-.257-.333-.255-.334-.253-.337-.251-.337-.249-.34-.248-.341-.245-.342-.243-.344-.241-.346-.239-.347-.237-.348-.235-.35-.233-.352-.231-.352-.229-.355-.226-.356-.225-.357-.223-.359-.22-.36-.218-.362-.217-.363-.214-.364-.212-.366-.209-.367-.208-.369-.206-.37-.203-.371-.201-.373-.199-.374-.197-.375-.194-.377-.192-.378-.191-.38-.187-.38-.186-.382-.183-.383-.181-.385-.179-.386-.177-.387-.174-.388-.172-.39-.17-.39-.167-.392-.165-.394-.163-.394-.16-.396-.158-.396-.156-.398-.153-.4-.151-.4-.148-.401-.147-.403-.143-.404-.142-.405-.139-.406-.136-.407-.134-.408-.132-.41-.129-.41-.127-.412-.124-.412-.122-.414-.119-.415-.117-.415-.114-.417-.112-.418-.109-.419-.107-.42-.104-.42-.102-.422-.099-.423-.096-.424-.094-.425-.092-.425-.088-.427-.087-.428-.083-.428-.081-.43-.079-.43-.075-.431-.073-.433-.071-.433-.068-.434-.065-.435-.062-.435-.06-.437-.057-.437-.054-.439-.052-.439-.049-.44-.046-.44-.044-.442-.04-.442-.038-.443-.036-.444-.032-.445-.03-.445-.027-.446-.024-.447-.021-.448-.019-.448-.015-.449-.013-.45-.01-.45-.007-.452-.005-.451-.001-.453.001-.453.005-.451.007-.452.01-.45.013-.45.015-.449.019-.448.021-.448.024-.447.027-.446.03-.445.032-.445.036-.444.038-.443.04-.442.044-.442.046-.44.049-.44.052-.439.054-.439.057-.437.06-.437.062-.435.065-.435.068-.434.071-.433.073-.432.075-.432.079-.43.081-.43.083-.428.087-.428.088-.426.092-.426.094-.425.096-.424.099-.423.102-.421.104-.421.107-.42.109-.419.112-.418.114-.417.117-.415.119-.415.122-.414.124-.412.127-.412.129-.41.132-.41.134-.408.136-.407.139-.406.142-.405.143-.404.147-.402.148-.402.151-.4.153-.4.156-.398.158-.396.16-.396.163-.394.165-.394.167-.392.17-.39.172-.39.174-.388.177-.387.179-.386.181-.385.183-.383.186-.382.187-.38.191-.38.192-.378.194-.377.197-.375.199-.374.201-.373.203-.371.206-.37.208-.369.209-.367.212-.366.214-.364.217-.363.218-.362.22-.36.223-.359.225-.357.226-.356.229-.354.231-.353.233-.352.235-.35.237-.348.239-.347.241-.346.243-.344.245-.342.248-.341.249-.34.251-.337.253-.337.255-.334.257-.333.259-.332.261-.33.263-.328.265-.327.266-.325.269-.323.27-.322.273-.32.274-.319.276-.317.278-.315.28-.313.282-.312.283-.31.286-.308.287-.307.289-.305.291-.303.292-.302.295-.299.296-.298.298-.296.299-.295.302-.292.303-.291.305-.289.307-.287.308-.286.31-.283.312-.282.313-.28.315-.278.317-.276.319-.274.32-.273.322-.27.323-.269.325-.266.327-.265.328-.263.33-.261.332-.259.333-.257.334-.255.337-.253.337-.251.34-.249.341-.247.342-.246.344-.243.346-.241.347-.239.348-.237.35-.235.352-.233.353-.231.354-.229.356-.226.357-.225.359-.223.36-.22.362-.218.363-.217.364-.214.366-.212.367-.209.369-.208.37-.206.371-.203.373-.201.374-.199.375-.197.377-.194.378-.192.38-.19.38-.188.382-.186.383-.183.385-.181.386-.179.387-.177.388-.174.39-.172.39-.17.392-.167.394-.165.394-.163.396-.16.396-.158.398-.156.4-.153.4-.151.401-.148.403-.146.404-.144.405-.142.406-.138.407-.137.408-.134.41-.132.41-.129.412-.127.412-.124.414-.122.415-.119.415-.117.417-.114.418-.112.419-.109.42-.107.421-.104.421-.102.423-.099.424-.096.425-.094.425-.092.427-.088.428-.087.428-.083.43-.081.43-.079.432-.075.432-.073.433-.071.434-.068.435-.065.435-.062.437-.06.437-.057.439-.054.439-.052.44-.049.44-.046.442-.044.442-.04.443-.038.444-.036.445-.032.445-.03.446-.027.447-.024.448-.021.448-.019.449-.015.45-.013.45-.01.452-.007.451-.005.453-.001.453.001zm-.864 6.5l-.411.004-.41.007-.409.009-.409.012-.408.014-.407.017-.406.019-.406.022-.405.024-.405.027-.403.03-.403.032-.402.034-.402.037-.4.039-.4.042-.399.045-.399.047-.397.049-.397.052-.396.054-.395.057-.394.059-.394.061-.393.064-.391.066-.392.069-.39.071-.389.073-.389.076-.387.078-.387.081-.386.083-.385.085-.384.087-.384.09-.382.092-.381.095-.381.097-.38.099-.378.101-.378.104-.377.105-.376.109-.375.11-.374.113-.373.114-.372.117-.371.12-.37.121-.369.124-.368.126-.367.128-.366.13-.365.133-.364.134-.363.137-.362.139-.361.141-.359.144-.359.145-.358.147-.356.15-.355.152-.355.154-.353.155-.352.158-.351.16-.35.163-.348.164-.348.166-.346.168-.345.171-.344.172-.343.174-.342.177-.34.178-.34.181-.338.182-.337.185-.335.186-.334.188-.334.191-.332.192-.33.194-.329.196-.329.198-.326.2-.326.202-.324.204-.323.206-.321.207-.321.21-.319.211-.317.213-.316.216-.315.217-.314.218-.312.221-.311.222-.309.225-.308.226-.307.228-.305.23-.304.231-.302.233-.301.235-.299.237-.298.239-.297.24-.295.242-.294.244-.292.246-.29.247-.29.249-.287.251-.286.252-.285.254-.283.256-.281.257-.28.259-.279.261-.277.262-.275.264-.273.266-.272.267-.271.269-.269.271-.267.272-.266.273-.264.275-.262.277-.261.279-.259.28-.257.281-.256.283-.254.285-.252.286-.251.287-.249.29-.247.29-.246.292-.244.294-.242.295-.24.297-.239.298-.237.299-.235.301-.233.302-.232.304-.229.305-.228.307-.226.308-.225.309-.222.311-.221.312-.219.314-.216.314-.216.317-.213.317-.211.319-.21.321-.207.321-.206.323-.204.324-.202.326-.2.327-.198.328-.196.329-.194.331-.192.331-.191.334-.188.334-.186.335-.185.337-.182.338-.181.34-.178.34-.177.342-.174.343-.172.344-.171.345-.168.346-.166.348-.164.348-.163.35-.16.351-.158.352-.156.353-.153.355-.152.355-.15.357-.147.357-.145.359-.144.36-.141.36-.139.362-.137.363-.134.364-.133.365-.13.366-.128.367-.126.368-.124.369-.121.37-.12.371-.117.372-.114.373-.113.374-.11.375-.109.376-.105.377-.104.378-.101.378-.099.38-.097.381-.095.381-.092.382-.09.384-.087.384-.085.385-.083.386-.081.387-.078.387-.076.389-.073.389-.071.39-.069.391-.066.392-.064.393-.061.394-.059.394-.057.395-.054.396-.052.397-.049.397-.047.399-.045.399-.042.4-.039.4-.037.402-.034.402-.032.403-.03.403-.027.405-.024.405-.022.406-.019.406-.017.407-.015.408-.011.409-.009.409-.007.41-.004.411-.001.411.001.411.004.411.007.41.009.409.011.409.015.408.017.407.019.406.022.406.024.405.027.405.03.403.032.403.034.402.037.402.039.4.042.4.045.399.047.399.049.397.052.397.054.396.057.395.059.394.061.394.064.393.066.392.069.391.071.39.073.389.076.389.078.387.081.387.083.386.085.385.087.384.09.384.092.382.095.382.097.38.099.38.101.378.104.378.105.377.109.376.11.375.113.374.114.373.117.372.12.371.121.37.124.369.126.368.128.367.13.366.133.365.134.364.137.363.139.362.141.361.144.359.145.359.147.358.15.356.152.355.153.355.156.353.158.352.16.351.163.35.164.348.166.348.168.346.171.346.172.343.174.343.177.342.178.34.181.34.182.338.185.337.186.335.188.334.191.334.192.332.194.33.196.329.198.329.2.326.202.326.204.324.206.323.207.322.21.32.211.319.213.317.216.317.216.314.219.314.221.312.222.311.225.309.226.308.228.307.23.305.231.304.233.302.235.301.237.299.239.299.24.296.242.295.244.294.246.292.247.29.249.29.251.287.252.286.254.285.256.283.257.281.259.28.261.279.262.277.264.275.266.273.267.273.269.27.271.269.272.267.273.266.275.264.277.262.279.261.28.259.281.257.283.256.285.254.286.253.287.25.29.249.29.247.292.246.294.244.295.242.297.24.298.239.299.237.301.235.302.233.304.232.305.229.307.228.308.226.309.225.311.222.312.221.314.219.314.216.317.216.317.213.319.211.32.21.322.207.323.206.324.204.326.202.326.2.329.198.329.196.33.194.332.193.334.19.334.188.335.186.337.185.338.182.34.181.34.178.342.177.343.174.343.172.346.171.346.168.348.166.348.164.35.163.351.16.352.158.353.156.355.153.355.152.356.15.358.147.359.145.36.144.36.141.362.139.363.137.364.134.365.133.366.13.367.128.368.126.369.124.37.121.371.12.372.117.373.114.374.113.375.11.376.109.377.105.378.104.378.101.38.099.38.097.382.095.382.092.384.09.384.087.385.085.386.083.387.081.387.078.389.076.389.073.39.071.391.069.392.066.393.064.394.061.394.059.395.057.396.054.397.052.397.049.399.047.399.045.4.042.4.039.402.037.402.034.403.032.403.03.405.027.405.024.406.022.406.02.407.016.408.015.409.011.409.009.41.007.411.004.411.001.411-.001.411-.004.41-.007.409-.009.409-.011.408-.015.407-.016.406-.02.406-.022.405-.024.405-.027.403-.03.403-.032.402-.034.402-.037.4-.039.4-.042.399-.045.399-.047.397-.049.397-.052.396-.054.395-.057.394-.059.394-.061.393-.064.392-.066.391-.069.39-.071.389-.073.389-.076.387-.078.387-.081.386-.083.385-.085.384-.087.384-.09.382-.092.382-.095.38-.097.38-.099.378-.101.378-.104.377-.105.376-.109.375-.11.374-.113.373-.114.372-.117.371-.12.37-.121.369-.124.368-.126.367-.128.366-.13.365-.133.364-.134.363-.137.362-.139.36-.141.36-.144.359-.145.357-.147.357-.15.355-.152.355-.153.353-.156.352-.158.351-.16.35-.163.348-.164.348-.166.346-.168.345-.171.344-.172.343-.174.342-.177.34-.178.34-.181.338-.182.337-.184.335-.187.334-.188.334-.191.332-.192.33-.194.329-.196.328-.198.327-.2.326-.202.324-.204.323-.206.322-.207.32-.21.319-.211.317-.213.317-.216.314-.216.314-.219.312-.221.311-.222.309-.225.308-.226.307-.228.305-.229.304-.232.302-.233.301-.235.299-.237.299-.239.296-.24.295-.242.294-.244.292-.246.29-.247.29-.249.287-.25.286-.253.285-.254.283-.256.281-.257.28-.259.279-.261.277-.262.275-.264.273-.266.272-.267.271-.269.269-.271.267-.272.266-.273.264-.276.262-.276.261-.279.259-.28.257-.281.256-.283.254-.285.252-.286.251-.288.249-.289.247-.29.246-.292.244-.294.242-.295.24-.297.239-.297.237-.3.235-.301.233-.302.231-.304.23-.305.228-.307.226-.308.224-.309.223-.311.221-.312.218-.314.218-.315.215-.316.213-.317.211-.319.21-.32.207-.322.206-.323.204-.324.202-.326.2-.326.198-.329.196-.329.194-.33.192-.332.191-.334.188-.334.186-.335.185-.337.182-.338.181-.34.178-.34.176-.342.175-.343.172-.344.171-.345.168-.346.166-.348.164-.348.163-.35.16-.351.158-.352.155-.353.154-.355.152-.355.15-.357.147-.357.145-.359.144-.359.141-.361.139-.362.137-.363.134-.364.133-.365.13-.366.128-.367.126-.368.124-.369.121-.37.12-.371.117-.372.114-.373.113-.374.11-.375.109-.376.105-.377.104-.378.101-.378.099-.38.097-.38.095-.382.092-.382.09-.383.087-.385.085-.385.083-.386.081-.387.078-.387.076-.389.073-.389.071-.39.069-.391.066-.392.064-.393.061-.394.059-.394.057-.395.054-.396.052-.397.049-.397.047-.399.045-.399.042-.4.039-.4.037-.402.034-.402.032-.403.03-.403.027-.405.024-.405.022-.406.019-.406.017-.407.015-.408.011-.409.009-.409.007-.41.004-.411.001-.411-.001-.411-.004-.411-.007-.41-.009-.409-.011-.409-.015-.408-.017-.407-.019-.406-.022-.406-.024-.405-.027-.405-.03-.403-.032-.403-.034-.402-.037-.402-.039-.4-.042-.4-.045-.399-.047-.398-.049-.398-.052-.397-.054-.396-.057-.395-.059-.394-.061-.394-.064-.393-.066-.392-.069-.391-.071-.39-.073-.389-.076-.389-.078-.387-.081-.387-.083-.386-.085-.385-.087-.385-.09-.383-.092-.382-.095-.381-.097-.381-.099-.38-.101-.378-.104-.378-.105-.377-.109-.376-.11-.375-.113-.374-.114-.373-.117-.372-.12-.371-.121-.37-.124-.369-.126-.368-.128-.367-.13-.366-.133-.365-.134-.364-.137-.363-.139-.362-.141-.361-.144-.359-.145-.359-.147-.357-.15-.357-.152-.355-.154-.355-.155-.353-.158-.352-.16-.351-.163-.35-.164-.348-.166-.348-.168-.346-.171-.345-.172-.344-.175-.343-.176-.342-.178-.34-.181-.34-.182-.338-.185-.337-.186-.335-.188-.334-.191-.334-.192-.331-.194-.331-.196-.329-.198-.328-.2-.327-.202-.326-.204-.324-.206-.323-.207-.322-.21-.32-.211-.319-.213-.317-.215-.316-.218-.315-.218-.314-.221-.312-.223-.311-.224-.309-.226-.308-.228-.307-.23-.305-.231-.304-.233-.302-.235-.301-.237-.3-.239-.297-.24-.297-.242-.295-.244-.294-.246-.292-.247-.29-.249-.289-.251-.288-.252-.286-.254-.285-.256-.283-.257-.281-.259-.28-.261-.279-.262-.276-.264-.276-.266-.273-.267-.272-.269-.271-.271-.269-.272-.267-.273-.266-.275-.264-.277-.262-.279-.261-.28-.259-.281-.257-.283-.256-.285-.254-.286-.252-.287-.251-.29-.249-.29-.247-.292-.246-.294-.244-.295-.242-.296-.24-.299-.239-.299-.237-.301-.235-.302-.233-.304-.231-.305-.23-.307-.228-.308-.226-.309-.225-.311-.222-.312-.221-.314-.218-.314-.217-.317-.216-.317-.213-.319-.211-.32-.21-.322-.207-.323-.206-.324-.204-.326-.202-.326-.2-.329-.198-.329-.196-.33-.194-.332-.192-.334-.191-.334-.188-.336-.187-.336-.184-.338-.182-.34-.181-.34-.178-.342-.177-.343-.174-.344-.172-.345-.171-.346-.168-.348-.166-.348-.164-.35-.163-.351-.16-.352-.158-.353-.155-.355-.154-.355-.152-.357-.15-.357-.147-.359-.145-.359-.144-.361-.141-.362-.139-.363-.137-.364-.134-.365-.133-.366-.13-.367-.128-.368-.126-.369-.124-.37-.121-.371-.12-.372-.117-.373-.114-.374-.113-.375-.11-.376-.109-.377-.105-.378-.104-.378-.101-.38-.099-.381-.097-.381-.095-.382-.092-.384-.09-.384-.087-.385-.085-.386-.083-.387-.081-.387-.078-.389-.076-.389-.073-.39-.071-.392-.069-.391-.066-.393-.064-.394-.061-.394-.059-.395-.057-.396-.054-.397-.052-.398-.049-.398-.047-.399-.045-.4-.042-.4-.039-.402-.037-.402-.034-.403-.032-.403-.03-.405-.027-.405-.024-.406-.022-.406-.019-.407-.017-.408-.014-.409-.012-.409-.009-.41-.007-.411-.004-.411-.001-.411.001z"
-></path>
-<path
-  fill="#565656"
-  d="M289.817 202.686a7.59 7.59 0 00-7.585-7.586H117.768a7.59 7.59 0 00-7.585 7.586v3.834a7.59 7.59 0 007.585 7.586h164.464a7.59 7.59 0 007.585-7.586v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M245.908 331.685a5.593 5.593 0 00-5.59-5.59h-80.636a5.593 5.593 0 00-5.59 5.59v2.813a5.593 5.593 0 005.59 5.59h80.636a5.593 5.593 0 005.59-5.59v-2.813zM306.311 281.891a5.686 5.686 0 00-5.684-5.683H99.373a5.686 5.686 0 00-5.684 5.683v2.674a5.687 5.687 0 005.684 5.684h201.254a5.687 5.687 0 005.684-5.684v-2.674zM326.619 259.129a5.686 5.686 0 00-5.683-5.683H72.341a5.686 5.686 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.675zM326.619 236.368a5.687 5.687 0 00-5.683-5.684H72.341a5.687 5.687 0 00-5.683 5.684v2.674a5.686 5.686 0 005.683 5.683h248.595a5.686 5.686 0 005.683-5.683v-2.674z"
-></path>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M176.939 83.289h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719h-46.122a3.72 3.72 0 01-3.719-3.719V87.009a3.721 3.721 0 013.719-3.72zm43.89 47.609V89.24h-41.658v41.658h41.658zM200 98.167c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976h-29.756a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-</svg>;
-icons.infoLeftOverlay = <svg
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 400 400"
->
-<path
-  fill="#fff"
-  stroke="#eee"
-  strokeWidth="5"
-  d="M390.993 131.261c0-7.864-5.144-14.248-11.479-14.248H20.486c-6.335 0-11.479 6.384-11.479 14.248v177.518c0 7.865 5.144 14.249 11.479 14.249h359.028c6.335 0 11.479-6.384 11.479-14.249V131.261z"
-></path>
-<path
-  fill="#fff"
-  d="M151.446 72.879c0-7.396-6.005-13.401-13.401-13.401H42.752c-7.396 0-13.4 6.005-13.4 13.401v88.334c0 7.395 6.004 13.4 13.4 13.4h95.293c7.396 0 13.401-6.005 13.401-13.4V72.879z"
-></path>
-<path
-  fill="#eee"
-  d="M138.219 59.48l.172.003.172.005.172.008.171.01.17.012.17.014.169.016.168.018.168.02.167.023.167.024.166.026.165.029.165.03.164.033.163.034.162.037.162.038.161.041.16.042.159.044.159.046.157.048.157.05.156.052.156.054.154.055.153.057.153.06.152.06.15.063.15.065.149.066.148.068.147.07.146.071.145.073.144.075.143.077.142.078.14.08.14.081.139.084.137.084.137.087.135.088.135.089.133.091.132.093.13.094.13.096.128.097.128.099.126.1.124.102.124.103.122.105.121.106.12.107.118.109.117.11.116.112.114.113.113.115.112.115.11.117.109.119.107.119.106.121.105.122.103.124.102.125.1.126.099.127.097.128.096.13.094.131.093.132.091.133.089.134.088.135.087.137.085.137.083.139.081.14.08.141.078.142.077.142.075.144.073.145.071.146.07.147.068.148.067.149.064.15.063.151.061.151.059.153.057.153.056.155.053.155.052.156.05.157.048.157.046.159.044.159.043.16.04.161.038.162.037.162.034.164.033.163.03.165.029.165.026.166.025.167.022.167.02.168.019.168.016.17.014.169.012.171.009.171.008.171.005.172.004.173.001.173v88.334l-.001.173-.004.172-.005.172-.008.172-.009.171-.012.17-.014.17-.016.169-.019.168-.02.168-.022.168-.025.166-.026.166-.029.165-.03.165-.033.164-.034.163-.037.162-.038.162-.04.161-.043.16-.044.159-.046.159-.048.158-.05.156-.052.156-.053.156-.056.154-.057.153-.059.153-.061.152-.063.15-.064.15-.067.149-.068.148-.07.147-.071.146-.073.145-.075.144-.077.143-.078.142-.08.14-.081.14-.083.139-.085.137-.087.137-.088.135-.089.135-.091.133-.093.132-.094.13-.096.13-.097.128-.099.128-.1.126-.102.124-.103.124-.105.122-.106.121-.107.12-.109.118-.11.117-.112.116-.113.114-.114.113-.116.112-.117.11-.118.109-.12.107-.121.106-.122.105-.124.103-.124.102-.126.1-.128.099-.128.097-.13.096-.13.094-.132.093-.133.091-.135.09-.135.087-.137.087-.137.085-.139.083-.14.081-.14.08-.142.078-.143.077-.144.075-.145.073-.146.072-.147.069-.148.068-.149.067-.15.064-.15.063-.152.061-.153.059-.153.057-.154.056-.156.053-.156.052-.157.05-.157.048-.159.046-.159.044-.16.043-.161.04-.162.038-.162.037-.163.034-.164.033-.165.03-.165.029-.166.026-.167.025-.167.022-.168.02-.168.019-.169.016-.17.014-.17.012-.171.009-.172.008-.172.005-.172.004-.174.001H42.752l-.173-.001-.172-.004-.172-.005-.172-.008-.171-.009-.17-.012-.17-.014-.169-.016-.169-.019-.167-.02-.168-.022-.166-.025-.166-.026-.165-.029-.165-.03-.164-.033-.163-.034-.162-.037-.162-.038-.161-.04-.16-.043-.159-.044-.159-.046-.158-.048-.157-.05-.156-.052-.155-.053-.154-.056-.154-.057-.152-.059-.152-.061-.151-.063-.149-.064-.149-.067-.148-.068-.147-.069-.146-.072-.145-.073-.144-.075-.143-.077-.142-.078-.141-.08-.139-.081-.139-.083-.138-.085-.136-.087-.136-.087-.134-.09-.133-.091-.132-.093-.131-.094-.129-.096-.129-.097-.127-.099-.126-.1-.124-.102-.124-.103-.122-.105-.121-.106-.12-.107-.118-.109-.117-.11-.116-.112-.114-.113-.113-.114-.112-.116-.11-.117-.109-.118-.108-.12-.106-.121-.104-.122-.103-.124-.102-.124-.1-.126-.099-.128-.097-.128-.096-.13-.094-.13-.093-.132-.091-.133-.09-.135-.088-.135-.086-.137-.085-.137-.083-.139-.082-.14-.079-.14-.079-.142-.076-.143-.075-.144-.073-.145-.072-.146-.069-.147-.068-.148-.067-.149-.064-.15-.063-.15-.061-.152-.059-.153-.057-.153-.056-.154-.053-.156-.052-.156-.05-.156-.048-.158-.046-.159-.045-.159-.042-.16-.04-.161-.039-.162-.036-.162-.035-.163-.032-.164-.031-.165-.028-.165-.026-.166-.025-.166-.022-.168-.021-.168-.018-.168-.016-.169-.014-.17-.012-.17-.01-.171-.007-.172-.006-.172-.003-.172-.001-.173V72.879l.001-.173.003-.173.006-.172.007-.171.01-.171.012-.171.014-.169.016-.17.018-.168.021-.168.022-.167.025-.167.026-.166.028-.165.031-.165.032-.163.035-.164.036-.162.039-.162.04-.161.042-.16.045-.159.046-.159.048-.157.05-.157.052-.156.053-.155.056-.155.057-.153.059-.153.061-.151.063-.151.064-.15.067-.149.068-.148.069-.147.072-.146.073-.145.075-.144.076-.142.079-.142.079-.141.082-.14.083-.139.085-.137.086-.137.088-.135.09-.134.091-.133.093-.132.094-.131.096-.13.097-.128.099-.127.1-.126.102-.125.103-.124.104-.122.106-.121.108-.119.109-.119.11-.117.112-.115.113-.115.114-.113.116-.112.117-.11.118-.109.12-.107.121-.106.122-.105.124-.103.124-.102.126-.1.127-.099.129-.097.129-.096.131-.094.132-.093.133-.091.134-.089.136-.088.136-.087.138-.084.139-.084.139-.081.141-.08.142-.078.143-.077.144-.075.145-.073.146-.071.147-.07.148-.068.149-.066.149-.065.151-.063.152-.06.152-.06.154-.057.154-.055.155-.054.156-.052.157-.05.158-.048.159-.046.159-.044.16-.042.161-.041.162-.038.162-.037.163-.034.164-.033.165-.03.165-.029.166-.026.166-.024.168-.023.167-.02.169-.018.169-.016.17-.014.17-.012.171-.01.172-.008.172-.005.172-.003.173-.002h95.293l.174.002zm-95.452 4.498l-.131.001-.115.002-.115.004-.115.005-.114.007-.113.008-.113.009-.113.011-.112.012-.111.013-.111.015-.111.016-.11.018-.109.019-.109.02-.109.021-.108.023-.107.024-.107.026-.107.027-.106.027-.105.03-.105.03-.104.032-.104.033-.103.034-.103.036-.102.036-.101.038-.101.039-.101.041-.1.041-.099.043-.098.044-.098.045-.098.046-.096.048-.097.048-.095.05-.095.051-.094.052-.094.053-.092.054-.093.055-.091.056-.091.058-.09.058-.089.06-.088.06-.088.062-.087.063-.087.063-.085.065-.085.066-.084.067-.083.068-.082.068-.082.07-.08.071-.08.071-.079.073-.078.073-.077.075-.076.075-.076.076-.074.078-.074.078-.072.078-.072.08-.07.081-.07.081-.069.083-.068.083-.066.084-.066.084-.065.086-.064.086-.062.087-.062.088-.061.088-.059.09-.059.09-.057.09-.056.092-.056.092-.054.093-.053.093-.052.094-.051.095-.049.096-.049.096-.047.097-.046.097-.045.098-.044.099-.043.099-.042.1-.04.1-.039.101-.038.102-.037.102-.035.102-.034.104-.033.103-.032.105-.031.104-.029.106-.028.106-.027.106-.025.107-.024.108-.023.108-.022.108-.02.109-.019.11-.017.11-.016.11-.015.111-.014.112-.012.112-.01.112-.01.113-.008.114-.006.114-.005.114-.004.115-.002.116-.001.13v88.305l.001.131.002.115.004.115.005.115.006.114.008.113.01.113.01.113.012.112.014.111.015.111.016.111.017.11.019.109.02.109.022.109.023.108.024.107.025.107.027.106.028.106.029.106.031.105.032.104.033.104.034.103.035.103.037.102.038.101.039.101.04.101.042.099.043.1.044.098.045.098.046.098.047.096.049.096.049.096.051.095.052.094.053.093.054.093.056.092.056.092.057.09.059.091.059.089.061.088.062.088.062.087.064.086.065.086.066.085.066.084.068.083.069.082.07.081.07.081.072.08.072.079.074.078.074.077.076.076.076.075.077.075.078.073.079.073.08.072.08.07.082.07.082.069.083.067.084.067.085.066.085.065.087.064.087.062.087.062.089.06.089.06.09.059.091.057.091.056.093.056.092.054.094.053.094.052.095.05.095.05.096.049.097.047.098.046.098.045.098.044.1.043.099.041.101.041.101.039.101.038.102.036.103.036.103.034.104.033.104.032.105.031.105.029.106.028.107.026.107.026.107.024.108.023.109.021.109.021.109.018.11.018.111.016.111.015.111.013.112.013.113.01.113.01.113.008.114.006.115.005.115.004.115.002.131.001h95.264l.131-.001.115-.002.115-.004.115-.005.113-.006.114-.008.113-.01.113-.01.112-.013.111-.013.111-.015.111-.016.109-.018.11-.018.109-.021.109-.021.108-.023.107-.024.107-.026.107-.026.105-.028.106-.029.104-.031.105-.032.104-.033.103-.034.103-.036.102-.036.101-.038.101-.039.1-.041.1-.041.099-.043.099-.044.098-.045.097-.046.097-.047.096-.049.096-.05.095-.05.094-.052.093-.053.093-.054.092-.056.092-.056.09-.057.09-.059.09-.06.088-.06.088-.062.087-.062.086-.064.086-.065.085-.066.083-.066.084-.068.082-.069.081-.07.081-.07.08-.072.079-.073.078-.073.077-.075.076-.075.075-.076.075-.077.073-.078.073-.079.071-.08.071-.081.07-.081.069-.082.067-.084.067-.083.066-.085.065-.086.063-.086.063-.087.062-.088.06-.088.06-.09.059-.09.057-.09.056-.092.056-.092.054-.092.053-.094.052-.094.05-.095.05-.096.049-.096.047-.096.046-.098.045-.098.044-.098.043-.1.041-.099.041-.101.039-.101.038-.101.036-.102.036-.103.034-.103.033-.104.032-.104.03-.105.03-.106.028-.106.026-.106.026-.107.024-.107.023-.108.021-.109.02-.109.019-.109.018-.11.016-.111.015-.111.013-.111.012-.112.011-.113.01-.113.007-.113.007-.114.005-.115.004-.115.002-.115.001-.131V72.893l-.001-.13-.002-.116-.004-.115-.005-.114-.007-.114-.007-.114-.01-.113-.011-.112-.012-.112-.013-.112-.015-.111-.016-.11-.018-.11-.019-.11-.02-.109-.021-.108-.023-.108-.024-.108-.026-.107-.026-.106-.028-.106-.03-.106-.03-.104-.032-.105-.033-.103-.034-.104-.036-.102-.036-.102-.038-.102-.039-.101-.041-.1-.041-.1-.043-.099-.044-.099-.045-.098-.046-.097-.047-.097-.049-.096-.05-.096-.05-.095-.052-.094-.053-.093-.055-.093-.055-.092-.056-.092-.058-.09-.058-.09-.06-.09-.06-.088-.062-.088-.063-.087-.063-.086-.065-.086-.066-.084-.067-.084-.067-.083-.069-.083-.07-.081-.071-.081-.071-.08-.073-.078-.073-.078-.075-.077-.075-.077-.076-.075-.077-.075-.078-.073-.079-.073-.08-.071-.081-.071-.081-.07-.082-.068-.084-.068-.083-.067-.085-.066-.086-.065-.086-.063-.087-.063-.088-.062-.088-.06-.09-.06-.09-.058-.09-.058-.092-.056-.092-.055-.093-.054-.093-.053-.094-.052-.095-.051-.096-.05-.096-.048-.096-.048-.098-.046-.098-.045-.098-.044-.1-.043-.1-.041-.1-.041-.101-.039-.101-.038-.102-.036-.103-.036-.103-.034-.104-.033-.104-.032-.105-.03-.106-.03-.106-.027-.106-.027-.107-.026-.107-.024-.108-.023-.109-.021-.109-.02-.109-.019-.11-.018-.111-.016-.111-.015-.111-.013-.112-.012-.113-.011-.113-.009-.113-.008-.115-.007-.114-.005-.115-.004-.115-.002-.131-.001H42.767z"
-></path>
-<path
-  fill="#888"
-  fillRule="nonzero"
-  d="M67.338 90.15h46.122a3.721 3.721 0 013.719 3.72v46.121a3.72 3.72 0 01-3.719 3.719H67.338a3.72 3.72 0 01-3.719-3.719V93.87a3.721 3.721 0 013.719-3.72zm43.89 47.609V96.101H69.57v41.658h41.658zm-20.829-32.731c0-3.273-2.678-5.951-5.951-5.951-3.273 0-5.951 2.678-5.951 5.951 0 3.273 2.678 5.951 5.951 5.951 3.273 0 5.951-2.678 5.951-5.951zm8.927 14.878s0-17.854 8.926-17.854v29.756a2.984 2.984 0 01-2.975 2.976H75.521a2.984 2.984 0 01-2.975-2.976v-20.829c5.951 0 8.926 11.902 8.926 11.902s2.976-8.926 8.927-8.926 8.927 5.951 8.927 5.951z"
-></path>
-<path
-  fill="#565656"
-  d="M299.746 200.62a7.59 7.59 0 00-7.585-7.586H35.651a7.589 7.589 0 00-7.585 7.586v3.834a7.588 7.588 0 007.585 7.585h256.51a7.588 7.588 0 007.585-7.585v-3.834z"
-></path>
-<path
-  fill="#888"
-  d="M372.406 264.094a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h333.112a5.687 5.687 0 005.683-5.684v-2.674zM266.41 291.342a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.675a5.686 5.686 0 005.683 5.683h227.116a5.686 5.686 0 005.683-5.683v-2.675zM370.188 236.219a5.686 5.686 0 00-5.683-5.683H33.611a5.685 5.685 0 00-5.683 5.683v2.674a5.686 5.686 0 005.683 5.684h330.894a5.687 5.687 0 005.683-5.684v-2.674z"
-></path>
-</svg>;
-icons.unlocked = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M5 12h14c.276 0 .525.111.707.293S20 12.724 20 13v7c0 .276-.111.525-.293.707S19.276 21 19 21H5c-.276 0-.525-.111-.707-.293S4 20.276 4 20v-7c0-.276.111-.525.293-.707S4.724 12 5 12zm3-2V7a3.988 3.988 0 011.169-2.831 3.983 3.983 0 012.821-1.174 3.985 3.985 0 012.652 1 4.052 4.052 0 011.28 2.209 1 1 0 101.958-.408 6.051 6.051 0 00-1.912-3.299A5.963 5.963 0 0011.995.995c-1.657.002-3.157.676-4.241 1.762S5.998 5.344 6 7v3H5a2.997 2.997 0 00-3 3v7a2.997 2.997 0 003 3h14a2.997 2.997 0 003-3v-7a2.997 2.997 0 00-3-3z"></path>
-</svg>;
-icons.locked = <svg
-xmlns="http://www.w3.org/2000/svg"
-width="24"
-height="24"
-viewBox="0 0 24 24"
->
-<path d="M5 12h14c.276 0 .525.111.707.293S20 12.724 20 13v7c0 .276-.111.525-.293.707S19.276 21 19 21H5c-.276 0-.525-.111-.707-.293S4 20.276 4 20v-7c0-.276.111-.525.293-.707S4.724 12 5 12zm13-2V7c0-1.657-.673-3.158-1.757-4.243S13.657 1 12 1s-3.158.673-4.243 1.757S6 5.343 6 7v3H5a2.997 2.997 0 00-3 3v7a2.997 2.997 0 003 3h14a2.997 2.997 0 003-3v-7a2.997 2.997 0 00-3-3zM8 10V7c0-1.105.447-2.103 1.172-2.828S10.895 3 12 3s2.103.447 2.828 1.172S16 5.895 16 7v3z"></path>
-</svg>;
-icons.px = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-  <path d="M2.896 6.603h1.419v.92h.027c.21-.394.504-.677.88-.848a2.926 2.926 0 011.223-.256c.534 0 1.001.094 1.399.283.399.188.73.447.993.775.263.329.46.712.591 1.15.132.438.197.907.197 1.407 0 .455-.059.898-.177 1.327a3.455 3.455 0 01-.539 1.137 2.699 2.699 0 01-.913.789c-.368.197-.802.295-1.302.295-.219 0-.438-.019-.657-.059a2.855 2.855 0 01-.631-.19 2.51 2.51 0 01-.558-.336 1.823 1.823 0 01-.427-.479h-.027v3.391H2.896V6.603zm5.231 3.404c0-.306-.039-.604-.118-.894a2.418 2.418 0 00-.355-.768 1.829 1.829 0 00-.592-.539 1.603 1.603 0 00-.814-.204c-.631 0-1.107.219-1.427.657-.319.438-.479 1.021-.479 1.748 0 .342.041.66.125.953.083.294.208.546.374.756.167.21.366.377.598.499.232.123.502.184.809.184.341 0 .63-.07.867-.21.237-.14.432-.322.585-.545a2.21 2.21 0 00.328-.763 3.86 3.86 0 00.099-.874zM12.714 9.823l-2.353-3.22h1.814l1.42 2.09 1.485-2.09h1.735l-2.313 3.141 2.602 3.654h-1.801l-1.721-2.51-1.67 2.51h-1.761l2.563-3.575z"></path>
-</g>
-</svg>;
-icons.em = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-  <path d="M6.248 9.402a2.401 2.401 0 00-.152-.683 1.704 1.704 0 00-.867-.967 1.56 1.56 0 00-.69-.151c-.263 0-.502.046-.716.138a1.633 1.633 0 00-.552.381 1.886 1.886 0 00-.368.572 2.002 2.002 0 00-.152.71h3.497zm-3.497.986c0 .263.038.517.112.762.075.245.186.46.335.644.149.184.338.331.565.44.228.11.5.165.815.165.438 0 .791-.095 1.058-.283.268-.188.467-.471.598-.848h1.42a2.824 2.824 0 01-1.104 1.716 2.99 2.99 0 01-.914.446c-.337.101-.69.152-1.058.152-.534 0-1.007-.088-1.419-.263a2.95 2.95 0 01-1.045-.736 3.12 3.12 0 01-.644-1.131 4.577 4.577 0 01-.217-1.445c0-.482.077-.94.23-1.374.153-.434.372-.815.657-1.143a3.13 3.13 0 011.032-.782c.403-.193.859-.29 1.367-.29a3.04 3.04 0 011.439.336c.425.223.778.519 1.058.887.281.368.484.791.611 1.268.127.478.16.971.099 1.479H2.751zM8.889 6.602h1.42v.947h.039c.114-.167.237-.32.368-.46a2.022 2.022 0 01.999-.585c.211-.057.451-.086.723-.086.412 0 .795.092 1.15.276.355.184.607.469.756.855a3.01 3.01 0 01.881-.828c.333-.202.749-.303 1.248-.303.719 0 1.277.176 1.676.526.399.351.598.938.598 1.761v4.693h-1.498v-3.97c0-.271-.009-.519-.027-.742a1.512 1.512 0 00-.151-.579.856.856 0 00-.374-.374c-.167-.088-.395-.132-.684-.132-.508 0-.876.158-1.104.473-.228.316-.342.763-.342 1.341v3.983h-1.498V9.034c0-.473-.085-.83-.256-1.071-.171-.241-.484-.362-.94-.362-.193 0-.379.04-.559.119a1.385 1.385 0 00-.473.341 1.703 1.703 0 00-.328.552 2.084 2.084 0 00-.125.75v4.035H8.889V6.602z"></path>
-</g>
-</svg>;
-icons.percent = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path
-fillRule="nonzero"
-d="M5.689 7.831c0 .246.017.476.053.69.035.215.092.401.17.559.079.158.182.283.309.375a.775.775 0 00.467.138.803.803 0 00.473-.138.978.978 0 00.315-.375 2.11 2.11 0 00.178-.559c.039-.214.059-.444.059-.69 0-.219-.015-.433-.046-.644a1.995 1.995 0 00-.164-.565 1.076 1.076 0 00-.316-.401.794.794 0 00-.499-.151.797.797 0 00-.5.151 1.02 1.02 0 00-.308.401 1.992 1.992 0 00-.152.565 5.253 5.253 0 00-.039.644zm1.012 2.616c-.394 0-.732-.07-1.012-.21a1.899 1.899 0 01-.684-.566 2.316 2.316 0 01-.381-.828 4.148 4.148 0 01-.118-1.012c0-.35.042-.685.125-1.005.083-.32.215-.598.394-.835.18-.236.408-.425.684-.565.276-.14.606-.21.992-.21s.716.07.992.21c.276.14.504.329.684.565.179.237.311.515.394.835.083.32.125.655.125 1.005 0 .36-.039.697-.118 1.012a2.3 2.3 0 01-.382.828 1.887 1.887 0 01-.683.566c-.28.14-.618.21-1.012.21zm5.586 1.722c0 .245.017.475.053.69.035.214.092.401.17.558.079.158.182.283.309.375a.775.775 0 00.467.138.803.803 0 00.473-.138.978.978 0 00.315-.375c.079-.157.138-.344.178-.558.039-.215.059-.445.059-.69 0-.219-.015-.434-.046-.644a1.992 1.992 0 00-.164-.566 1.065 1.065 0 00-.316-.4.795.795 0 00-.499-.152.798.798 0 00-.5.152 1.01 1.01 0 00-.308.4 1.99 1.99 0 00-.152.566c-.026.21-.039.425-.039.644zm1.012 2.615c-.394 0-.732-.07-1.012-.21a1.885 1.885 0 01-.683-.565 2.317 2.317 0 01-.382-.828 4.16 4.16 0 01-.118-1.012c0-.351.042-.686.125-1.006.083-.32.215-.598.394-.834.18-.237.408-.425.684-.566.276-.14.606-.21.992-.21s.716.07.992.21c.276.141.504.329.684.566.179.236.311.514.394.834.083.32.125.655.125 1.006 0 .359-.039.696-.118 1.012a2.332 2.332 0 01-.381.828 1.897 1.897 0 01-.684.565c-.28.14-.618.21-1.012.21zm-1.341-9.7h.999l-5.086 9.832H6.846l5.112-9.832z"
-></path>
-</svg>;
-icons.rem = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M.731 7.079H1.94v1.13h.023c.038-.158.111-.313.22-.463.11-.151.241-.289.396-.413.154-.124.326-.224.514-.3.189-.075.381-.113.577-.113.151 0 .254.004.311.012l.175.022v1.244a5.951 5.951 0 00-.277-.04 2.393 2.393 0 00-.277-.017 1.424 1.424 0 00-1.119.514c-.143.17-.256.379-.339.628a2.712 2.712 0 00-.125.859v2.781H.731V7.079zM8.519 9.486a2.057 2.057 0 00-.13-.587 1.534 1.534 0 00-.294-.492 1.433 1.433 0 00-.452-.339 1.327 1.327 0 00-.593-.13c-.226 0-.432.039-.616.118-.185.08-.343.189-.475.328-.132.14-.238.304-.317.492a1.727 1.727 0 00-.13.61h3.007zm-3.007.848c0 .226.032.445.096.656.064.211.161.395.289.554.128.158.29.284.486.378.196.095.429.142.701.142.376 0 .68-.081.91-.243.229-.162.401-.405.514-.73h1.221a2.422 2.422 0 01-.95 1.476 2.597 2.597 0 01-.785.384c-.291.087-.594.13-.91.13-.46 0-.867-.075-1.221-.226a2.535 2.535 0 01-.899-.633 2.69 2.69 0 01-.554-.972 3.964 3.964 0 01-.186-1.244c0-.414.066-.808.198-1.181.131-.373.32-.701.565-.983.245-.283.54-.507.887-.673A2.692 2.692 0 017.05 6.92c.459 0 .872.096 1.237.289.366.192.669.446.91.763.242.316.417.68.526 1.09.109.411.138.835.085 1.272H5.512zM10.791 7.079h1.221v.813h.034c.098-.143.203-.275.317-.395a1.722 1.722 0 01.859-.503c.18-.049.388-.074.621-.074.355 0 .684.079.989.238.306.158.522.403.65.734.219-.301.471-.538.758-.712.286-.173.644-.26 1.074-.26.618 0 1.098.151 1.441.452.343.302.514.807.514 1.515v4.036h-1.288V9.509c0-.234-.008-.447-.023-.639a1.292 1.292 0 00-.13-.497.737.737 0 00-.322-.322c-.143-.076-.339-.113-.588-.113-.437 0-.754.135-.95.407-.195.271-.293.655-.293 1.153v3.425h-1.289V9.17c0-.407-.074-.714-.221-.921-.146-.208-.416-.311-.808-.311a1.192 1.192 0 00-.887.395 1.48 1.48 0 00-.283.475 1.815 1.815 0 00-.107.644v3.471h-1.289V7.079z"></path>
-</g>
-</svg>;
-icons.vh = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M3.271 6.603H4.9l1.722 5.218h.026l1.656-5.218h1.551l-2.431 6.795H5.742L3.271 6.603zM10.762 4.014h1.499v3.483h.026c.184-.307.458-.563.821-.769a2.425 2.425 0 011.216-.309c.745 0 1.332.193 1.761.578.43.386.644.964.644 1.735v4.666h-1.498V9.127c-.017-.535-.131-.923-.342-1.164-.21-.241-.538-.361-.985-.361-.254 0-.482.046-.684.138a1.484 1.484 0 00-.512.381c-.141.162-.25.353-.329.572a2.035 2.035 0 00-.118.696v4.009h-1.499V4.014z"></path>
-</g>
-</svg>;
-icons.vw = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M1.621 6.603h1.63l1.722 5.218h.026l1.656-5.218h1.551l-2.432 6.795H4.092L1.621 6.603zM8.495 6.603h1.59l1.328 5.073h.026l1.275-5.073h1.512l1.222 5.073h.026l1.38-5.073h1.525l-2.129 6.795h-1.538L13.45 8.351h-.026l-1.249 5.047h-1.577L8.495 6.603z"></path>
-</g>
-</svg>;
-icons.none = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M17.401 4.69L15.31 2.599 2.599 15.31l2.091 2.091L17.401 4.69z"></path>
-<path d="M4.69 2.599L2.599 4.69 15.31 17.401l2.091-2.091L4.69 2.599z"></path>
-</svg>;
-icons.solid = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M18.988 11.478V8.522H1.012v2.956h17.976z"></path>
-</svg>;
-icons.dashed = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M12.512 11.478V8.522H7.488v2.956h5.024zM14.004 8.522v2.956h4.984V8.522h-4.984zM1.012 8.522v2.956H6.05V8.522H1.012z"></path>
-</svg>;
-icons.dotted = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<circle cx="2.503" cy="10" r="1.487"></circle>
-<circle cx="17.486" cy="10" r="1.487"></circle>
-<circle cx="12.447" cy="10" r="1.487"></circle>
-<circle cx="7.455" cy="10" r="1.487"></circle>
-</svg>;
-icons.double = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<path d="M1.02 6.561v2.957h17.968V6.561H1.02zM1.012 10.586v2.956H18.98v-2.956H1.012z"></path>
-</svg>;
-icons.lowercase = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M9.1 13.181c0 .184.024.315.072.394.048.079.142.118.283.118h.157a.953.953 0 00.211-.026v1.038a3.222 3.222 0 01-.204.06 3.035 3.035 0 01-.743.111c-.306 0-.561-.061-.762-.184-.202-.122-.333-.337-.394-.644-.298.289-.664.5-1.098.631a4.303 4.303 0 01-1.255.197c-.307 0-.6-.041-.881-.125a2.302 2.302 0 01-.742-.368 1.764 1.764 0 01-.513-.617 1.904 1.904 0 01-.19-.874c0-.421.076-.763.23-1.026.153-.262.354-.468.604-.617.25-.149.53-.257.841-.322.311-.066.625-.116.94-.152.272-.052.53-.089.776-.111a3.6 3.6 0 00.65-.112c.189-.053.337-.134.447-.243.11-.11.164-.274.164-.493a.765.765 0 00-.138-.473.931.931 0 00-.341-.283 1.5 1.5 0 00-.454-.131 3.61 3.61 0 00-.473-.033c-.421 0-.767.088-1.038.263-.272.175-.425.447-.46.815H3.29c.027-.438.132-.802.316-1.091.184-.289.418-.522.703-.697.285-.175.607-.298.966-.368s.727-.105 1.104-.105c.333 0 .662.035.986.105.324.07.615.184.874.342.258.158.466.361.624.611.158.25.237.554.237.914v3.496zm-1.499-1.893a1.822 1.822 0 01-.841.27 10.7 10.7 0 00-.999.138 3.18 3.18 0 00-.46.111c-.149.048-.28.114-.394.197a.874.874 0 00-.27.329c-.065.136-.098.3-.098.493 0 .166.048.307.144.42.097.114.213.204.349.27.136.066.285.112.447.138.162.026.309.039.44.039.166 0 .346-.022.539-.065.193-.044.374-.119.545-.224.171-.105.314-.239.427-.401.114-.162.171-.361.171-.598v-1.117zM10.756 5.308h1.498v3.47h.026c.106-.167.233-.316.382-.447.149-.132.313-.243.492-.335a2.62 2.62 0 01.566-.211c.197-.048.392-.072.584-.072.535 0 1.002.094 1.4.283.399.188.73.446.993.775.262.329.46.712.591 1.15.132.438.197.907.197 1.406 0 .456-.059.899-.177 1.328a3.475 3.475 0 01-.539 1.137 2.713 2.713 0 01-.914.789c-.368.197-.801.295-1.301.295-.228 0-.458-.015-.69-.046a2.614 2.614 0 01-.664-.177c-.21-.088-.4-.202-.571-.342a1.645 1.645 0 01-.427-.552h-.027v.933h-1.419V5.308zm5.231 5.993c0-.306-.04-.604-.118-.894a2.44 2.44 0 00-.355-.768 1.818 1.818 0 00-.592-.539 1.604 1.604 0 00-.815-.204c-.631 0-1.106.219-1.426.657-.32.438-.479 1.021-.479 1.748 0 .342.041.66.124.953.084.294.209.546.375.756.166.21.366.377.598.499.232.123.502.184.808.184.342 0 .631-.07.868-.21.236-.14.431-.322.585-.545a2.21 2.21 0 00.328-.763 3.86 3.86 0 00.099-.874z"></path>
-</g>
-</svg>;
-icons.uppercase = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M4.789 5.308h1.748l3.614 9.384H8.39l-.881-2.484H3.763l-.88 2.484H1.187l3.602-9.384zm-.579 5.651h2.866L5.669 6.924H5.63l-1.42 4.035zM12.688 9.238h2.681c.394 0 .723-.112.986-.335.263-.224.394-.546.394-.967 0-.473-.118-.806-.355-.998-.236-.193-.578-.29-1.025-.29h-2.681v2.59zm-1.643-3.93h4.561c.841 0 1.516.193 2.024.578.508.386.762.968.762 1.748 0 .473-.116.879-.348 1.216-.232.337-.563.598-.993.782v.026c.579.123 1.017.397 1.315.822.298.425.447.957.447 1.597 0 .368-.066.712-.197 1.031a2.11 2.11 0 01-.618.828c-.281.233-.64.417-1.078.553-.438.135-.959.203-1.564.203h-4.311V5.308zm1.643 8.044h2.905c.499 0 .887-.13 1.163-.388.276-.259.414-.624.414-1.098 0-.464-.138-.821-.414-1.071-.276-.25-.664-.374-1.163-.374h-2.905v2.931z"></path>
-</g>
-</svg>;
-icons.capitalize = <svg width="24" height="24"
-xmlns="http://www.w3.org/2000/svg"
-fillRule="evenodd"
-strokeLinejoin="round"
-strokeMiterlimit="2"
-clipRule="evenodd"
-viewBox="0 0 20 20"
->
-<g fillRule="nonzero">
-<path d="M5.393 5.216h1.748l3.615 9.384H8.995l-.881-2.484H4.368L3.487 14.6H1.792l3.601-9.384zm-.578 5.651H7.68L6.274 6.832h-.04l-1.419 4.035zM11.479 5.216h1.498v3.47h.026c.105-.167.232-.316.381-.447a2.38 2.38 0 01.493-.335 2.62 2.62 0 01.566-.211c.197-.048.392-.072.584-.072.535 0 1.001.094 1.4.283.399.188.73.446.993.775.262.329.46.712.591 1.15.131.438.197.907.197 1.406 0 .456-.059.899-.177 1.328a3.475 3.475 0 01-.539 1.137 2.713 2.713 0 01-.914.789c-.368.197-.801.295-1.301.295-.228 0-.458-.015-.69-.046a2.614 2.614 0 01-.664-.177 2.278 2.278 0 01-.571-.342 1.66 1.66 0 01-.428-.552h-.026v.933h-1.419V5.216zm5.231 5.993c0-.306-.04-.604-.119-.894a2.416 2.416 0 00-.354-.768 1.818 1.818 0 00-.592-.539 1.604 1.604 0 00-.815-.204c-.631 0-1.106.219-1.426.657-.32.438-.48 1.021-.48 1.748 0 .342.042.66.125.953.084.294.208.546.375.756.166.21.366.377.598.499.232.123.502.184.808.184.342 0 .631-.07.868-.21.236-.14.431-.322.585-.545.153-.224.262-.478.328-.763a3.86 3.86 0 00.099-.874z"></path>
-</g>
-</svg>;
-export default icons;
